% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixtools.R
\docType{package}
\name{admixtools}
\alias{admixtools}
\alias{admixtools-package}
\title{Lightweight R implementations of some AdmixTools programs.}
\description{
The admixtools package provides implementations of some AdmixTools programs
which make use of pre-computed f2 block jackknife statistics.
Aside from the core functions \code{\link{qp3pop}}, \code{\link{qpdstat}}, \code{\link{qpadm}}, and \code{\link{qpgraph}}, there are also wrapper functions and parsing functions around the original AdmixTools software, as well as functions to precompute f2 block jackknife statistics.
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics
}
\seealso{
\url{https://uqrmaie1.github.io/admixtools/index.html}
}
\author{
Robert Maier \email{<rmaier@broadinstitute.org>}

Nick Patterson
}
