% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{afs_to_counts}
\alias{afs_to_counts}
\title{Compute count blocks and write them to disk}
\usage{
afs_to_counts(
  genodir,
  outdir,
  chunk1,
  chunk2,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{genodir}{Directory with genotype files split in chunks created by \code{\link{extract_afs}} (with each individual its own population).}

\item{outdir}{Directory where allele count data will be stored}

\item{chunk1}{Index of the first chunk of individuals}

\item{chunk2}{Index of the second chunk of individuals}

\item{overwrite}{Overwrite existing files (default \code{FALSE})}

\item{verbose}{Print progress updates}
}
\description{
This is intended for computing allele count data for a large number of individuals,
too many to do everything in working memory.
It assumes that the allele frequencies have already been computed and are stored in \code{.rds} files,
split into consecutive blocks for a set of individuals.
}
\seealso{
\code{\link{extract_counts}} Does the same thing in one step for smaller data.
}
