% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ancestrymap_to_afs}
\alias{ancestrymap_to_afs}
\title{Read allele frequencies from ancestrymap files}
\usage{
ancestrymap_to_afs(
  pref,
  inds = NULL,
  pops = NULL,
  adjust_pseudohaploid = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{pref}{Prefix of ancestrymap files (files have to end in \code{.geno}, \code{.ind}, \code{.snp})}

\item{inds}{Individuals from which to compute allele frequencies}

\item{pops}{Populations from which to compute allele frequencies. If \code{NULL} (default), populations will be extracted from the third column in the \code{.ind} file. If population labels are provided, they should have the same length as \code{inds}, and will be matched to them by position}

\item{adjust_pseudohaploid}{Genotypes of pseudohaploid samples are usually coded as \code{0} or \code{2}, even though only one allele is observed. \code{adjust_pseudohaploid} ensures that the observed allele count increases only by \code{1} for each pseudohaploid sample. If \code{TRUE} (default), samples that don't have any genotypes coded as \code{1} among the first 1000 SNPs are automatically identified as pseudohaploid. This leads to slightly more accurate estimates of f-statistics. Setting this parameter to \code{FALSE} is equivalent to the ADMIXTOOLS \code{inbreed: NO} option.}

\item{verbose}{Print progress updates}
}
\value{
A list with three data frames: allele frequency data, allele counts, and SNP metadata
}
\description{
Read allele frequencies from ancestrymap files
}
\examples{
\dontrun{
afdat = ancestrymap_to_afs(prefix, pops = pops)
afs = afdat$afs
counts = afdat$counts
}
}
