% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ancestrymap_to_aftable}
\alias{ancestrymap_to_aftable}
\title{Read allele frequencies from ancestrymap files}
\usage{
ancestrymap_to_aftable(pref, inds = NULL, pops = NULL,
  blocksize = 1000, verbose = TRUE)
}
\arguments{
\item{pref}{prefix of ancestrymap files (files have to end in \code{.geno}, \code{.ind}, \code{.snp}).}

\item{inds}{vector of samples from which to compute allele frequencies.}

\item{pops}{vector of populations from which to compute allele frequencies. If \code{NULL} (default), populations will be extracted from the third column in the \code{.ind} file}

\item{verbose}{print progress updates}
}
\value{
a list with two items: allele frequency data and allele counts.
}
\description{
Read allele frequencies from ancestrymap files
}
\examples{
\dontrun{
afdat = ancestrymap_to_aftable(prefix, pops = pops)
afs = afdat$afs
counts = afdat$counts
}
}
