% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{boo_list}
\alias{boo_list}
\title{Generate a list of bootstrap resampled arrays}
\usage{
boo_list(arr, nboot = dim(arr)[3])
}
\arguments{
\item{arr}{3d array with blocked estimates, with blocks in the 3rd dimension}

\item{nboot}{Number of bootstrap iterations}
}
\value{
A list of bootstrap resampled arrays, with 3rd dimension equal to \code{nboot}
}
\description{
Generate a list of bootstrap resampled arrays
}
\seealso{
\code{\link{loo_list}} \code{\link{est_to_boo}} \code{\link{qpgraph_resample_snps2}}
}
