% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{compare_fits}
\alias{compare_fits}
\title{Compare the fit of two qpgraph models}
\usage{
compare_fits(scores1, scores2)
}
\arguments{
\item{scores1}{Scores for the first graph}

\item{scores2}{Scores for the second graph}
}
\description{
Takes the bootstrap score distribution of two fits on the same populations and tests whether the scores of one graph are significantly better than the scores of the other.
}
\examples{
\dontrun{
fits = qpgraph_resample_multi(f2_blocks, list(graph1, graph2), nboot = 100)
compare_fits(fits[[1]]$score_test, fits[[2]]$score_test)
}
}
\seealso{
\code{\link{qpgraph_resample_multi}}
}
