% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{compare_fits2}
\alias{compare_fits2}
\title{Compare the fit of two qpgraph models}
\usage{
compare_fits2(fits1, fits2, boot = FALSE)
}
\arguments{
\item{fits1}{The fits of the first graph}

\item{fits2}{The fits of the second graph}

\item{boot}{should match the \code{boot} parameter in \code{qpgraph_resample_snps} (\code{FALSE} by default)}
}
\description{
Takes two data frames with model fits computed on two graphs for on the same populations and tests whether the scores of one graph are significantly better than the scores of the other
}
\examples{
\dontrun{
nblocks = dim(example_f2_blocks)[3]
train = sample(1:nblocks, round(nblocks/2))
fits1 = qpgraph_resample_snps(example_f2_blocks[,,train], graph = graph1, f2_blocks_test = example_f2_blocks[,,-train])
fits2 = qpgraph_resample_snps(example_f2_blocks[,,train], graph = graph2, f2_blocks_test = example_f2_blocks[,,-train])
compare_fits2(fit1, fit2)
}
}
