% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{compare_fits4}
\alias{compare_fits4}
\title{Compare the fit of two qpgraph models}
\usage{
compare_fits4(fit1, fit2, f2_blocks, f2_blocks_test, boot = FALSE, seed = NULL)
}
\arguments{
\item{fit1}{The fit of the first graph}

\item{fit2}{The fit of the second graph}

\item{f2_blocks}{f2 blocks used for fitting \code{fit1} and \code{fit2}. Used in combination with \code{f2_blocks_test} to compute f-statistics covariance matrix.}

\item{f2_blocks_test}{f2 blocks which were not used for fitting \code{fit1} and \code{fit2}}

\item{boot}{If \code{TRUE}, bootstrap resampling will be used on \code{f2_blocks_test} with the number of resamplings equal to the number of blocks. If \code{FALSE} jackknife will be used. If set to a number, bootstrap resampling will be used on \code{f2_blocks_test} with the number of resamplings equal to \code{boot}. If bootstrap resampling is enabled, empirical p-values (\code{p_emp}) and 95 confidence intervals (\code{ci_low} and \code{ci_high}) will be reported.}

\item{seed}{Random seed used if \code{boot} is \code{TRUE}. Does not need to match a seed used in fitting the models}
}
\description{
Takes two data frames with model fits computed on two graphs for on the same populations and tests whether the scores of one graph are significantly better than the scores of the other.
}
\examples{
\dontrun{
nblocks = dim(example_f2_blocks)[3]
train = sample(1:nblocks, round(nblocks/2))
fit1 = qpgraph(example_f2_blocks[,,train], graph1)
fit2 = qpgraph(example_f2_blocks[,,train], graph2)
compare_fits4(fit1, fit2, example_f2_blocks[,,train], example_f2_blocks[,,-train])
}
}
