% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{count_snps}
\alias{count_snps}
\title{Count SNPs in an array of per block f2-statistics}
\usage{
count_snps(f2_blocks)
}
\arguments{
\item{f2_blocks}{A 3d array of per block f2-statistics}
}
\value{
The total number of SNPs across all blocks
}
\description{
This function adds up all block lengths, which are stored in the names along the third dimension of the array
}
