% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{count_zero_edges}
\alias{count_zero_edges}
\title{Count zero-length edges}
\usage{
count_zero_edges(edges, epsilon = 1e-06)
}
\arguments{
\item{edges}{Edges data frame from fitted admixture graph}

\item{epsilon}{Every edge with length}
}
\value{
The number of edges with length < epsilon
}
\description{
\code{agraph} is the format used by the \code{admixturegraph} packge. \code{igraph} is used by the \code{admixtools} package
}
\examples{
\dontrun{
fit = qpgraph(example_f2_blocks, example_igraph)
count_zero_edges(fit$edges)
}
}
