% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{decompose_graph}
\alias{decompose_graph}
\title{Find all trees which are part of the admixture graph}
\usage{
decompose_graph(graph)
}
\arguments{
\item{graph}{Admixture graph in \code{igraph} format}
}
\value{
A data frame with columns \code{name} and \code{graph}
}
\description{
Find all trees which are part of the admixture graph
}
\examples{
\dontrun{
trees = decompose_graph(example_igraph)
# now evaluate the trees
trees \%>\%
  rowwise \%>\%
  mutate(res = list(qpgraph(example_f2_blocks, graph))) \%>\%
  unnest_wider(res)
}
}
