% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{delete_groups}
\alias{delete_groups}
\title{Delete groups}
\usage{
delete_groups(dir, groups = NULL, verbose = TRUE)
}
\arguments{
\item{dir}{directory with precomputed individual pair data}

\item{groups}{the groups to delete. defaults to all groups}

\item{verbose}{print progress updates}
}
\value{
invisibly returns sample IDs in deleted groups as character vector
}
\description{
This function deletes data for groups created by \code{group_samples}
}
\examples{
\dontrun{
dir = 'my/f2/dir/'
inds = c('ind1', 'ind2', 'ind3', 'ind4', 'ind5')
pops = c('pop_A', 'pop_A', 'pop_A', 'pop_B', 'pop_B')
group_samples(dir, inds, pops)
}
}
\seealso{
\code{\link{group_samples}}
}
