% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{delete_leaf}
\alias{delete_leaf}
\title{Remove population from graph}
\usage{
delete_leaf(graph, leaf, desimplify = TRUE)
}
\arguments{
\item{graph}{An admixture graph}

\item{leaf}{Population to be removed}

\item{desimplify}{Desimplify graph (\code{\link{desimplify_graph}})}
}
\value{
Admixture graph with removed population
}
\description{
Remove population from graph
}
\seealso{
\code{\link{insert_leaf}}
}
