% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{edges_to_igraph}
\alias{edges_to_igraph}
\title{Convert data frame graph to igraph}
\usage{
edges_to_igraph(edges)
}
\arguments{
\item{edges}{An admixture graph as an edge list data frame}
}
\value{
An \code{igraph} object
}
\description{
Convert data frame graph to igraph
}
