% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{est_to_boo}
\alias{est_to_boo}
\title{Turn per-block estimates into bootstrap estimates}
\usage{
est_to_boo(arr, nboot = dim(arr)[3])
}
\arguments{
\item{arr}{3d array with blocked estimates, with blocks in the 3rd dimension.}

\item{nboot}{Number of bootstrap iterations}
}
\value{
A 3d array with bootstrap estimates. The first two dimensions are equal to those of \code{arr}.
The 3rd dimension is equal to \code{nboot}.
}
\description{
This works for any statistics which, when computed across \code{N} blocks, are equal
to the weighted mean of the statistics across the \code{N} blocks.
}
\seealso{
\code{\link{est_to_loo}}
}
