% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{est_to_loo}
\alias{est_to_loo}
\title{Turn per-block estimates into leave-one-out estimates}
\usage{
est_to_loo(arr, block_lengths = NULL)
}
\arguments{
\item{arr}{3d array with blocked estimates, with blocks in the 3rd dimension}

\item{block_lengths}{Optional block lengths. If \code{NULL}, will be parsed from 3rd dimnames in blocks}
}
\value{
A 3d array with leave-one-out estimates for jackknife. Dimensions are equal to those of \code{arr}.
}
\description{
This works for any statistics which, when computed across \code{N} blocks, are equal
to the weighted mean of the statistics across the \code{N} blocks.
}
\seealso{
\code{\link{loo_to_est}} \code{\link{est_to_boo}}
}
