% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixtools.R
\docType{data}
\name{example_anno}
\alias{example_anno}
\title{Data frame with sample annotations}
\format{
A data frame with sample annotations
}
\usage{
example_anno
}
\description{
Data frame with sample annotations
}
\keyword{datasets}
