% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixtools.R
\docType{data}
\name{example_f2_blocks}
\alias{example_f2_blocks}
\title{Blocked f2-statistics for 7 populations}
\format{
A 3d array with populations along the first two dimensions,
and SNP blocks along the 3rd dimension
}
\usage{
example_f2_blocks
}
\description{
Blocked f2-statistics for 7 populations
}
\keyword{datasets}
