% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixtools.R
\docType{data}
\name{example_f2sim1}
\alias{example_f2sim1}
\title{Simulated f2-statistics for 5 populations}
\format{
A 3d array of f2-statistics computed from msprime simulation output,
with populations along the first two dimensions and SNP blocks along the 3rd
}
\usage{
example_f2sim1
}
\description{
Simulated f2-statistics for 5 populations
}
\keyword{datasets}
