% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixtools.R
\docType{data}
\name{example_graph}
\alias{example_graph}
\title{Admixture graph for 7 populations}
\format{
A two column matrix where each row represents one edge
}
\usage{
example_graph
}
\description{
Admixture graph for 7 populations
}
\keyword{datasets}
