% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixtools.R
\docType{data}
\name{example_igraph}
\alias{example_igraph}
\title{Admixture graph for 7 populations}
\format{
An \code{igraph} object
}
\usage{
example_igraph
}
\description{
Admixture graph for 7 populations
}
\keyword{datasets}
