% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixtools.R
\docType{data}
\name{example_qpgraph_ref_results}
\alias{example_qpgraph_ref_results}
\title{example_graph fitted using qpGraph}
\format{
A list with parsed qpGraph output
}
\usage{
example_qpgraph_ref_results
}
\description{
example_graph fitted using qpGraph
}
\keyword{datasets}
