% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixtools.R
\docType{data}
\name{example_triples}
\alias{example_triples}
\title{Data frame with population triples}
\format{
A data frame with population triples, generated by \code{summarize_triples}
}
\usage{
example_triples
}
\description{
Data frame with population triples
}
\keyword{datasets}
