% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixtools.R
\docType{data}
\name{example_winner}
\alias{example_winner}
\title{Data frame with one fitted admixture graph}
\format{
A data frame with a fitted admixture graph, generated by \code{find_graphs}
}
\usage{
example_winner
}
\description{
Data frame with one fitted admixture graph
}
\keyword{datasets}
