% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixtools.R
\docType{data}
\name{example_winners}
\alias{example_winners}
\title{Data frame with fitted admixture graphs}
\format{
A data frame with fitted admixture graphs, generated by \code{find_graphs}
}
\usage{
example_winners
}
\description{
Data frame with fitted admixture graphs
}
\keyword{datasets}
