% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{extract_afs}
\alias{extract_afs}
\title{Compute and store blocked allele frequency data}
\usage{
extract_afs(pref, outdir, inds = NULL, pops = NULL, dist = 0.05,
  cols_per_chunk = 20, maxmiss = 0.25, minmaf = 0, maxmaf = 0.5,
  transitions = TRUE, transversions = TRUE, keepsnps = NULL,
  format = NULL, poly_only = TRUE, ignore_ploidy = FALSE,
  randomize_alleles = TRUE, seed = NULL, verbose = TRUE)
}
\arguments{
\item{pref}{prefix of packedancestrymap or PLINK files. packedancestrymap has to end in \code{.geno}, \code{.snp}, \code{.ind},
PLINK has to end in \code{.bed}, \code{.bim}, \code{.fam}}

\item{outdir}{the directory to which to write data}

\item{inds}{the individuals for which data should be extracted.}

\item{pops}{the populations for which data should be extracted. \code{pops} and \code{inds} cannot be specified
at the same time. If none are specified, all populations will be extracted.}

\item{dist}{genetic distance in Morgan. Default is 0.05 (50 cM).}

\item{cols_per_chunk}{number of populations per chunk. higher value will lead to fewer,
but more memory intensive jobs when computing f2}

\item{maxmiss}{discard SNPs which are missing in a fraction of populations higher than \code{maxmiss}}

\item{minmaf}{discard SNPs with minor allele frequency less than \code{minmaf}}

\item{maxmaf}{discard SNPs which minor allele frequency greater than than \code{maxmaf}}

\item{transitions}{set to \code{FALSE} to exclude transition SNPs}

\item{transversions}{set to \code{FALSE} to exclude transversion SNPs}

\item{keepsnps}{SNP IDs of SNPs to keep. overrides other SNP filtering options}

\item{format}{supply this if the prefix can refer to genotype data in different formats
and you want to choose which one to read}

\item{ignore_ploidy}{treat pseudohaploid samples as if they were diploid samples (each observed genotype in a sample increases the allele count by two). This is equivalent to the ADMIXTOOLS \code{inbreed: NO} option. The default (\code{FALSE}) produces slightly more accurate f2-statistics by increasing the allele count for each pseudohaploid sample only by one.}

\item{randomize_alleles}{randomly report allele frequencies for reference or for alternative allele for each SNP. For data with large amounts of missingness, this can help to reduce bias of f4-statistics computed from allele frequency products.}

\item{seed}{random seed for \code{randomize_alleles}}

\item{verbose}{print progress updates}
}
\description{
Prepare data for various admixtools functions. Reads data from packedancestrymap or PLINK files,
and computes allele frequencies for selected populations and stores it as \code{.rds} files in outdir.
}
\examples{
\dontrun{
pref = 'my/genofiles/prefix'
outdir = 'dir/for/afdata/'
extract_afs(pref, outdir)
}
}
