% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{extract_counts}
\alias{extract_counts}
\title{Extract and store data needed to compute blocked f2}
\usage{
extract_counts(pref, outdir, inds = NULL, dist = 0.05, maxmiss = 1,
  minmaf = 0, maxmaf = 0.5, transitions = TRUE,
  transversions = TRUE, keepsnps = NULL, maxmem = 8000,
  overwrite = FALSE, format = NULL, verbose = TRUE)
}
\arguments{
\item{pref}{prefix of packedancestrymap or PLINK files. packedancestrymap has to end in \code{.geno}, \code{.snp}, \code{.ind},
PLINK has to end in \code{.bed}, \code{.bim}, \code{.fam}}

\item{outdir}{the directory to which to write data}

\item{inds}{the individuals for which to extract data}

\item{dist}{genetic distance in Morgan. Default is 0.05 (50 cM).}

\item{maxmiss}{discard SNPs which are missing in a fraction of individuals greater than \code{maxmiss}}

\item{minmaf}{discard SNPs with minor allele frequency less than \code{minmaf}}

\item{maxmaf}{discard SNPs which minor allele frequency greater than than \code{maxmaf}}

\item{transitions}{set to \code{FALSE} to exclude transition SNPs}

\item{transversions}{set to \code{FALSE} to exclude transversion SNPs}

\item{keepsnps}{SNP IDs of SNPs to keep. overrides other SNP filtering options}

\item{maxmem}{split up allele frequency data into blocks, if memory requirements exceed \code{maxmem} MB.}

\item{overwrite}{should existing files be overwritten?}

\item{format}{supply this if the prefix can refer to genotype data in different formats
and you want to choose which one to read}

\item{verbose}{print progress updates}
}
\description{
Prepare data for various admixtools functions. This function reads data from packedancestrymap or PLINK files,
and extracts data required to compute blocked f-statistics for any sets of samples. The data consists of
\code{.rds} files with total and alternative allele counts for each individual, and products of total
and alternative allele counts for each pair.
The function calls \code{\link{packedancestrymap_to_aftable}} or \code{\link{plink_to_aftable}}
and \code{\link{afs_to_f2_blocks}}.
}
\examples{
\dontrun{

}
}
