% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{extract_data}
\alias{extract_data}
\title{Extract subset of genotype data}
\usage{
extract_data(pref, outdir = NULL, pops = NULL, inds = NULL,
  na.action = "none", maxmem = 1000, overwrite = FALSE,
  verbose = TRUE)
}
\arguments{
\item{pref}{prefix of packedancestrymap or PLINK files. packedancestrymap has to end in \code{.geno}, \code{.snp}, \code{.ind}, PLINK has to end in \code{.bed}, \code{.bim}, \code{.fam}}

\item{outdir}{the directory to which to write data}

\item{pops}{the populations for which data should be extracted. Defaults to all populations, which may require a lot of memory. If \code{NULL}, each individual will be its own population.}

\item{inds}{the individuals for which data should be extracted. \code{pops} and \code{inds} cannot be specified at the same time. If none are specified, all populations will be extracted}

\item{na.action}{what to do with SNPs with missing data.
\itemize{
\item \code{none} (default) output table will have missing data
\item \code{impute} allele frequencies will be imputed from other populations
\item \code{remove} SNPs with missing data will be removed
}}

\item{maxmem}{split up allele frequency data into blocks, if memory requirements exceed \code{maxmem} MB.}

\item{overwrite}{should existing files be overwritten?}

\item{verbose}{print progress updates}
}
\value{
Nothing, if \code{outdir} is specified. Otherwise a list with \code{f2_blocks} and \code{block_lengths}
}
\description{
Prepare data for various admixtools functions. Reads data from packedancestrymap or PLINK files, and computes allele frequencies and f2 block jackknife statistics for selected populations. Data is either written to disk in files for each population and population pair, or returned as list. This function calls \code{\link{packedancestrymap_to_aftable}} / \code{\link{plink_to_aftable}} and \code{\link{afs_to_f2_blocks}}. The default treatment of missing data is \code{na.action = 'none'} which is similar to \code{useallsnps: YES}. To get results that are closer to \code{useallsnps: NO}, use \code{na.action = 'remove'}.
}
\examples{
\dontrun{
pref = 'my/genofiles/prefix'
data = extract_data(pref, pops = c('Protoss', 'Zerg', 'Terran'))
f2_blocks = data$f2_blocks
block_lengths = data$block_lengths
}
}
