% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{extract_f2}
\alias{extract_f2}
\title{Compute and store blocked f2 statistics}
\usage{
extract_f2(pref, outdir, inds = NULL, pops = NULL, dist = 0.05,
  maxmem = 8000, maxmiss = 0.25, minmaf = 0, maxmaf = 0.5,
  transitions = TRUE, transversions = TRUE, keepsnps = NULL,
  overwrite = FALSE, format = NULL, verbose = TRUE)
}
\arguments{
\item{pref}{prefix of packedancestrymap or PLINK files. packedancestrymap has to end in \code{.geno}, \code{.snp}, \code{.ind},
PLINK has to end in \code{.bed}, \code{.bim}, \code{.fam}}

\item{outdir}{the directory to which to write data}

\item{inds}{the individuals for which data should be extracted.}

\item{pops}{the populations for which data should be extracted. \code{pops} and \code{inds} cannot be specified
at the same time. If none are specified, all populations will be extracted.}

\item{dist}{genetic distance in Morgan. Default is 0.05 (50 cM).}

\item{maxmem}{split up allele frequency data into blocks, if memory requirements exceed \code{maxmem} MB.}

\item{maxmiss}{discard SNPs which are missing in a fraction of populations higher than \code{maxmiss}}

\item{minmaf}{discard SNPs with minor allele frequency less than \code{minmaf}}

\item{maxmaf}{discard SNPs which minor allele frequency greater than than \code{maxmaf}}

\item{transitions}{set to \code{FALSE} to exclude transition SNPs}

\item{transversions}{set to \code{FALSE} to exclude transversion SNPs}

\item{keepsnps}{SNP IDs of SNPs to keep. overrides other SNP filtering options}

\item{overwrite}{should existing files be overwritten?}

\item{format}{supply this if the prefix can refer to genotype data in different formats
and you want to choose which one to read}

\item{verbose}{print progress updates}
}
\description{
Prepare data for various admixtools functions. Reads data from packedancestrymap or PLINK files,
and computes allele frequencies and f2 block jackknife statistics for selected populations.
Data is either written to disk in files for each population and population pair, or returned as list.
This function calls \code{\link{packedancestrymap_to_aftable}} or \code{\link{plink_to_aftable}}
and \code{\link{afs_to_f2_blocks}}.
}
\examples{
\dontrun{
pref = 'my/genofiles/prefix'
extract_f2(pref, pops = c('popA', 'popB', 'popC'))
}
}
