% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{extract_f2_subset}
\alias{extract_f2_subset}
\title{Copy a subset of f2-statistics to a new directory}
\usage{
extract_f2_subset(from, to, pops)
}
\arguments{
\item{from}{directory with f2-statistics}

\item{to}{target directory}

\item{pops}{the populations to copy}

\item{verbose}{print progress updates}
}
\description{
Copy a subset of f2-statistics to a new directory
}
\examples{
\dontrun{
pref = 'my/genofiles/prefix'
outdir = 'dir/for/afdata/'
extract_f2_subset(pref, outdir)
}
}
