% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{extract_indpairs}
\alias{extract_indpairs}
\title{Write data for individual pairs to disk}
\usage{
extract_indpairs(pref, outdir, inds = NULL, na.action = "none",
  maxmem = 8000, overwrite = FALSE, verbose = TRUE)
}
\description{
Write data for individual pairs to disk
}
