% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract samples from packedancestrymap files}
\usage{
extract_samples(inpref, outpref, inds = NULL, pops = NULL)
}
\arguments{
\item{inpref}{Prefix of the packedancestrymap input files}

\item{outpref}{Prefix of the PLINK output files}

\item{inds}{Individuals which should be extracted}

\item{pops}{Populations which should be extracted. Can not be provided together with 'inds'}
}
\description{
This function extracts samples from packedancestrymap files and saves them as PLINK files, using the package 'genio'
}
