% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{f2}
\alias{f2}
\title{Estimate f2 statistics}
\usage{
f2(pop1 = NULL, pop2 = NULL, f2_blocks = NULL,
  block_lengths = NULL, f2_dir = NULL, f2_denom = 1, sure = FALSE,
  unique_only = TRUE, verbose = FALSE)
}
\arguments{
\item{pop1}{one of the following four:
\enumerate{
\item \code{NULL}: all possible pairs of the populations in \code{f2_blocks} will be returned
\item a vector of population labels
\item a data frame in which the first four columns specify the population quadruples to be tested. \code{pop2} will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or population quadruples to be tested. \code{pop2} will be ignored.
}}

\item{pop2}{a vector of population labels}

\item{f2_blocks}{3d array of block-jackknife leave-one-block-out estimates of f2 statistics. output of \code{\link{afs_to_f2_blocks}}. they are weighted by inverse of outgroup heterozygosity, if outgroup was specified.}

\item{block_lengths}{the jackknife block lengths used in computing the f2 statistics. see \code{\link{get_block_lengths}}.}

\item{f2_dir}{a directory with f2 statistics for each population pair in the graph. must contain 'block_lengths.RData'.}

\item{f2_denom}{scales f2-statistics. 1 correspondes to \code{f4mode: YES}. 1/4.75 is similar to \code{f4mode: NO}.}

\item{pop3}{a vector of population labels}
}
\value{
If \code{printonly}, the \code{qpDstat} command, otherwise a data frame with parsed \code{qpDstat} output
}
\description{
Computes f2 statistics from f2 jackknife blocks of the form \eqn{F_2(P_1, P_2)}
}
\examples{
\dontrun{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
f2(pop1, pop2, f2_blocks = f2_blocks, block_lengths = block_lengths)
f2(pop1, pop2, f2_dir = f2_dir)
}
}
