% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{f2_from_geno}
\alias{f2_from_geno}
\title{Read f2 jackknife blocks from genotype data}
\usage{
f2_from_geno(
  pref,
  inds = NULL,
  pops = NULL,
  maxmem = 8000,
  blgsize = 0.05,
  format = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pref}{Prefix of \emph{(packed)ancestrymap} or \emph{PLINK} files. \emph{(packed)ancestrymap} has to end in \code{.geno}, \code{.snp}, \code{.ind},
PLINK has to end in \code{.bed}, \code{.bim}, \code{.fam}}

\item{inds}{Individuals for which data should be extracted}

\item{pops}{Populations for which data should be extracted. If both \code{pops} and \code{inds} are provided, they should have the same length and will be matched by position. If only \code{pops} is provided, all individuals from the \code{.ind} or \code{.fam} file in those populations will be extracted. If only \code{inds} is provided, each indivdual will be assigned to its own population of the same name. If neither \code{pops} nor \code{inds} is provided, all individuals and populations in the \code{.ind} or \code{.fam} file will be extracted.}

\item{maxmem}{Maximum amount of memory to be used. If the required amount of memory exceeds \code{maxmem}, allele frequency data will be split into blocks, and the computation will be performed separately on each block pair.}

\item{blgsize}{SNP block size in Morgan. Default is 0.05 (50 cM).}

\item{format}{Supply this if the prefix can refer to genotype data in different formats
and you want to choose which one to read. Should be \code{plink} to read \code{.bed}, \code{.bim}, \code{.fam} files, or \code{ancestrymap}, or \code{packedancestrymap} to read \code{.geno}, \code{.snp}, \code{.ind} files.}

\item{verbose}{Print progress updates}
}
\description{
This function is currently out of date. The idea is to read a 3d array of f2-statistics directly from genotype data without the intermediate step of writing files to disk. It should combine \code{\link{extract_f2}} and \code{\link{f2_from_precomp}} and could be useful when f2-statistics don't need to be stored on disk. Might be easier to add an option to \code{\link{extract_f2}} to not write to disk.
}
\examples{
\dontrun{
pref = 'my/genofiles/prefix'
f2_blocks = f2_from_geno(pref, pops = c('pop1', 'pop2', 'pop3'))
}
}
