% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{f2_from_precomp}
\alias{f2_from_precomp}
\title{Read f2 jackknife blocks from disk}
\usage{
f2_from_precomp(dir, poplist = NULL, verbose = TRUE)
}
\arguments{
\item{dir}{directory with precomputed f2 jackknife blocks}

\item{poplist}{the populations for which data should be extracted. Defaults to all populations, which may require a lot of memory. If \code{NULL}, each individual will be its own population.}

\item{verbose}{print progress updates}
}
\description{
Read f2 jackknife blocks from disk
}
\examples{
\dontrun{
dir = 'my/f2/dir/'
f2_blocks = f2_from_precomp(dir, pops = c('Protoss', 'Zerg', 'Terran'))
}
}
