% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{f2dat_f4dat}
\alias{f2dat_f4dat}
\title{Turn f2 data to f4 data}
\usage{
f2dat_f4dat(f2dat, popcomb = NULL)
}
\arguments{
\item{f2dat}{A data frame of f2-statistics with columns \code{pop1}, \code{pop2}, \code{f2}}
}
\value{
A data frame with f4-statistics
}
\description{
Turn f2 data to f4 data
}
