% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{f3blockdat_from_geno}
\alias{f3blockdat_from_geno}
\title{f3 from genotype data}
\usage{
f3blockdat_from_geno(
  pref,
  popcombs,
  auto_only = TRUE,
  blgsize = 0.05,
  block_lengths = NULL,
  allsnps = FALSE,
  adjust_pseudohaploid = TRUE,
  poly_only = FALSE,
  apply_corr = TRUE,
  outgroupmode = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{pref}{Prefix of genotype files}

\item{popcombs}{A data frame with one population combination per row, and columns \code{pop1}, \code{pop2}, \code{pop3}, \code{pop4}. If there is an additional integer column named \code{model} and \code{allsnps = FALSE}, only SNPs present in every population in any given model will be used to compute f4-statistics for that model.}

\item{auto_only}{Use only chromosomes 1 to 22.}

\item{blgsize}{SNP block size in Morgan. Default is 0.05 (5 cM). If \code{blgsize} is 100 or greater, if will be interpreted as base pair distance rather than centimorgan distance.}

\item{block_lengths}{An optional vector with block lengths. If \code{NULL}, block lengths will be computed.}

\item{allsnps}{Use all SNPs with allele frequency estimates in every population of any given population quadruple. If \code{FALSE} (the default) only SNPs which are present in all populations in \code{popcombs} (or any given model in it) will be used. Setting \code{allsnps = TRUE} in the presence of large amounts of missing data might lead to false positive results.}

\item{adjust_pseudohaploid}{Genotypes of pseudohaploid samples are usually coded as \code{0} or \code{2}, even though only one allele is observed. \code{adjust_pseudohaploid} ensures that the observed allele count increases only by \code{1} for each pseudohaploid sample. If \code{TRUE} (default), samples that don't have any genotypes coded as \code{1} among the first 1000 SNPs are automatically identified as pseudohaploid. This leads to slightly more accurate estimates of f-statistics. Setting this parameter to \code{FALSE} is equivalent to the ADMIXTOOLS \code{inbreed: NO} option. Setting \code{adjust_pseudohaploid} to an integer \code{n} will check the first \code{n} SNPs instead of the first 1000 SNPs.}

\item{apply_corr}{With \code{apply_corr = FALSE}, no bias correction is performed. With \code{apply_corr = TRUE} (the default), a bias correction term based on the heterozygosity in the first population is subtracted from the f3 estimate. With \code{apply_corr = 2}, the bias correction term is calculated based on all 3 populations. This option is not generally recommended, and only exists to match how the f3-statistics are estimated in certain scenarios in the original qpGraph program.}

\item{outgroupmode}{Default \code{FALSE} if f3 computed directly from genotype data, and \code{TRUE} if f3 is computed from f2-statistics. With \code{outgroupmode = FALSE}, estimates of f3 will be normalized by estimates of the heterozygosity of the target population. With \code{outgroupmode = TRUE}, this normalization is simply ommited.}

\item{verbose}{Print progress updates}
}
\value{
A data frame with per-block f4-statistics for each population quadruple.
}
\description{
Compute per-block f3-statistics directly from genotype data
}
