% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{f4_from_afdat}
\alias{f4_from_afdat}
\title{Compute f4 from allele frequencies}
\usage{
f4_from_afdat(afdat, popcombs)
}
\arguments{
\item{afdat}{A data frame with allele frequencies, and SNP metadata. Can be grouped.}

\item{popcombs}{A data frame with population combinations. Columns \code{pop1} to \code{pop4}}
}
\description{
Compute f4 from allele frequencies
}
\examples{
\dontrun{
afs = plink_to_afs('/my/geno/prefix', pops = c('p1', 'p2', 'p3', 'p4', 'p5'))
afdat = bind_cols(afs$snpfile, afs$afs \%>\% as_tibble()) \%>\%
        mutate(gr = paste0(pmin(A1, A2), pmax(A1, A2))) \%>\%
        group_by(gr)
popcombs = tibble(pop1 = c('p1', 'p5'), pop2 = 'p2', pop3 = 'p3', pop4 = 'p4')
out = f4_from_afdat(afdat, popcombs)
out \%>\% ggplot(aes(gr, est)) + geom_point() +
          geom_errorbar(aes(ymin = est - se, ymax = est + se)) +
          facet_wrap(~paste(pop1, pop2, pop3, pop4), scales = 'free')
}
}
