% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{f4_from_f2}
\alias{f4_from_f2}
\title{Get per-block f4-statistics}
\usage{
f4_from_f2(f2_data, pop1, pop2 = NULL, pop3 = NULL, pop4 = NULL)
}
\arguments{
\item{f2_data}{A 3d array with blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{f2_from_geno}}
Alternatively, a directory with precomputed data. See \code{\link{extract_f2}} and \code{\link{extract_counts}}.}

\item{pop1}{Either the name(s) of the first population(s), or a four column matrix with the names of all four populations.}

\item{pop2}{Population 2 (same length as \code{pop1})}

\item{pop3}{Population 3 (same length as \code{pop1})}

\item{pop4}{Population 4 (same length as \code{pop1})}
}
\value{
A matrix of per-block f4-statistics (\verb{popcomb x block})
}
\description{
This function turns per-block f2-statistics into per-block f4-statistics of the form \verb{f4(pop1, pop2; pop3, pop4)}
}
