% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{f4_from_f2_pops}
\alias{f4_from_f2_pops}
\title{Get per-block f4-statistics f4(pop1, pop2; pop3, pop4)}
\usage{
f4_from_f2_pops(f2_blocks, pop1, pop2, pop3, pop4)
}
\arguments{
\item{f2_blocks}{3d array of f2-statistics}

\item{pop1}{Population 1}

\item{pop2}{Population 2}

\item{pop3}{Population 3}

\item{pop4}{Population 4}
}
\value{
A vector of per-block f4-statistics
}
\description{
Get per-block f4-statistics f4(pop1, pop2; pop3, pop4)
}
