% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{f4blockdat_to_f4blocks}
\alias{f4blockdat_to_f4blocks}
\title{Turn f4 block data to 3d array}
\usage{
f4blockdat_to_f4blocks(f4blockdat, remove_na = TRUE)
}
\arguments{
\item{f4blockdat}{f4 block data frame generated by \code{\link{f4blockdat_from_geno}}}

\item{remove_na}{Remove blocks with missing values}
}
\description{
Turn f4 block data to 3d array
}
