% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{find_admixedges}
\alias{find_admixedges}
\title{Find admixture edges}
\usage{
find_admixedges(graph)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
A data frame with columns \code{from} and \code{to} with admixture edges
}
\description{
Find admixture edges
}
\seealso{
\code{\link{find_normedges}} \code{\link{find_newedges}}
}
