% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{find_newedges}
\alias{find_newedges}
\title{Find possible new edges}
\usage{
find_newedges(graph, fix_outgroup = TRUE, all = TRUE)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
A data frame with columns \code{from} and \code{to}. New edges which begin above \code{from}
and end above \code{to} could be inserted
}
\description{
Find possible new edges
}
\seealso{
\code{\link{find_normedges}} \code{\link{find_admixedges}}
}
