% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{flipadmix_random}
\alias{flipadmix_random}
\title{Modify a graph flipping the direction of an admixture edge}
\usage{
flipadmix_random(graph, desimplify = TRUE, fix_outgroup = TRUE)
}
\arguments{
\item{graph}{An admixture graph}

\item{desimplify}{Desimplify graph (\code{\link{desimplify_graph}})}

\item{fix_outgroup}{Keep outgroup in place (has no effect here)}
}
\value{
A new admixture graph
}
\description{
Modify a graph flipping the direction of an admixture edge
}
