% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{fst}
\alias{fst}
\title{Compute Fst}
\usage{
fst(data, pop1 = NULL, pop2 = NULL, boot = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{A directory which contains pre-computed f2- and fst-statistics}

\item{pop1}{One of the following four:
\enumerate{
\item \code{NULL}: all possible population combinations will be returned
\item A vector of population labels. All combinations with the other \code{pop} arguments will be returned
\item A matrix with population combinations to be tested, with one population per column and one
combination per row. Other \code{pop} arguments will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or
population combinations to be tested. Other \code{pop} arguments will be ignored.
}}

\item{pop2}{A vector of population labels}

\item{boot}{If \code{FALSE} (the default), block-jackknife resampling will be used to compute standard errors.
Otherwise, block-bootstrap resampling will be used to compute standard errors. If \code{boot} is an integer, that number
will specify the number of bootstrap resamplings. If \code{boot = TRUE}, the number of bootstrap resamplings will be
equal to the number of SNP blocks.}

\item{verbose}{Print progress updates}
}
\description{
This function reads Fst from a directory with precomputed f2-statistics, and turns per-block data
into estimates and standard errors for each population pair. See \code{details} for how Fst is computed.
}
\details{
The Hudson Fst estimator used here is described in the two publications below.
For two populations with estimated allele frequency vectors \code{p1} and \code{p2},
and allele count vectors \code{n1} and \code{n2}, it is calculated as follows:\cr\cr
\code{num = (p1 - p2)^2 - p1*(1-p1)/(n1-1) - p2*(1-p2)/(n2-1)}\cr
\code{denom = p1 + p2 - 2*p1*p2}\cr
\code{fst = mean(num)/mean(denom)}\cr\cr
This is done independently for each SNP block, and is stored on disk for each population pair.
Jackknifing or bootstrapping across these per-block estimates yields the overall estimates and standard errors.
}
\examples{
\dontrun{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
fst(f2_dir, pop1, pop2)
}
}
\references{
Reich, D. (2009) \emph{Reconstructing Indian population history} Nature

Bhatia, G. (2013) \emph{Estimating and interpreting Fst: the impact of rare variants} Genome Research
}
