% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{generate_all_graphs}
\alias{generate_all_graphs}
\title{Generate all graphs}
\usage{
generate_all_graphs(leaves, nadmix = 0, verbose = TRUE)
}
\arguments{
\item{leaves}{The leaf nodes}

\item{nadmix}{The number of admixture nodes}

\item{verbose}{Print progress updates}
}
\value{
A list of graphs in \code{igraph} format
}
\description{
This functions generates all possible admixture graphs with a set number of admixture events for a given set of leaf nodes.
}
