% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{generate_all_trees}
\alias{generate_all_trees}
\title{Generate all trees}
\usage{
generate_all_trees(leaves)
}
\arguments{
\item{leaves}{The leaf nodes}
}
\value{
A list of trees in \code{igraph} format
}
\description{
This functions generates all possible trees with for a given set of leaf nodes.
}
