% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{get_block_lengths}
\alias{get_block_lengths}
\title{Finds LD-independent blocks.}
\usage{
get_block_lengths(afdat, dist = 0.05, distcol = "cm")
}
\arguments{
\item{afdat}{data frame with columns 'CHR' and either 'POS' or 'cm'}

\item{dist}{minimum distance between blocks}

\item{distcol}{which column should be used as distance column}
}
\value{
a numeric vector where the ith element lists the number of SNPs in the ith block.
}
\description{
starts new block at the SNP after the first SNP which is not within dist of the last block start
afdat needs to be ordered first by 'CHR', then by 'POS' or 'cm'
}
\examples{
\dontrun{
prefix = 'path/to/packedancestrymap_prefix'
pops = c('Stark', 'Targaryen', 'Lannister')
afdat = packedancestrymap_to_aftable(prefix, pops)
block_lengths = get_block_lengths(afdat)
}
}
