% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{get_f2}
\alias{get_f2}
\title{Turns f2_data into f2_blocks}
\usage{
get_f2(f2_data, pops, pops2 = pops, afprod = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{f2_data}{f2 data as genotype file prefix, f2 directory, or f2 blocks}

\item{pops}{Populations for which to extract f2-stats}

\item{pops2}{Optional second vector of populations. Can result in non-square array}

\item{afprod}{Get allele frequency products}

\item{verbose}{Print progress updates}

\item{...}{Additional arguments passed to \code{\link{f2_from_precomp}} or \code{\link{f2_from_geno}}}
}
\value{
A 3d array with f2-statistics
}
\description{
Turns f2_data into f2_blocks
}
