% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{get_leafnames}
\alias{get_leafnames}
\title{Get the population names of a graph}
\usage{
get_leafnames(graph)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
Population names
}
\description{
Get the population names of a graph
}
