% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{get_outpop}
\alias{get_outpop}
\title{Get the outgroup from a graph (if it exists)}
\usage{
get_outpop(graph)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
Outgroup name
}
\description{
Get the outgroup from a graph (if it exists)
}
