% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{graph_addleaf}
\alias{graph_addleaf}
\title{Add a population to an admixture graph}
\usage{
graph_addleaf(graph, pop)
}
\arguments{
\item{graph}{An admixture graph}

\item{pop}{Population to add to the graph}
}
\value{
Admixture graph with the added population
}
\description{
Add a population to an admixture graph
}
\seealso{
\code{\link{insert_leaf}} adds pop at a specific position
}
