% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{graph_distances}
\alias{graph_distances}
\title{Pairwise distance estimates for graphs}
\usage{
graph_distances(graphlist)
}
\arguments{
\item{graphlist}{List of graphs}
}
\value{
A data frame with graph distances
}
\description{
Computes a distance estimate for each graph pair. Each graph is first summarized as a vector which counts for every leaf pair how many internal nodes reach that pair. The distance between two graphs is the Euclidean distance between the vectors of two graphs, and is scaled to fall between 0 and 1.
}
