% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{graph_equations}
\alias{graph_equations}
\title{Find well fitting admixture graphs}
\usage{
graph_equations(graph, substitute = TRUE, nam = c("a", "e", "f"))
}
\arguments{
\item{graph}{An admixture graph}

\item{substitute}{Should edge names be represented by shorter symbols?}

\item{nam}{Symbols used to shorten edge names}
}
\value{
A list with two data frames: \code{equations} holds the equtions for all f2-statistics; \code{coding} has the mapping from edge names to edge symbols, which is used when \code{substitute = TRUE}
}
\description{
This function generates and evaluates admixture graphs in \code{numgen} iterations
to find well fitting admixturegraphs.
}
