% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{graph_f2_function}
\alias{graph_f2_function}
\title{Make a function representing a graph}
\usage{
graph_f2_function(graph, admix_default = 0.5, drift_default = 0.01)
}
\arguments{
\item{graph}{An admixture graph}

\item{admix_default}{The default weights for admixture edges}

\item{drift_default}{The default weights for drift edges}
}
\value{
A function mapping edge weights to f2-statistics
}
\description{
This function takes an igraph object and turns it into a function that takes edge weights as input,
and outputs the expected f2-statistics.
}
\examples{
\dontrun{
mygraph = graph_f2_function(example_igraph)
mygraph(N3N8 = 0.1, `N2N1|Vindija.DG` = 0.4)
}
}
