% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{graph_flipadmix}
\alias{graph_flipadmix}
\title{Find all valid graphs which result from flipping one admixture edge}
\usage{
graph_flipadmix(graph)
}
\arguments{
\item{graph}{Admixture graph in \code{igraph} format}
}
\value{
A data frame with columns \code{from}, \code{to}, and \code{graph}
}
\description{
Find all valid graphs which result from flipping one admixture edge
}
\examples{
\dontrun{
newgraphs = graph_flipadmix(example_igraph)
# now evaluate the new graphs
newgraphs \%>\%
  rowwise \%>\%
  mutate(res = list(qpgraph(example_f2_blocks, graph))) \%>\%
  unnest_wider(res)
}
}
