% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{graph_hash}
\alias{graph_hash}
\title{Get unique hash of an admixture graph}
\usage{
graph_hash(graph)
}
\arguments{
\item{graph}{Admixture graph in igraph format}
}
\value{
A hash string of the admixture graph
}
\description{
This can be used to check if two graphs are identical. For two graphs to be identical, the topology and leaf nodes have to match, but internal node names do not matter.
}
