% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{graph_minusplus}
\alias{graph_minusplus}
\title{Find all graphs which result from adding and removing one admixture edge}
\usage{
graph_minusplus(graph, desimplify = TRUE)
}
\arguments{
\item{graph}{Admixture graph in \code{igraph} format}

\item{desimplify}{Separate admixture edges from drift edges (default is \code{TRUE})}
}
\value{
A data frame with columns \code{from1}, \code{to1}, \code{from2}, \code{to2}, and \code{graph}
}
\description{
Find all graphs which result from adding and removing one admixture edge
}
\examples{
\dontrun{
newgraphs = graph_minusplus(example_igraph)
# now evaluate the new graphs
newgraphs \%>\%
  rowwise \%>\%
  mutate(res = list(qpgraph(example_f2_blocks, graph))) \%>\%
  unnest_wider(res)
}
}
