% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{graph_to_pcs}
\alias{graph_to_pcs}
\title{Simulate PCs under an admixture graph}
\usage{
graph_to_pcs(
  graph,
  nsnps = 10000,
  drift_default = 0.02,
  admix_default = 0.5,
  leaves_only = TRUE
)
}
\arguments{
\item{graph}{An admixture graph as igraph object, or as edge list data frame
with a column \code{weight}, as returned by \code{qpgraph()$edges}}

\item{nsnps}{Number of SNPs to simulate}

\item{drift_default}{Default branch lengths. Ignored if \code{graph} is a data frame with weights}

\item{admix_default}{Default admixture weights. Ignored if \code{graph} is a data frame with weights}

\item{leaves_only}{Return PCs for leaf nodes only}
}
\value{
A data frame with PCs for each population
}
\description{
This function simulates PCA of allele frequencies under an admixture graph model
}
\examples{
\dontrun{
pcs = graph_to_pcs(example_igraph)
pcs \%>\% ggplot(aes(PC1, PC2, label = pop)) + geom_text() + geom_point()
}
}
\seealso{
\code{\link{graph_to_afs}}
}
