% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{group_samples}
\alias{group_samples}
\title{Group precomputed data}
\usage{
group_samples(dir, inds, pops, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{dir}{Directory with precomputed individual pair data}

\item{inds}{Individuals to group}

\item{pops}{Group names, either length 1, or same length as \code{inds}}

\item{overwrite}{Overwrite existing files in \code{outdir}}

\item{verbose}{print progress updates}
}
\description{
Computing f2 statistics for populations consisting of many individuals can be slow and require a lot of memory.
To speed this up, this function groups individuals into populations, computes allele counts and products of
pairwise allele counts with all individuals and groups, and writes the data to disk. f2 statistics for a
combination of grouped and ungrouped precomputed data can then be read using \code{\link{f2_from_precomp}},
replacing individual IDs of the grouped samples with the new group labels.
All groupings are listed in \code{{dir}/groups/{groupname}.rds}
}
\examples{
\dontrun{
dir = 'my/f2/dir/'
inds = c('ind1', 'ind2', 'ind3', 'ind4', 'ind5')
pops = c('pop_A', 'pop_A', 'pop_A', 'pop_B', 'pop_B')
group_samples(dir, inds, pops)
}
}
\seealso{
\code{\link{delete_groups}}
}
