% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{igraph_to_agraph}
\alias{igraph_to_agraph}
\title{Convert igraph to agraph}
\usage{
igraph_to_agraph(igraph)
}
\arguments{
\item{igraph}{An admixture graph in igraph format (as used by admixtools).}
}
\value{
An admixture graph in agraph format (as used by the \code{admixturegraph} package).
}
\description{
\code{agraph} is the format used by the \code{admixturegraph} package. \code{igraph} is used by the \code{admixtools} package
}
\examples{
\dontrun{
igraph_to_agraph(example_igraph)
}
}
