% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{insert_admix}
\alias{insert_admix}
\title{Insert random admixture nodes}
\usage{
insert_admix(edges, n = 1)
}
\arguments{
\item{edges}{an admixture graph.}

\item{n}{the number of admixture nodes to insert.}
}
\value{
a new graph with inserted admixture nodes.
}
\description{
First edge should be R -> outgroup.
Does not always result in correct graphs. Check topology before using them!
}
