% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{insert_admix_multi}
\alias{insert_admix_multi}
\title{Insert admixture edges into graph}
\usage{
insert_admix_multi(
  graph,
  n = 1,
  source_from = NULL,
  source_to = NULL,
  dest_from = NULL,
  dest_to = NULL,
  substitute = FALSE
)
}
\arguments{
\item{graph}{An admixture graph}

\item{substitute}{Should another edge be inserted, if the one specified doesn't work?}

\item{from}{List of nodes. New edges will originate above these nodes.}

\item{to}{List of nodes. New edges will end above these nodes.}
}
\value{
Admixture graph with inserted edges
}
\description{
Insert admixture edges into graph
}
\seealso{
\code{\link{insert_admix}} \code{\link{delete_admix}}
}
