% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{insert_admix_n}
\alias{insert_admix_n}
\title{Insert admixture edges into graph}
\usage{
insert_admix_n(graph, n = 1, fix_outgroup = TRUE)
}
\arguments{
\item{graph}{An admixture graph}

\item{n}{Number of admixture edges to insert (default 1).}

\item{fix_outgroup}{Prevent insertion of edges that affect the root-to-outgroup edge (default \code{TRUE}).}
}
\value{
Admixture graph with inserted edges
}
\description{
Insert admixture edges into graph
}
\seealso{
\code{\link{insert_admix}} \code{\link{delete_admix}}
}
