% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{insert_admix_old}
\alias{insert_admix_old}
\title{Insert admixture edges into graph}
\usage{
insert_admix_old(
  graph,
  fromnodes,
  tonodes,
  substitute_missing = FALSE,
  allow_below_admix = FALSE
)
}
\arguments{
\item{graph}{An admixture graph}

\item{fromnodes}{List of nodes. New edges will originate above these nodes.}

\item{tonodes}{List of nodes. New edges will end above these nodes.}

\item{substitute_missing}{If \code{TRUE}, an attempt will be made to insert random other edges
if some of the provided edges could not be inserted.}

\item{allow_below_admix}{Allow insertion of edges which begin or end directly underneath an admixture node}
}
\value{
Adxmiture graph with inserted edges
}
\description{
Insert admixture edges into graph
}
