% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{insert_edge}
\alias{insert_edge}
\title{Insert a single edge into graph}
\usage{
insert_edge(graph, from, to)
}
\arguments{
\item{graph}{An admixture graph}

\item{from}{List of nodes. New edges will originate above this node}

\item{to}{List of nodes. New edges will end above this node}
}
\value{
Adxmiture graph with inserted edge
}
\description{
Insert a single edge into graph
}
\seealso{
\code{\link{insert_edges}}, \code{\link{delete_admix}}
}
