% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{insert_edges}
\alias{insert_edges}
\title{Insert admixture edges into graph}
\usage{
insert_edges(graph, from, to)
}
\arguments{
\item{graph}{An admixture graph}

\item{from}{List of nodes. New edges will originate above these nodes.}

\item{to}{List of nodes. New edges will end above these nodes.}
}
\value{
Adxmiture graph with inserted edges
}
\description{
Insert admixture edges into graph
}
\seealso{
\code{\link{insert_edge}} \code{\link{delete_admix}}
}
