% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{insert_leaf}
\alias{insert_leaf}
\title{Add population to graph}
\usage{
insert_leaf(graph, leaf, from, to)
}
\arguments{
\item{graph}{An admixture graph}

\item{leaf}{Population to be added}

\item{from}{Source node of edge onto which \code{leaf} should be added}

\item{to}{Target node of edge onto which \code{leaf} should be added}
}
\value{
Admixture graph with added population
}
\description{
Add population to graph
}
\seealso{
\code{\link{delete_leaf}}, \code{\link{graph_addleaf}} to add \code{leaf} at any position
}
