% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{is_valid}
\alias{is_valid}
\title{Test if an admixture graph is valid}
\usage{
is_valid(graph)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
\code{TRUE} if graph is valid, otherwise \code{FALSE}
}
\description{
Test if an admixture graph is valid
}
