% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{isomorphism_classes}
\alias{isomorphism_classes}
\title{Find identical graphs}
\usage{
isomorphism_classes(igraphlist)
}
\arguments{
\item{igraphlist}{A list with admixture graphs}
}
\value{
An integer vector with isomorphism classes.
Graphs with the same number have identical topology (but may have different labels).
}
\description{
Find identical graphs
}
\seealso{
\code{\link{isomorphism_classes2}}
}
