% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{joint_sfs}
\alias{joint_sfs}
\title{Joint site frequency spectrum}
\usage{
joint_sfs(afs, pref = NULL)
}
\arguments{
\item{afs}{A named list of length two where the first element (\code{afs}) contains the allele frequency matrix,
and the second element (\code{counts}) contains the allele count matrix.}

\item{pref}{Instead of \code{afs}, the prefix of genotype files can be provided.}
}
\value{
A data frame with the number of times each possible combination of allele counts is observed.
}
\description{
This function computes the joint site frequency spectrum from genotype files or allele frequency and count matrices.
The joint site frequency spectrum lists how often each combination of haplotypes is observed.
The number of combinations is equal to the product of one plus the number of haplotypes in each population.
For example, five populations with a single diploid individual each have 3^5 possible combinations.
}
