% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{joint_spectrum}
\alias{joint_spectrum}
\title{Estimate joint allele frequency spectrum}
\usage{
joint_spectrum(afs, ...)
}
\arguments{
\item{afs}{A matrix or data frame of allele frequencies}
}
\value{
A data frame with columns \code{pattern} and \code{proportion}
}
\description{
Estimate joint allele frequency spectrum
}
\examples{
\dontrun{
dat = plink_to_afs('/my/plink/file', pops = c('pop1', 'pop2', 'pop3', 'pop4', 'pop5'))

# Spectrum across all SNPs
joint_spectrum(dat$afs)

# Stratify by allele frequency in one population
dat$afs \%>\% as_tibble \%>\% select(1:4) \%>\%
  group_by(grp = cut(pop1, 10)) \%>\%
  group_modify(joint_spectrum) \%>\%
  ungroup

# Stratify by mutation class
dat$afs \%>\% as_tibble \%>\% select(1:4) \%>\%
  mutate(mut = paste(dat$snpfile$A1, dat$snpfile$A2)) \%>\%
  group_by(mut) \%>\%
  group_modify(joint_spectrum) \%>\%
  ungroup
}
}
