% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{lazadm}
\alias{lazadm}
\title{Estimate admixture weights}
\usage{
lazadm(target, left, right, f2_blocks = NULL, block_lengths = NULL,
  f2_dir = NULL, f2_denom = 1, getcov = FALSE, constrained = TRUE)
}
\arguments{
\item{target}{target population}

\item{left}{source populations}

\item{right}{outgroup populations}

\item{f2_blocks}{3d array of block-jackknife leave-one-block-out estimates of f2 statistics. output of \code{\link{afs_to_f2_blocks}}}

\item{block_lengths}{the jackknife block lengths used in computing the f2 statistics. see \code{\link{get_block_lengths}}.}

\item{f2_dir}{a directory with f2 statistics for each population pair in the graph. must contain \code{block_lengths.RData}.}

\item{f2_denom}{scales f2-statistics. A value of around 0.278 converts F2 to Fst.}

\item{getcov}{should standard errors be returned? Currently not implemented.}

\item{constrained}{if \code{TRUE} (default), admixture weights will all be positive. if \code{FALSE}, they can be negative, as in \code{\link{qpadm}}}
}
\value{
a data frame with weights and standard errors for each left population
}
\description{
Models target as a mixture of left populations, and outgroup right populations. Uses Lazaridis method based non-negative least squares of f4 matrix.
}
\examples{
target = 'Denisova.DG'
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
lazadm(target, left, right, example_f2_blocks, example_block_lengths)
lazadm(target, left, right, example_f2_blocks, example_block_lengths, constrained = FALSE)
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics

Haak, W. et al. (2015) \emph{Massive migration from the steppe was a source for Indo-European languages in Europe.} Nature (SI 9)
}
\seealso{
\code{\link{qpadm}}
}
