% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{loo_list}
\alias{loo_list}
\title{Generate a list of leave-one-out arrays}
\usage{
loo_list(arr)
}
\arguments{
\item{arr}{3d array with blocked estimates, with blocks in the 3rd dimension}
}
\value{
A list of leave-one-out arrays, each of which is 1 element shorter than \code{arr} along the 3rd dimension
}
\description{
Generate a list of leave-one-out arrays
}
\seealso{
\code{\link{boo_list}} \code{\link{est_to_loo}}
}
