% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{make_resample_inds_fun}
\alias{make_resample_inds_fun}
\title{Takes a function \code{qpfun} which takes f2_blocks as input
Returns a function which will repeadetly evaluate \code{qpfun} on
a subset of all samples, returning a nested data frame}
\usage{
make_resample_inds_fun(qpfun)
}
\description{
Takes a function \code{qpfun} which takes f2_blocks as input
Returns a function which will repeadetly evaluate \code{qpfun} on
a subset of all samples, returning a nested data frame
}
