% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{move_admixedge_once}
\alias{move_admixedge_once}
\title{Modify a graph by moving an admixture edge}
\usage{
move_admixedge_once(graph, desimplify = TRUE)
}
\arguments{
\item{graph}{An admixture graph}

\item{desimplify}{Desimplify graph (\code{\link{desimplify_graph}})}
}
\value{
A new admixture graph
}
\description{
Modify a graph by moving an admixture edge
}
