% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{msprime_sim}
\alias{msprime_sim}
\title{Simulate an admixture graph in msprime v1.x}
\usage{
msprime_sim(
  graph,
  outpref = "msprime_sim",
  nsnps = 1000,
  neff = 1000,
  ind_per_pop = 1,
  mutation_rate = 0.001,
  time = 1000,
  fix_leaf = FALSE,
  admix_default = 0.5,
  run = FALSE,
  numcores = NULL,
  ghost_lineages = FALSE,
  shorten_admixed_leaves = FALSE
)
}
\arguments{
\item{graph}{A graph as an \code{igraph} object or edge list with columns ’from’ and ’to’. If it is an edge list with a column ’weight’ (derived possibly from a fitted graph), the admixture weights will be used. Otherwise, all admixture edges will have a weight of 0.5.}

\item{outpref}{A prefix of output files.}

\item{nsnps}{The number of SNPs to simulate. All SNPs will be simulated independently of each other.}

\item{neff}{Effective population size (in diploid individuals). If a scalar value, it will be constant across all populations. Alternatively, it can be a named vector with a different value for each population (e.g., \code{c('R'=100, 'A'=50, 'B'=50)}).}

\item{ind_per_pop}{The number of diploid individuals to simulate for each population. If a scalar value, it will be constant across all populations.
Alternatively, it can be a named vector with a different value for each population (e.g., \code{c('A'=10, 'B'=20)} to sample 10 and 20 diploid individuals from populations A and B, respectively).}

\item{mutation_rate}{Mutation rate per site per generation. The default is set to a high value (0.001 per site per generation) to obtain more polymorphic SNPs in order to speed up the simulation.}

\item{time}{Either a scalar value (1000 generations by default) with the dates generated by \code{\link{pseudo_dates}} function, or a named vector with dates for each graph nodes (in generations).}

\item{fix_leaf}{A boolean value specifying if the dates  of the leaf nodes will be fixed at time 0. If \code{TRUE}, all samples will be drawn at the end of the simulation (i.e., from “today”).}

\item{admix_default}{A float value specifying default admixture proportion for all admixture nodes. The default is \code{0.5}.
If another value between 0 and 1 is specified, admixture weights for each admixture event will be (value, 1-value).}

\item{run}{If \code{FALSE}, the function will terminate after writing the msprime script. If \code{TRUE}, it will try to execute the msprime script with the default python installation. If you want to use some other python installation, you can set \verb{run = /my/python}.}

\item{numcores}{The number of cores to use when simulating data.}

\item{ghost_lineages}{A boolean value specifying whether ghost lineages will be allowed.
If \code{TRUE}, admixture happens at the time points defined by the y-axis generated while plotting the graph by \code{\link{plot_graph}}.
If \code{FALSE} (default), admixture occurs at the time of the previous split event.}

\item{shorten_admixed_leaves}{If \code{TRUE} simulate the behavior of treemix where drift after admixture is not allowed}
}
\value{
The file name and path of the simulation script
}
\description{
This function generates an msprime simulation script, and optionally executes it in python. Unlike the \code{\link{msprime_genome}} function, this function simulates independent SNP sites.
}
\examples{
\dontrun{
results = qpgraph(example_f2_blocks, example_graph)
msprime_sim(results$edges, nsnps=100)
}
}
