% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{mutate_n}
\alias{mutate_n}
\title{Modify a graph by applying n mutation functions}
\usage{
mutate_n(
  graph,
  n = 2,
  funs = list(spr_all, spr_leaves, swap_leaves, move_admixedge_once, flipadmix_random,
    mutate_n),
  desimplify = TRUE,
  fix_outgroup = TRUE
)
}
\arguments{
\item{graph}{An admixture graph}

\item{n}{Number of functions to apply}

\item{funs}{List of function from which to choose}

\item{desimplify}{Desimplify graph (\code{\link{desimplify_graph}})}

\item{fix_outgroup}{Keep outgroup in place}
}
\value{
A new admixture graph
}
\description{
Modify a graph by applying n mutation functions
}
