% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{namedList}
\alias{namedList}
\title{Create a named list from arguments}
\usage{
namedList(...)
}
\arguments{
\item{...}{Objects to include in the list.}
}
\value{
A named list.
}
\description{
Like \code{list()} but uses argument names as list names when names are not
explicitly supplied.
}
