% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{newick_to_edges}
\alias{newick_to_edges}
\title{Turn a newick format tree to a matrix of edges}
\usage{
newick_to_edges(newick, node = "R", edgemat = matrix(NA, 0, 2))
}
\arguments{
\item{newick}{Tree in newick format.}

\item{node}{Root label of the tree.}

\item{edgemat}{Used for recursive function calls.}
}
\value{
Tree as two column matrix of edges (adjacency list)
}
\description{
Turn a newick format tree to a matrix of edges
}
\examples{
newick = random_newick(c('a', 'b', 'c', 'd'))
newick
newick_to_edges(newick)
}
