% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{node_counts}
\alias{node_counts}
\title{Count how often each node in graph occurs in other graphs}
\usage{
node_counts(graph, graphlist)
}
\arguments{
\item{graph}{An admixture graph}

\item{graphlist}{List of graphs}
}
\description{
Count how often each node in graph occurs in other graphs
}
