% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{node_signature}
\alias{node_signature}
\title{Returns a signature of a graph consisting of the left and right descendent leaf nodes of each internal node (sorted and concatenated)}
\usage{
node_signature(graph)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
A graph signature as character vector
}
\description{
Can be used to determine how often internal nodes occur in a list of other well fitting models
}
\examples{
\dontrun{
sigs = example_winners \%>\% mutate(sig = map(igraph, node_signature)) \%$\%
         sig \%>\% unlist \%>\% table \%>\% c
node_signature(example_winners$igraph[[1]])
}
}
