% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{numadmix}
\alias{numadmix}
\title{Count number of admixture nodes}
\usage{
numadmix(graph)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
Number of admixture nodes
}
\description{
Count number of admixture nodes
}
