% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{optimize_admixturegraph}
\alias{optimize_admixturegraph}
\title{Find a well fitting admixturegraph}
\usage{
optimize_admixturegraph(pops, f3_jest, ppinv, numrep = 10,
  numgraphs = 50, numgen = 5, numsel = 5, numadmix = 0,
  outpop = NA, verbose = TRUE, cpp = TRUE, debug = FALSE, ...)
}
\arguments{
\item{pops}{population names for which to fit admixturegraphs}

\item{f3_jest}{3d array with all pairwise f3 estimates with the outgroup}

\item{ppinv}{inverse of the f3 block jackknife covariance matrix}

\item{numgraphs}{number of graphs in each generation}

\item{numgen}{number of generations}

\item{numsel}{number of graphs which are selected in each generation}

\item{numadmix}{number of admixture events within each graph}

\item{outpop}{outgroup population}
}
\value{
a nested data frame with one model per line
}
\description{
This function generates and evaluates admixturegraphs in \code{numgen} iterations across \code{numrep} independent repeats to find well fitting admixturegraphs. It uses the function \code{\link{future_map}} from the \code{\link{furrr}} package to parallelize across the independent repeats. The function \code{\link{future::plan}} can be called to specify the details of the parallelization. This can be used to parallelize across cores or across nodes on a compute cluster.
}
\examples{
\dontrun{
optimize_admixturegraph(pops, f3_jest, ppinv, numrep=200, numgraphs=100, numgen=20, numsel=5, numadmix=3)
}
}
