% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{optimize_admixturegraph_single}
\alias{optimize_admixturegraph_single}
\title{Find well fitting admixturegraphs}
\usage{
optimize_admixturegraph_single(pops, f3_jest, ppinv, numgraphs = 50,
  numgen = 5, numsel = 5, numadmix = 0, outpop = NA,
  verbose = TRUE, cpp = TRUE, debug = FALSE, keep = "all", ...)
}
\arguments{
\item{pops}{population names for which to fit admixturegraphs}

\item{f3_jest}{3d array with all pairwise f3 estimates with the outgroup}

\item{ppinv}{inverse of the f3 block jackknife covariance matrix}

\item{numgraphs}{number of graphs in each generation}

\item{numgen}{number of generations}

\item{numsel}{number of graphs which are selected in each generation}

\item{numadmix}{number of admixture events within each graph}

\item{outpop}{outgroup population}
}
\value{
a nested data frame with one model per line
}
\description{
This function generates and evaluates admixturegraphs in \code{numgen} iterations to find well fitting admixturegraphs.
}
\examples{
pops = get_leafnames(igraph1)
optimize_admixturegraph_single(igraph1, f3_jest, ppinv, numgraphs=10, numgen=2, numsel=2, numadmix=1)
}
