% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{packedancestrymap_to_plink}
\alias{packedancestrymap_to_plink}
\title{Convert packedancestrymap to PLINK}
\usage{
packedancestrymap_to_plink(inpref, outpref, inds = NULL, pops = NULL)
}
\arguments{
\item{inpref}{Prefix of the packedancestrymap input files}

\item{outpref}{Prefix of the PLINK output files}

\item{inds}{Individuals which should be extracted}

\item{pops}{Populations which should be extracted. Can not be provided together with 'inds'}
}
\description{
This function converts \emph{packedancestrymap} format files to \emph{PLINK files}, using \code{\link[genio]{write_plink}}.
When \code{inds} or \code{pops} is provided, only a subset of samples will be extracted.
}
