% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parse_dot}
\alias{parse_dot}
\title{Read graph in dot format}
\usage{
parse_dot(dotfile)
}
\arguments{
\item{dotfile}{Name of a file with a dot formatted admixture graph}
}
\description{
Read graph in dot format
}
\examples{
\dontrun{
graph = parse_dot("/my/graph.dot")
}
}
