% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{parse_fstats}
\alias{parse_fstats}
\title{Parse qpGraph fstats output file}
\usage{
parse_fstats(outfile, denom1 = 1000, denom2 = 1e+07)
}
\arguments{
\item{outfile}{Path to the qpGraph fstats output file.}

\item{denom1}{Scaling denominator for f3-statistics.}

\item{denom2}{Scaling denominator for f4-statistics.}
}
\value{
A list of tibbles with parsed f-statistics.
}
\description{
Parses the fstats output file written by the original qpGraph program.
}
