% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parse_qp3pop_output}
\alias{parse_qp3pop_output}
\title{Read qp3Pop output file}
\usage{
parse_qp3pop_output(outfile)
}
\arguments{
\item{outfile}{output file generated by qp3Pop.}
}
\value{
tibble with output data.
}
\description{
Read qp3Pop output file
}
