% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parse_qpadm_output}
\alias{parse_qpadm_output}
\title{Reads qpAdm output file}
\usage{
parse_qpadm_output(outfile)
}
\arguments{
\item{outfile}{output file generated by qpAdm.}
}
\value{
tibble with output data.
}
\description{
Reads qpAdm output file
}
