% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parse_qpdstat_output}
\alias{parse_qpdstat_output}
\title{Read qpDstat output file}
\usage{
parse_qpdstat_output(outfile)
}
\arguments{
\item{outfile}{output file generated by qpDstat.}
}
\value{
tibble with output data.
}
\description{
Read qpDstat output file
}
