% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parse_qpdstat_parfile}
\alias{parse_qpdstat_parfile}
\title{Read qpDstat parameter file}
\usage{
parse_qpdstat_parfile(parfile)
}
\arguments{
\item{parfile}{parameter file for qpDstat.}
}
\value{
named list of qpGraph parameters.
}
\description{
Read qpDstat parameter file
}
