% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{parse_qpf4ratio_output}
\alias{parse_qpf4ratio_output}
\title{Read qpF4ratio output file}
\usage{
parse_qpf4ratio_output(parfile)
}
\arguments{
\item{outfile}{Output file generated by qpF4ratio}
}
\value{
Data frame with output data
}
\description{
Read qpF4ratio output file
}
