% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{parse_qpgraph_output}
\alias{parse_qpgraph_output}
\title{Read qpGraph output file}
\usage{
parse_qpgraph_output(outfile, logfile = outfile)
}
\arguments{
\item{outfile}{Output file generated by qpGraph.}

\item{logfile}{Log file for reading score and outliers; defaults to \code{outfile}.}
}
\value{
list of output data.
}
\description{
Read qpGraph output file
}
