% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parse_qpgraph_parfile}
\alias{parse_qpgraph_parfile}
\title{Reads qpGraph parameter file}
\usage{
parse_qpgraph_parfile(parfile)
}
\arguments{
\item{parfile}{parameter file for qpGraph.}
}
\value{
named list of qpGraph parameters.
}
\description{
Reads qpGraph parameter file
}
