% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{permute_leaves}
\alias{permute_leaves}
\title{Modify a graph by permuting leaf nodes}
\usage{
permute_leaves(graph, desimplify = TRUE, fix_outgroup = TRUE)
}
\arguments{
\item{graph}{An admixture graph}

\item{fix_outgroup}{Keep outgroup in place}
}
\value{
A new admixture graph
}
\description{
Modify a graph by permuting leaf nodes
}
