% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{place_root_random}
\alias{place_root_random}
\title{Modify a graph by changing the position of the root node}
\usage{
place_root_random(graph, fix_outgroup = TRUE)
}
\arguments{
\item{graph}{An admixture graph}

\item{fix_outgroup}{Keep outgroup in place}
}
\value{
A new admixture graph
}
\description{
Modify a graph by changing the position of the root node
}
