% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{plink_to_afs}
\alias{plink_to_afs}
\title{Read allele frequencies from \code{PLINK} files}
\usage{
plink_to_afs(
  pref,
  inds = NULL,
  pops = NULL,
  adjust_pseudohaploid = TRUE,
  first = 1,
  last = NULL,
  numblocks = 1,
  poly_only = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{pref}{prefix of \code{PLINK} files (files have to end in \code{.bed}, \code{.bim}, \code{.fam}).}

\item{inds}{Individuals from which to compute allele frequencies}

\item{pops}{Populations from which to compute allele frequencies. If \code{NULL} (default), populations will be extracted from the third column in the \code{.ind} file. If population labels are provided, they should have the same length as \code{inds}, and will be matched to them by position}

\item{adjust_pseudohaploid}{Genotypes of pseudohaploid samples are usually coded as \code{0} or \code{2}, even though only one allele is observed. \code{adjust_pseudohaploid} ensures that the observed allele count increases only by \code{1} for each pseudohaploid sample. If \code{TRUE} (default), samples that don't have any genotypes coded as \code{1} among the first 1000 SNPs are automatically identified as pseudohaploid. This leads to slightly more accurate estimates of f-statistics. Setting this parameter to \code{FALSE} is equivalent to the ADMIXTOOLS \code{inbreed: NO} option. Setting \code{adjust_pseudohaploid} to an integer \code{n} will check the first \code{n} SNPs instead of the first 1000 SNPs.}

\item{numblocks}{Number of blocks in which to read genotype file. Setting this to a number
greater than one is more memory efficient, but slower.}

\item{poly_only}{Only keep SNPs with mean allele frequency not equal to 0 or 1.}

\item{verbose}{Print progress updates}
}
\value{
A list with three items: Allele frequency matrix, allele count matrix, and SNP meta data.
}
\description{
Read allele frequencies from \code{PLINK} files
}
\examples{
\dontrun{
afdat = plink_to_afs(prefix, pops)
afs = afdat$afs
counts = afdat$counts
}
}
