% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{plink_to_aftable}
\alias{plink_to_aftable}
\title{Read allele frequencies from \code{PLINK} files}
\usage{
plink_to_aftable(pref, inds = NULL, pops = NULL, verbose = FALSE)
}
\arguments{
\item{pref}{prefix of \code{PLINK} files (files have to end in \code{.bed}, \code{.bim}, \code{.fam}).}

\item{inds}{vector of samples from which to compute allele frequencies.}

\item{pops}{vector of populations from which to compute allele frequencies. If \code{NULL} (default), populations will be extracted from the third column in the \code{.ind} file}

\item{verbose}{print progress updates}
}
\value{
a list with two items: allele frequency data and individual counts.
}
\description{
Read allele frequencies from \code{PLINK} files
}
\examples{
\dontrun{
afdat = plink_to_aftable(prefix, pops)
afs = afdat$afs
counts = afdat$counts
}
}
