% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot an admixture graph}
\usage{
plot_graph(edges, fix = TRUE, title = "", color = TRUE)
}
\arguments{
\item{edges}{an admixture graph. If it's an edge list matrix or data frame with a \code{label} column, those values will displayed on the edges}

\item{title}{plot title}
}
\value{
a ggplot object
}
\description{
Plot an admixture graph
}
\examples{
plot_graph(graph1)
}
