% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_graph_map}
\alias{plot_graph_map}
\title{Plots an admixture graph on a map.}
\usage{
plot_graph_map(grph, leafcoords, shapedata = NULL)
}
\arguments{
\item{grph}{a two column matrix specifying the admixture graph. first column is source node, second column is target node. the first edge has to be root -> outgroup. admixture nodes are inferred as those nodes which are the targets of two different sources.}

\item{leafcoords}{data frame with columns \code{group}, \code{lon}, \code{lat}}
}
\value{
a plotly object.
}
\description{
Plots an admixture graph on a map.
}
\examples{
\dontrun{
plot_graph_map(igraph1, anno_v41)
}
}
