% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_graph_map2}
\alias{plot_graph_map2}
\title{Plots an admixture graph on a map.}
\usage{
plot_graph_map2(grph, leafcoords, map_layout = 1)
}
\arguments{
\item{grph}{an admixturegraph as an igraph object.}

\item{leafcoords}{data frame with columns \code{group}, \code{lon}, \code{lat}}

\item{map_layout}{1 or 2}
}
\value{
a ggplot object.
}
\description{
Plots an admixture graph on a map.
}
\examples{
\dontrun{
p1 = plot_graph_map2(igraph1, anno_v41, 1)
p2 = plot_graph_map2(igraph1, anno_v41, 2)
}
}
