% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_graph_old}
\alias{plot_graph_old}
\title{Plots an admixture graph.}
\usage{
plot_graph_old(edges, layout = "tree", fix = TRUE, title = "")
}
\arguments{
\item{edges}{a two column matrix specifying the admixture graph. first column is source node, second column is target node. the first edge has to be root -> outgroup. admixture nodes are inferred as those nodes which are the targets of two different sources.}

\item{layout}{argument passed to \code{\link[ggdag]{tidy_dagitty}} to modify plot layout.}
}
\value{
a ggplot object.
}
\description{
Plots an admixture graph.
}
