% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_map}
\alias{plot_map}
\title{Plot samples on a map}
\usage{
plot_map(leafcoords, map_layout = 1, color = "yearsbp",
  colorscale = "Portland", collog10 = TRUE)
}
\arguments{
\item{leafcoords}{data frame with columns \code{iid}, \code{lon}, \code{lat}}

\item{map_layout}{1 or 2}

\item{color}{color}

\item{colorscale}{colorscale}

\item{collog10}{collog10}
}
\value{
a plotly object.
}
\description{
Plot samples on a map
}
\examples{
\dontrun{
plot_map(example_anno, 1)
}
}
