% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotly_comparison}
\alias{plotly_comparison}
\title{Compare two models}
\usage{
plotly_comparison(out1, out2, name1 = NULL, name2 = NULL)
}
\arguments{
\item{out1}{First model}

\item{out2}{Second model}

\item{name1}{Optional name for first model}

\item{name2}{Optional name for second model}
}
\value{
A plotly object.
}
\description{
Plot a comparison of two runs of qpgraph or two runs of qpadm
}
\examples{
fit1 = qpgraph(example_f2_blocks, example_graph, lsqmode = FALSE)
fit2 = qpgraph(example_f2_blocks, example_graph, lsqmode = TRUE)
plotly_comparison(fit1, fit2)
}
