% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotly_graph}
\alias{plotly_graph}
\title{Plot an admixture graph using plotly}
\usage{
plotly_graph(graph, collapse_threshold = 0, fix = FALSE,
  shift_down = TRUE)
}
\arguments{
\item{graph}{an admixture graph}

\item{fix}{if \code{TRUE}, there will be an attempt to rearrange the nodes to minimize
the number of intersecting edges. This can take very long for large graphs.
By default this is only done for graphs with fewer than 10 leaves.}

\item{shift_down}{shift descendent nodes down}
}
\value{
a plotly object
}
\description{
Plot an admixture graph using plotly
}
\examples{
plotly_graph(example_graph)
}
