% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qp3pop}
\alias{qp3pop}
\alias{f3}
\title{Estimate f3 statistics}
\usage{
qp3pop(
  data,
  pop1 = NULL,
  pop2 = NULL,
  pop3 = NULL,
  boot = FALSE,
  sure = FALSE,
  unique_only = TRUE,
  blgsize = NULL,
  block_lengths = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Input data in one of three forms: \enumerate{
\item A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{f2_from_geno}} (fastest option)
\item A directory which contains pre-computed f2-statistics
\item The prefix of genotype files (slowest option)
}}

\item{pop1}{One of the following four:
\enumerate{
\item \code{NULL}: all possible population combinations will be returned
\item A vector of population labels. All combinations with the other \code{pop} arguments will be returned
\item A matrix with population combinations to be tested, with one population per column and one
combination per row. Other \code{pop} arguments will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or
population combinations to be tested. Other \code{pop} arguments will be ignored.
}}

\item{pop2}{A vector of population labels}

\item{pop3}{A vector of population labels}

\item{boot}{If \code{FALSE} (the default), block-jackknife resampling will be used to compute standard errors.
Otherwise, block-bootstrap resampling will be used to compute standard errors. If \code{boot} is an integer, that number
will specify the number of bootstrap resamplings. If \code{boot = TRUE}, the number of bootstrap resamplings will be
equal to the number of SNP blocks.}

\item{sure}{The number of population combinations can get very large. This is a safety option that stops you
from accidently computing all combinations if that number is large.}

\item{unique_only}{If \code{TRUE} (the default), redundant combinations will be excluded}

\item{verbose}{Print progress updates}

\item{...}{Additional arguments passed to \code{\link{f3blockdat_from_geno}} if \code{data} is a genotype prefix, or to \code{\link{get_f2}} otherwise}
}
\value{
\code{qp3pop} returns a data frame with f3 statistics
}
\description{
Computes f3 statistics of the form \eqn{f3(A; B, C)}. When using the same SNPs for all populations, this is equivalent to
\eqn{(f2(A, B) + f2(A, C) - f2(B, C)) / 2} and to \eqn{f4(A, B; A, C)}
}
\details{
There are several arguments that can be passed via ... which affect the estimated f3-statistics.
The default options are the same as in the original qp3pop program,
but some options are not effective when using precomputed f2-statistics. See \code{examples} for more information.
}
\section{Alias}{

\code{f3}
}

\examples{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
pop3 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
qp3pop(example_f2_blocks, pop1, pop2, pop3)

pops = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
qp3pop(example_f2_blocks, pops)
qp3pop(example_f2_blocks, pops, unique_only = FALSE)
\dontrun{
qp3pop(f2_dir, pop1, pop2, pop3)

Below are three scenarios, and in each one `qp3pop()` and `qp3pop_wrapper()`
should give the same or very similar estimates. Note that to compute `f3(A; B, C)`,
`qp3pop_wrapper()` expects the populations to be in the order `B`, `C`, `A`.

prefix = '/path/to/geno/prefix'
qp3popbin = '/path/to/AdmixTools/bin/qp3Pop'
pops = dimnames(example_f2_blocks)[[1]]
qp3pop_wrapper(prefix, pops[2], pops[3], pops[1], bin = qp3popbin, outgroupmode = FALSE)
qp3pop(prefix, pops[1], pops[2], pops[3])
qp3pop(prefix, pops[1], pops[2], pops[3], poly_only = TRUE)
qp3pop_wrapper(prefix, pops[2], pops[3], pops[1], bin = qp3popbin, outgroupmode = TRUE)
qp3pop(prefix, pops[1], pops[2], pops[3], outgroupmode = TRUE)
f2b = f2_from_geno(prefix, pops = pops[1:3], poly_only = FALSE)
qp3pop(f2b, pops[1], pops[2], pops[3])
qp3pop_wrapper(prefix, pops[1], pops[3], pops[2], bin = qp3popbin, outgroupmode = TRUE)
qp3pop(prefix, pops[2], pops[1], pops[3], outgroupmode = TRUE, apply_corr = FALSE)
}
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history} Genetics

Peter, B. (2016) \emph{Admixture, Population Structure, and F-Statistics} Genetics
}
