% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qp3pop_wrapper}
\alias{qp3pop_wrapper}
\title{Wrapper function around the original qp3Pop program}
\usage{
qp3pop_wrapper(source1, source2 = NULL, target = NULL,
  bin = "./qp3Pop", pref = NULL, outdir = ".", parfile = NULL,
  inbreed = "NO", outgroupmode = "YES", useallsnps = "NO",
  printonly = FALSE, env = "")
}
\arguments{
\item{source1}{one of the following four:
\enumerate{
\item \code{NULL}: populations will be read from \code{poplistname} or \code{popfilename} specified in \code{parfile}
\item a vector of population labels
\item a data frame in which the first four columns specify the population triples to be tested. \code{source2}, \code{target} will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or population quadruples to be tested. \code{source2} and \code{target} will be ignored.
}}

\item{source2}{a vector of population labels}

\item{target}{a vector of population labels}

\item{bin}{path to the qp3Pop binary file}

\item{pref}{path to and prefix of the packedancestrymap genotype files}

\item{outdir}{the output directory. files \code{out}, \code{parfile}, \code{poplistname}, \code{popfilename} may be overwritten}

\item{parfile}{qp3Pop parameter file. If this is specified, \code{source1}, \code{source2}, \code{target} will be ignored.}

\item{printonly}{should the command be printed or executed?}
}
\value{
If \code{printonly}, the \code{qp3Pop} command, otherwise a data frame with parsed \code{qp3Pop} output
}
\description{
Computes f3 statistics of the form \eqn{F_3(A; B, C)}. Equivalent to \eqn{(F_2(A, B) + F_2(A, C) - F_2(B, C)) / 2} and to \eqn{F_4(A, B; A, C)}. Requires a working installation of qp3Pop, which will be called using \code{\link{system}}
}
\examples{
\dontrun{
target = 'Denisova.DG'
source1 = c('Altai_Neanderthal.DG', 'Vindija.DG')
source2 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
qp3pop_wrapper(source1, source2, target,
  bin = 'path/to/qp3Pop', pref = 'path/to/packedancestrymap_prefix',
  env = 'export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/path/to/blas/')

qp3pop_wrapper(bin = 'path/to/qp3Pop', parfile = 'path/to/parfile')
}
}
