% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{qpadm_p}
\alias{qpadm_p}
\title{Faster version of \code{\link{qpadm}} with reduced output}
\usage{
qpadm_p(
  f2_data,
  left,
  right,
  target = NULL,
  fudge = 1e-04,
  boot = FALSE,
  constrained = FALSE,
  rnk = length(setdiff(left, target)) - 1,
  cpp = TRUE,
  f4blocks = NULL
)
}
\arguments{
\item{left}{Left populations (sources)}

\item{right}{Right populations (outgroups)}

\item{target}{Target population}

\item{fudge}{Value added to diagonal matrix elements before inverting}

\item{boot}{If \code{FALSE} (the default), block-jackknife resampling will be used to compute standard errors.
Otherwise, block-bootstrap resampling will be used to compute standard errors. If \code{boot} is an integer, that number
will specify the number of bootstrap resamplings. If \code{boot = TRUE}, the number of bootstrap resamplings will be
equal to the number of SNP blocks.}

\item{constrained}{Constrain admixture weights to be non-negative}

\item{rnk}{Rank of f4-matrix. Defaults to one less than full rank.}

\item{cpp}{Use C++ functions. Setting this to \code{FALSE} will be slower but can help with debugging.}

\item{f4blocks}{Instead of f2 blocks, f4 blocks can be supplied. This is used by \code{\link{qpadm_multi}}}
}
\value{
Data frame with \code{f4rank}, \code{dof}, \code{chisq}, \code{p}, \code{feasible}
}
\description{
Models target as a mixture of left populations, given a set of outgroup right populations.
Can be used to estimate admixture proportions, and to estimate the number of independent
admixture events.
}
\examples{
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
target = 'Denisova.DG'
qpadm_p(example_f2_blocks, left, right, target)
}
\seealso{
\code{\link{qpadm}}
}
