% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{qpadm_pairs}
\alias{qpadm_pairs}
\title{Compute all pairwise qpadm p-values}
\usage{
qpadm_pairs(f2_blocks, left, right)
}
\arguments{
\item{f2_blocks}{a 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}}.}

\item{left}{left populations}

\item{right}{right populations}
}
\description{
For all pairs of left populations qpadm Chi-squared statistics and p-values will be computed
}
