% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpadm_wrapper}
\alias{qpadm_wrapper}
\title{Wrapper function around the original qpAdm program}
\usage{
qpadm_wrapper(left, right, target = NULL, bin = "~np29/o2bin/qpAdm",
  pref = NULL, outdir = "./", parfile = NULL, useallsnps = "NO",
  fancyf4 = "YES", f4mode = "YES", inbreed = "NO",
  printonly = FALSE, env = "", verbose = TRUE)
}
\arguments{
\item{left}{Left populations (or leftlist file)}

\item{right}{Right populations (or rightlist file)}

\item{target}{Target population}

\item{bin}{Path to the qpAdm binary file}

\item{pref}{Path to and prefix of the packedancestrymap genotype files}

\item{outdir}{Output directory. files \code{out}, \code{parfile}, \code{leftlist},
\code{rightlist} will be overwritten}

\item{printonly}{Should the command be printed or executed?}

\item{verbose}{Print progress updates}
}
\value{
If not printonly, a data frame with parsed qpAdm output
}
\description{
This requires a working installation of qpAdm, which will be called using \code{\link{system}}
}
\examples{
\dontrun{
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
target = 'Denisova.DG'
qpadm_wrapper(left, right, target,
  bin = 'path/to/qpAdm', pref = 'path/to/packedancestrymap_prefix',
  env = 'export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/path/to/blas/')
}
}
