% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qpdstat}
\alias{qpdstat}
\alias{f4}
\title{Estimate f4 statistics}
\usage{
qpdstat(
  data,
  pop1 = NULL,
  pop2 = NULL,
  pop3 = NULL,
  pop4 = NULL,
  boot = FALSE,
  sure = FALSE,
  unique_only = TRUE,
  comb = TRUE,
  blgsize = NULL,
  block_lengths = NULL,
  f4mode = TRUE,
  afprod = TRUE,
  cpp = TRUE,
  verbose = is.character(data),
  ...
)
}
\arguments{
\item{data}{Input data in one of three forms: \enumerate{
\item A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{extract_f2}} (fastest option)
\item A directory which contains pre-computed f2-statistics
\item The prefix of genotype files (slowest option)
}}

\item{pop1}{One of the following four:
\enumerate{
\item \code{NULL}: all possible population combinations will be returned
\item A vector of population labels. All combinations with the other \code{pop} arguments will be returned
\item A matrix with population combinations to be tested, with one population per column and one
combination per row. Other \code{pop} arguments will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or
population combinations to be tested. Other \code{pop} arguments will be ignored.
}}

\item{pop2}{A vector of population labels}

\item{pop3}{A vector of population labels}

\item{pop4}{A vector of population labels}

\item{boot}{If \code{FALSE} (the default), block-jackknife resampling will be used to compute standard errors.
Otherwise, block-bootstrap resampling will be used to compute standard errors. If \code{boot} is an integer, that number
will specify the number of bootstrap resamplings. If \code{boot = TRUE}, the number of bootstrap resamplings will be
equal to the number of SNP blocks.}

\item{sure}{The number of population combinations can get very large. This is a safety option that stops you
from accidently computing all combinations if that number is large.}

\item{unique_only}{If \code{TRUE} (the default), redundant combinations will be excluded}

\item{comb}{Generate all combinations of \code{pop1}, \code{pop2}, \code{pop3}, \code{pop4}. If \code{FALSE}, \code{pop1}, \code{pop2}, \code{pop3}, \code{pop4} should all be vectors of the same length.}

\item{blgsize}{SNP block size in Morgan. Default is 0.05 (50 cM). Only used when \code{data} is the prefix of genotype files}

\item{block_lengths}{Vector with lengths of each jackknife block. \code{sum(block_lengths)} has to
match the number of SNPs. only used when \code{data} is the prefix of genotype files}

\item{f4mode}{If \code{TRUE}: f4 is computed from allele frequencies \code{a}, \code{b}, \code{c}, and \code{d} as \code{(a-b)*(c-d)}. if \code{FALSE}, D-statistics are computed instead, defined as \code{(a-b)*(c-d) / ((a + b - 2*a*b) * (c + d - 2*c*d))}, which is the same as \code{(P(ABBA) - P(BABA)) / (P(ABBA) + P(BABA))}. \code{f4mode = FALSE} is only available when \code{data} is the prefix of genotype files}

\item{afprod}{Compute f4 from allele frequency products instead of f2. Only used if \code{data} is a directory with precomputed data.}

\item{cpp}{Use C++ functions. Setting this to \code{FALSE} will be slower but can help with debugging.}

\item{verbose}{Print progress updates}

\item{...}{Additional arguments passed to \code{\link{f4blockdat_from_geno}} if \code{data} is a genotype file prefix or \code{\link{f2_from_precomp}} if \code{data} is a directory with f2-statistics}
}
\value{
\code{qpdstat} returns a data frame with f4 statistics
}
\description{
Computes f4 statistics from f2 blocks of the form \eqn{f4(A, B; C, D)}. Equivalent to
\eqn{(f2(A, D) + f2(B, C) - f2(A, C) - f2(B, D)) / 2}
}
\section{Alias}{

\code{f4}
}

\examples{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
pop3 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
pop4 = 'Switzerland_Bichon.SG'
qpdstat(example_f2_blocks, pop1, pop2, pop3, pop4)
\dontrun{
qpdstat(f2_dir, pop1, pop2, pop3, pop4)
}
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history} Genetics

Peter, B. (2016) \emph{Admixture, Population Structure, and F-Statistics} Genetics
}
