% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qpdstat_wrapper}
\alias{qpdstat_wrapper}
\title{Wrapper function around the original qpDstat program}
\usage{
qpdstat_wrapper(pop1 = NULL, pop2 = NULL, pop3 = NULL, pop4 = NULL,
  bin = "./qpDstat", pref = NULL, outdir = ".", parfile = NULL,
  f4mode = "YES", printonly = FALSE, env = "")
}
\arguments{
\item{pop1}{one of the following four:
\enumerate{
\item \code{NULL}: populations will be read from \code{poplistname} or \code{popfilename} specified in \code{parfile}
\item a vector of population labels
\item a data frame in which the first four columns specify the population quadruples to be tested. \code{pop2}, \code{pop3}, \code{pop4} will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or population quadruples to be tested. \code{pop2}, \code{pop3}, \code{pop4} will be ignored.
}}

\item{pop2}{a vector of population labels}

\item{pop3}{a vector of population labels}

\item{pop4}{a vector of population labels}

\item{bin}{path to the qpDstat binary file}

\item{pref}{path to and prefix of the packedancestrymap genotype files}

\item{outdir}{the output directory. files \code{out}, \code{parfile}, \code{poplistname}, \code{popfilename} may be overwritten}

\item{parfile}{qpDstat parameter file. If this is specified, \code{pop}, \code{pop2}, \code{pop3}, and \code{pop4} will be ignored.}

\item{printonly}{should the command be printed or executed?}
}
\value{
If \code{printonly}, the \code{qpDstat} command, otherwise a data frame with parsed \code{qpDstat} output
}
\description{
This requires a working installation of qpDstat, which will be called using \code{\link{system}}
}
\examples{
\dontrun{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
pop3 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
pop4 = 'Switzerland_Bichon.SG'
qpdstat_wrapper(pop1, pop2, pop3, pop4,
  bin = 'path/to/qpDstat', pref = 'path/to/packedancestrymap_prefix',
  env = 'export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/path/to/blas/')

qpdstat_wrapper(bin = 'path/to/qpDstat', parfile = 'path/to/parfile')
}
}
