% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{qpgraph}
\alias{qpgraph}
\title{Compute the fit of an admixture graph}
\usage{
qpgraph(
  f2_blocks,
  graph,
  f2_denom = 1,
  boot = FALSE,
  fudge = 1e-04,
  fudge_cov = 1e-05,
  fnscale = 1e-06,
  lsqmode = FALSE,
  numstart = 10,
  seed = NULL,
  cpp = TRUE,
  return_f4 = FALSE,
  f3precomp = NULL,
  ppinv = NULL,
  f2_blocks_test = NULL,
  low_q = 0,
  high_q = 1,
  verbose = FALSE
)
}
\arguments{
\item{f2_blocks}{A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}}.}

\item{graph}{An admixture graph represented as a matrix of edges, an \code{\link{igraph}} object, or the path to a qpGraph graph file. Edges can be constrained by providing a matrix or data frame of edges with columns titled \code{lower} and \code{upper} with lower and upper bounds, respectively. By default, admixture edges are constrained to be between zero and one (with paired edges summing to one), and drift edges have a lower bound at zero.}

\item{f2_denom}{Scales f2-statistics. A value of around 0.278 converts F2 to Fst.}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix of
f3 statistics will be computed using block-jackknife. Otherwise bootstrap resampling is performed \code{n} times,
where \code{n} is either equal to \code{boot} if it is an integer, or equal to the number of blocks if \code{boot} is \code{TRUE}.
The covariance matrix of f3 statistics will be computed using bootstrapping.}

\item{fudge}{Regularization term added to the diagonal elements of the covariance matrix of fitted branch lengths (after scaling by the matrix trace).}

\item{fudge_cov}{Regularization term added to the diagonal elements of the covariance matrix of estimated f3 statistics (after scaling by the matrix trace).}

\item{fnscale}{Optimization parameter passed to \code{control} in \code{\link{optim}}}

\item{lsqmode}{Least-squares mode. sets the offdiagonal elements of the block-jackknife covariance matrix to zero.}

\item{numstart}{Number of random initializations. Defaults to 10.}

\item{seed}{Seed for generating starting weights.}

\item{cpp}{Use C++ functions. Setting this to \code{FALSE} will be slower but can help with debugging.}

\item{return_f4}{Return all f4-statistics. Defaults to \code{FALSE} because this can be slow.}

\item{f3precomp}{Precomputed f3-statistics. This should be the output of \code{qpgraph_precompute_f3} and can be provided instead of \code{f2_blocks}. This can speed things up if many graphs are evaluated using the same set of f3-statistics.}

\item{ppinv}{Inverse of f3-statistics covariance matrix. Can be used for \code{\link{compare_fits3}}.}

\item{f2_blocks_test}{An optional 3d array of f2-statistics used for computing an out-of-sample score. Ideally this contains SNP blocks which are not part of \code{f2_blocks}. This allows to estimate the fit of a graph without overfitting and will not be used during the optimization step}

\item{low_q}{Reported lower quantile of fitted admixture edge weights across random initializations}

\item{high_q}{Reported upper quantile of fitted admixture edge weights across random initializations}

\item{verbose}{Print progress updates}
}
\value{
A list with data describing the model fit:
\enumerate{
\item \code{edges} a data frame where each row is an edge in the graph. For regular edges,
the column \code{weight} is the estimated edge length, and for admixture edges, it is the estimated admixture weight.
\item \code{score} the likelihood score of the fitted graph. lower values correspond to better fits.
the score is calculated as the inner product of the residuals (difference between estimated and
fitted f3 statistics), weighted by the inverse of the f3 covariance matrix.
\item \code{f2} estimated and fitted f2 statistics
\item \code{f3} estimated and fitted f3 statistics
\item \code{f4} estimated and fitted f4 statistics (if \code{return_f4 = TRUE})
\item \code{opt} a data frame with details of the weight-fitting step, including the randomly sampled starting weights.
}
}
\description{
Computes the fit of an admixturegraph for a given graph topology and empirical f2-block-jackknife statistics.
}
\examples{
out = qpgraph(example_f2_blocks, example_graph)
plot_graph(out$edges)
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics
}
\seealso{
\code{\link{qpgraph_wrapper}} for a wrapper functions which call the original qpGraph program.
}
