% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{qpgraph_precompute_f3}
\alias{qpgraph_precompute_f3}
\title{Compute f3-statistics from f2-statistics.}
\usage{
qpgraph_precompute_f3(f2_data, pops, f2_denom = 1, fudge = 0.001,
  lsqmode = FALSE)
}
\arguments{
\item{f2_data}{a 3d array of block-jackknife leave-one-block-out estimates of f2 statistics, output of \code{\link{afs_to_f2_blocks}}. alternatively, a directory with f2 statistics. see \code{\link{extract_data}}.}

\item{pops}{populations for which to compute f3-statistics}

\item{f2_denom}{scales f2-statistics. A value of around 0.278 converts F2 to Fst.}

\item{fudge}{try increasing this, if you get the error message \code{constraints are inconsistent, no solution!}.}

\item{lsqmode}{least-squares mode. sets the offdiagonal elements of the block-jackknife covariance matrix to zero.}
}
\value{
a list with four items
\enumerate{
\item \code{f3_jest} a matrix with f3-statistics for all population pairs with the output
\item \code{ppinv} a matrix with the inverse of the f3-statistic covariance matrix
\item \code{f2out} a data frame with f2 estimates
\item \code{f3out} a data frame with f3 estimates
}
}
\description{
Takes a 3d array of f2 block jackknife estimates and computes f3-statistics between the first population \code{p1} and all population pairs \code{i, j}: \code{f3(p1; p_i, p_j)}
}
\examples{
pops = get_leafnames(example_igraph)
qpgraph_precompute_f3(example_f2_blocks, pops)$f3_jest
\dontrun{
qpgraph_precompute_f3(f2_dir, pops, f2_denom = 0.278)
}
}
