% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{qpgraph_resample_multi}
\alias{qpgraph_resample_multi}
\title{Evaluate a qpgraph models many times}
\usage{
qpgraph_resample_multi(f2_blocks, graphlist, nboot, verbose = TRUE, ...)
}
\arguments{
\item{f2_blocks}{3d array of f2-statistics}

\item{graphlist}{A list of admixture graphs}

\item{nboot}{Number of bootstrap iterations}

\item{verbose}{Print progress updates}

\item{...}{Arguments passed to \code{\link{qpgraph}}}
}
\value{
A list of same length as \code{graphlist} with data frames with \code{\link{qpgraph}} results for each iteration of bootstrap resampled f2-statistics
}
\description{
This function is used in combination with \code{\link{compare_fits}} in order to test whether one graph has a significantly better fit than another. It creates bootstrap resampled SNP block training and test sets, and uses them to evaluate multiple graphs.
}
\seealso{
\code{\link{compare_fits}}
}
