% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{qpgraph_resample_snps2}
\alias{qpgraph_resample_snps2}
\title{Evaluate a qpgraph model many times}
\usage{
qpgraph_resample_snps2(f2_blocks, graph, f2_blocks_test, verbose = TRUE, ...)
}
\arguments{
\item{f2_blocks}{A list of bootstrap resampled f2-statistics}

\item{graph}{An admixture graph}

\item{f2_blocks_test}{A list of f2-statistics from all blocks which were not used in the corresponding f2-array in \code{f2_blocks}}

\item{verbose}{Print progress updates}

\item{...}{Arguments passed to \code{\link{qpgraph}}}
}
\value{
A data frame with \code{\link{qpgraph}} results for each iteration of bootstrap resampled f2-statistics
}
\description{
This function is used in combination with \code{\link{compare_fits3}} in order to test whether one graph has a significantly better fit than another. using a list of bootstrap resampled f2-statistics and corresponding test-set f2-statistics
}
\seealso{
\code{\link{compare_fits3}} \code{\link{boo_list}}
}
