% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{qpgraph_wrapper2}
\alias{qpgraph_wrapper2}
\title{A wrapper around qpGraph which creates 'parfile' and 'graphfile'}
\usage{
qpgraph_wrapper2(graph, bin, pref, outpop = "NULL", outdir = ".",
  printonly = FALSE, lambdascale = -1, lsqmode = "NO",
  diag = 1e-04, hires = "NO", forcezmode = "NO", allsnps = "NO",
  bigiter = 100)
}
\arguments{
\item{graph}{a graph as two column edge matrix .}

\item{bin}{location of the qpGraph binary file.}

\item{pref}{prefix of the packedancestrymap format genotype files.}

\item{outpop}{an outgroup population.}

\item{outdir}{output directory}

\item{printonly}{should output be executed or the command just be printed?}

\item{lambdascale}{lambdascale}

\item{lsqmode}{least-squares mode. sets the offdiagonal elements of the block-jackknife covariance matrix to zero.}

\item{diag}{diag}

\item{hires}{hires}

\item{forcezmode}{forcezmode}

\item{allsnps}{allsnps}

\item{bigiter}{bigiter}
}
\value{
a list with parsed qpGraph output
\enumerate{
\item \code{edges}: data frame
\item \code{score}: scalar
\item \code{f2}: data frame
}
}
\description{
A wrapper around qpGraph which creates 'parfile' and 'graphfile'
}
\examples{
\dontrun{
qpgraph_wrapper2(graph1,
                 bin = 'path/to/qpGraph',
                 pref = 'path/to/packedancestrymap_prefix')
}
}
