% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{qpwave_pairs}
\alias{qpwave_pairs}
\title{Compute all pairwise qpwave p-values}
\usage{
qpwave_pairs(f2_data, left, right)
}
\arguments{
\item{f2_data}{A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}}. Alternatively, a directory with f2 statistics.}

\item{left}{Left populations}

\item{right}{Right populations}
}
\description{
For all pairs of left populations qpwave rank 0 Chi-squared statistics and p-values will be computed. This tests for each pair of left populations whether it forms a clade with respect to the right populations.
}
\examples{
\dontrun{
left = pops[5:7]
right = pops[1:4]
f2_blocks = f2_from_precomp('/my/f2/dir/', pops = left, pops2 = right, afprod = TRUE)
qpwave_pairs(f2_blocks, left, right)
# If f2-stats are too big to load them into memory,
# the following will read the data for each pair from disk:
qpwave_pairs('/my/f2/dir/', left, right)
}
}
