% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpwave_wrapper}
\alias{qpwave_wrapper}
\title{Wrapper function around the original qpWave program}
\usage{
qpwave_wrapper(
  pref,
  left,
  right,
  target = NULL,
  bin = "~np29/o2bin/qpAdm",
  outdir = "./",
  parfile = NULL,
  allsnps = "NO",
  blgsize = 0.05,
  fancyf4 = "NO",
  f4mode = "YES",
  inbreed = "NO",
  printonly = FALSE,
  env = "",
  verbose = TRUE
)
}
\arguments{
\item{left}{Left populations (sources)}

\item{right}{Right populations (outgroups)}

\item{target}{Target population}

\item{verbose}{Print progress updates}
}
\description{
Wrapper function around the original qpWave program
}
