% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{random_newick}
\alias{random_newick}
\title{Generate a random binary graph}
\usage{
random_newick(n, start = "", end = "", outpop = NULL)
}
\arguments{
\item{n}{The number of terminal nodes, or a vector of population labels.}

\item{start}{Prefix.}

\item{end}{Postfix.}

\item{outpop}{Outgroup (if \code{n} is a vector of labels).}
}
\value{
Tree in newick format.
}
\description{
Generate a random binary graph
}
\examples{
random_newick(5)
random_newick(c('a', 'b', 'c', 'd')) # toplogy random, but pop order fixed
random_newick(sample(c('a', 'b', 'c', 'd'))) # toplogy and pop order random
}
