% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{random_sim}
\alias{random_sim}
\title{Generate a random graph and simulate it in msprime v1.x}
\usage{
random_sim(
  nleaf,
  nadmix,
  outpref = "random_sim",
  max_depth = NULL,
  ind_per_pop = 1,
  mutation_rate = 1.25e-08,
  admix_weights = 0.5,
  neff = 1000,
  time = 1000,
  fix_leaf = FALSE,
  outpop = NULL,
  nchr = 1,
  recomb_rate = 2e-08,
  seq_length = 1000,
  ghost_lineages = TRUE,
  run = FALSE
)
}
\arguments{
\item{nleaf}{Number of leaf nodes (i.e., number of final populations)}

\item{nadmix}{Number of admixture events}

\item{outpref}{Prefix of output files}

\item{max_depth}{Constraint specifying maximum depth of admixture tree (in generation)}

\item{ind_per_pop}{The number of individuals to simulate for each population. If a scalar value, it will be constant across all populations. Alternatively, it can be a named vector with a different value for each population.}

\item{mutation_rate}{Mutation rate per site per generation. The default is \code{1.25e-8}.}

\item{admix_weights}{Admixture weights. If a float value (0 < value < 1), admixture weights for each admixture event will be (value, 1-value). Alternatively, it can be a range, i.e., \code{c(0.1, 0.4)} specifying lower and upper limits of an uniform distribution from which admixture weight value will be drawn. By default, all admixture edges will have weight 0.5 (i.e., 50\\%-50\\%).}

\item{neff}{Effective population size. If a scalar value, it will be constant across all populations. Alternatively, it can be a range, i.e., \code{c(500, 1000)} specifying lower and upper limits of an uniform distribution from which values will be drawn}

\item{time}{Time between nodes. Either a scalar value (1000 by default) which the dates generated by \code{\link{pseudo_dates}}, or a range, i.e., \code{c(500, 1000)} specifying lower and upper limits of an uniform distribution from which values will be drawn (see \code{\link{random_dates}}).}

\item{fix_leaf}{Boolean specifying if leaf nodes will be fixed to 0. If TRUE, all samples will be drawn at the end of the simulation (i.e., from today).}

\item{outpop}{Name of the outgroup population.}

\item{nchr}{Number of chromosomes to simulate}

\item{recomb_rate}{A float value specifying recombination rate along the chromosomes. The default is \code{2e-8}.}

\item{seq_length}{The sequence length of the chromosomes. If it is a scalar value, the sequence length will be constant for all chromosomes.
Alternatively, it can be a vector with a length equal to number of chromosomes (i.e., \code{c(100,50)} to simulate 2 chromosomes with the lengths of 100 and 50).}

\item{ghost_lineages}{If \code{TRUE} (default), admixture events occur at time points shown on admixture graph (assumes ghost lineages) by \code{\link{plot_graph}}}

\item{run}{If \code{FALSE}, the function will terminate after writing the msprime script. If \code{TRUE}, it will try and execute the function with the default python installation. If you want to use some other python installation, you can set \verb{run = /my/python}.}
}
\value{
A list with the path of simulation script, data frame of graph edges, dates and population sizes:
\itemize{
\item{out - }{File name and path of simulation script}
\item{edges - }{An edge dataframe with weights}
\item{dates - }{A named vector dates for each node}
\item{neffs - }{A named vector effective population size for each node}
}
}
\description{

}
\examples{
# Create simulation script that simulates 2 chromosomes that are 50base long where maximum depth of the tree is 5000 generations, and plot the output graph
out = random_sim(nleaf=4, nadmix=0, max_depth=5000, nchr=2, seq_length=50)
plot_graph(out$edges, dates = out$dates, neff = out$neff hide_weights = TRUE)
}
