% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_bl}
\alias{read_bl}
\title{Read block_lengths from disk}
\usage{
read_bl(f2_dir)
}
\arguments{
\item{f2_dir}{directory from which to read files}
}
\value{
a vector with block_lengths
}
\description{
Read block_lengths from disk
}
\examples{
\dontrun{
read_bl(f2_dir)
}
}
