% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_eigenstrat}
\alias{read_eigenstrat}
\title{Read genotype data from \emph{EIGENSTRAT} files}
\usage{
read_eigenstrat(
  pref,
  inds = NULL,
  first = 1,
  last = Inf,
  transpose = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{pref}{Prefix of the packedancestrymap files}

\item{inds}{Individuals for which data should be read. Defaults to all individuals}

\item{first}{Index of first SNP to read}

\item{last}{Index of last SNP to read}

\item{transpose}{Transpose genotype matrix (default is \code{snps} x \code{individuals})}

\item{verbose}{Print progress updates}
}
\value{
A list with the genotype data matrix, the \code{.ind} file, and the \code{.snp} file
}
\description{
Read genotype data from \emph{EIGENSTRAT} files
}
\examples{
\dontrun{
samples = c('Ind1', 'Ind2', 'Ind3')
geno = read_packedancestrymap(prefix, samples)
}
}
