% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_packedancestrymap}
\alias{read_packedancestrymap}
\title{Read genotype data from packedancestrymap files}
\usage{
read_packedancestrymap(pref, inds = NULL, blocksize = 1000,
  na.action = "none", auto_only = TRUE, verbose = TRUE)
}
\arguments{
\item{pref}{prefix of packedancestrymap files (files have to end in \code{.geno}, \code{.ind}, \code{.snp}).}

\item{inds}{optional vector of samples to read in}

\item{blocksize}{number of SNPs read in each block.}

\item{na.action}{what to do with SNPs with missing data.
\itemize{
\item \code{none} (default) output table will have missing data
\item \code{impute} allele frequencies will be imputed from other populations
\item \code{remove} SNPs with missing data will be removed
}}

\item{auto_only}{If \code{TRUE} (the default), all SNPs with chromosome outside the range of 1 to 22 will be discarded.}

\item{verbose}{print progress updates}
}
\value{
a list with the genotype data matrix, the \code{.ind} file, and the \code{.snp} file
}
\description{
This is currently slower than \code{read_plink} because it is implemented only in \code{R}, not in \code{C++}.
}
\examples{
\dontrun{
samples = c('Ind1', 'Ind2', 'Ind3')
geno = read_packedancestrymap(prefix, samples)
}
}
