% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_plink}
\alias{read_plink}
\title{Read genotype data from \code{PLINK} files}
\usage{
read_plink(pref, inds = NULL, verbose = FALSE)
}
\arguments{
\item{pref}{Prefix of packedancestrymap files (files have to end in \code{.geno}, \code{.ind}, \code{.snp})}

\item{inds}{Individuals for which data should be read. Defaults to all individuals}

\item{verbose}{Print progress updates}
}
\value{
A list with the genotype data matrix, the \code{.ind} file, and the \code{.snp} file
}
\description{
See \href{https://www.rdocumentation.org/packages/genio}{genio} for a dedicated \code{R} package for
reading and writing \code{PLINK} files.
}
\examples{
\dontrun{
samples = c('Ind1', 'Ind2', 'Ind3')
geno = read_packedancestrymap(prefix, samples)
}
}
