% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resample_inds}
\alias{resample_inds}
\alias{qp3pop_resample_inds}
\alias{qpdstat_resample_inds}
\alias{qpadm_resample_inds}
\alias{qpgraph_resample_inds}
\title{Run models while leaving out individuals}
\usage{
qp3pop_resample_inds(dir, inds, pops, verbose = TRUE, ...)

qpdstat_resample_inds(dir, inds, pops, verbose = TRUE, ...)

qpadm_resample_inds(dir, inds, pops, verbose = TRUE, ...)

qpgraph_resample_inds(dir, inds, pops, verbose = TRUE, ...)
}
\arguments{
\item{dir}{directory with precomputed data}

\item{inds}{vector of individual names}

\item{pops}{vector of population names. Should be the same length as \code{inds}}

\item{verbose}{print progress updates}

\item{...}{named arguments passed to the \code{qp} function.}
}
\value{
a nested data frame where each model is a row, and the columns are model parameters and model outputs
}
\description{
These are wrapper functions around various \code{qp} functions, which will evaluate many models at once.
The models are formed by leaving out every individual which belongs to a population with more than one sample,
one at a time.
}
\examples{
\dontrun{
res = qpadm_resample_inds
unnest(res, weights)
}
}
