% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resample_snps}
\alias{resample_snps}
\alias{qp3pop_resample_snps}
\alias{qpdstat_resample_snps}
\alias{qpadm_resample_snps}
\alias{qpgraph_resample_snps}
\title{Run models while leaving out SNP blocks}
\usage{
qp3pop_resample_snps(f2_blocks, boot = FALSE, verbose = TRUE, ...)

qpdstat_resample_snps(f2_blocks, boot = FALSE, verbose = TRUE, ...)

qpadm_resample_snps(f2_blocks, boot = FALSE, verbose = TRUE, ...)

qpgraph_resample_snps(f2_blocks, boot = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{f2_blocks}{a 3d array of blocked f2 statistics}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time.
Otherwise bootstrap resampling is performed \code{n} times, where \code{n} is either equal to \code{boot} if it is an integer,
or equal to the number of blocks if \code{boot} is \code{TRUE}.}

\item{verbose}{print progress updates}

\item{...}{named arguments which are passed to the \code{qp} function.}
}
\value{
a nested data frame where each model is a row, and the columns are model parameters and model outputs
}
\description{
These are wrapper functions around various \code{qp} functions, which will evaluate a model on multiple subdivisions of the data.
The subdivisions are formed by leaving out one or more SNP block at a time (see \code{boot} for details).
}
\examples{
\dontrun{
res = qpadm_resample_snps(example_f2_blocks, target = target, left = left, right = right)
unnest(res, weights)
}
}
