% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{rotate_models}
\alias{rotate_models}
\title{Rotate populations between left and right}
\usage{
rotate_models(leftright, target, rightfix = NULL)
}
\arguments{
\item{leftright}{Populations which will be distributed between left and right}

\item{target}{Target population}

\item{rightfix}{Populations which will be on the right side in all models}
}
\value{
A data frame with Chi-squared statistics and p-values for each population combination
}
\description{
This functions creates a data frame with population combinations which can be used as the input for \code{\link{qpadm_multi}}
}
\examples{
\dontrun{
pops = dimnames(example_f2_blocks)[[1]]
rotate_models(leftright = pops[1:4],
             target = pops[5], rightfix = pops[6:7])
}
}
