% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{satisfies_constraints}
\alias{satisfies_constraints}
\title{Test constraints on a graph}
\usage{
satisfies_constraints(
  graph,
  nonzero_f4 = NULL,
  admix_constraints = NULL,
  event_order = NULL
)
}
\arguments{
\item{graph}{An admixture graph}

\item{nonzero_f4}{A data frame or matrix with four columns. One row for each f4-statistic which is
observed to be significantly non-zero}

\item{admix_constraints}{A data frame with columns \code{pop}, \code{min}, \code{max}}

\item{event_order}{A data frame with columns \code{earlier1}, \code{earlier2}, \code{later1}, \code{later2}}
}
\value{
\code{TRUE} if all admixture constraints are satisfied, else \code{FALSE}
}
\description{
This function tests whether a graph satisfies a number of different types of constraints
}
\seealso{
\code{\link{satisfies_numadmix}}, \code{\link{satisfies_zerof4}}, \code{\link{satisfies_eventorder}}
\dontrun{
# At least one admixture event for C, and none for D:
constrain_cd = tibble(pop = c('C', 'D'), min = c(1, NA), max = c(NA, 0))
satisfies_numadmix(random_admixturegraph(5, 2), constrain_cd)
}
}
