% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{satisfies_numadmix}
\alias{satisfies_numadmix}
\title{Test admixture constraints on a graph}
\usage{
satisfies_numadmix(graph, admix_constraints)
}
\arguments{
\item{graph}{An admixture graph}

\item{admix_constraints}{A data frame with columns \code{pop}, \code{min}, \code{max}}
}
\value{
\code{TRUE} if all admixture constraints are satisfied, else \code{FALSE}
}
\description{
This function returns \code{TRUE} if and only if the admixture graph satisfies all constraints on
the number of admixture events for the populations in \code{admix_constraints}
}
\examples{
\dontrun{
# At least one admixture event for C, and none for D:
constrain_cd = tribble(
  ~pop, ~min, ~max,
  'C', 1, NA,
  'D', 0, NA)
satisfies_numadmix(random_admixturegraph(5, 2), constrain_cd)
}
}
