% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{shortest_unique_prefixes}
\alias{shortest_unique_prefixes}
\title{Return shortest unique prefixes}
\usage{
shortest_unique_prefixes(strings, min_length = 1)
}
\arguments{
\item{strings}{A character vector}

\item{min_length}{Minimum length of prefixes}
}
\value{
Character vector with shortest unique prefixes
}
\description{
Return shortest unique prefixes
}
\examples{
strings = c("Abra", "Abracadabra", "Simsalabim")
shortest_unique_prefixes(strings)
}
