% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{simplify_graph}
\alias{simplify_graph}
\title{Removes redundant edges from graph.}
\usage{
simplify_graph(grph)
}
\arguments{
\item{grph}{an igraph object}
}
\description{
Nodes with in and out degree of one will be removed.
}
\examples{
simple = simplify_graph(igraph)
plot_graph2(igraph)
plot_graph2(simple)
}
