% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{split_afmat}
\alias{split_afmat}
\title{Split an allele frequency matrix into blocks}
\usage{
split_afmat(afmat, pops_per_block, outprefix, verbose = TRUE)
}
\arguments{
\item{afmat}{the allele frequency matrix to be split}

\item{pops_per_block}{the number of populations per block}

\item{outprefix}{the prefix of the output files. Directories are not created on the fly.}
}
\description{
This function splits a large allele frequency matrix into smaller blocks with \code{pops_per_block} populations per block, and saves them as \code{.RData} files in \code{outprefix}
}
\examples{
\dontrun{
afmatall = packedancestrymap_to_aftable('path/to/packedancestrymap_prefix', allpopulations,
                                         na.action = 'none', return_matrix = TRUE)
split_afmat(afmatall, pops_per_block = 20, outprefix = 'afmatall_split_v41.1/afmatall_')
}
}
\seealso{
\code{\link{packedancestrymap_to_aftable}}, \code{\link{write_split_f2_block}}
}
