% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{split_mat}
\alias{split_mat}
\title{Split a matrix into blocks}
\usage{
split_mat(mat, cols_per_chunk, prefix, overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{mat}{The matrix to be split}

\item{cols_per_chunk}{Number of columns per block}

\item{prefix}{Prefix of output files}

\item{overwrite}{Overwrite existing files (default \code{TRUE})}

\item{verbose}{Print progress updates}
}
\description{
This function splits a large matrix into smaller blocks with \code{cols_per_chunk} columns per block,
and saves them as \code{.rds} files with prefix \code{prefix}
}
\examples{
\dontrun{
afdat = packedancestrymap_to_afs('path/to/packedancestrymap_prefix', allpopulations)
split_mat(afdat$afs, cols_per_chunk = 20, prefix = 'afdat_split_v42.1/afs')
}
}
\seealso{
\code{\link{packedancestrymap_to_afs}}, \code{\link{afs_to_f2}}
}
