% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{split_multifurcations}
\alias{split_multifurcations}
\title{Split nodes with more than two edges}
\usage{
split_multifurcations(graph)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
A new admixture graph in which no node has more than two incoming or outgoing edges
}
\description{
Split nodes with more than two edges
}
