% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{spr_leaves}
\alias{spr_leaves}
\title{Modify a graph by regrafting a leaf}
\usage{
spr_leaves(graph, fix_outgroup = TRUE)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
A new admixture graph
}
\description{
Modify a graph by regrafting a leaf
}
