% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_descendants}
\alias{summarize_descendants}
\title{List leaf nodes for all internal nodes}
\usage{
summarize_descendants(graph)
}
\arguments{
\item{graph}{An admixture graphs}
}
\value{
A data frame with columns \code{from}, \code{to}, \code{id}
}
\description{
List leaf nodes for all internal nodes
}
\examples{
\dontrun{
summarize_descendants(graph)
}
}
