% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_descendants_list}
\alias{summarize_descendants_list}
\title{List leaf nodes for all internal nodes in a list of graphs}
\usage{
summarize_descendants_list(graphlist, rename = FALSE)
}
\arguments{
\item{graphlist}{A list of admixture graphs}
}
\value{
A data frame with columns \code{graph}, \code{from}, \code{n}, \code{frac}
}
\description{
List leaf nodes for all internal nodes in a list of graphs
}
\examples{
\dontrun{
summarize_descendants_list(graphlist)
}
}
