% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_eventorder}
\alias{summarize_eventorder}
\title{List population split events in a graph}
\usage{
summarize_eventorder(graph, unique_only = TRUE)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
A data frame with columns \code{earlier1}, \code{earlier2}, \code{later1}, \code{later2}
}
\description{
List population split events in a graph
}
\examples{
\dontrun{
summarize_eventorder(example_igraph)
}
}
