% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_eventorder_list}
\alias{summarize_eventorder_list}
\title{List population split events in a list of graphs}
\usage{
summarize_eventorder_list(graphlist)
}
\arguments{
\item{graphlist}{A list of admixture graphs}
}
\value{
A data frame with columns \code{earlier1}, \code{earlier2}, \code{later1}, \code{later2}, \code{n}, \code{frac}
}
\description{
List population split events in a list of graphs
}
\examples{
\dontrun{
summarize_eventorder(graphlist)
}
}
