% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{summarize_fits}
\alias{summarize_fits}
\title{Summarize graph fits}
\usage{
summarize_fits(fits, q_low = 0.05, q_high = 0.95)
}
\arguments{
\item{fits}{A nested data frame where each line is the output of \code{qpgraph()}}

\item{q_low}{Lower quantile for the weight confidence interval (default 0.05).}

\item{q_high}{Upper quantile for the weight confidence interval (default 0.95).}
}
\value{
A data frame summarizing the fits
}
\description{
This function takes the output of \code{\link{qpgraph_resample_snps}} and creates a data frame with summaries of the estimated graph parameters. \code{weight} has the mean of all fits, while \code{low} and \code{high} have lower and upper quantiles.
}
\seealso{
\code{\link{qpgraph_resample_snps}}
}
