% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_numadmix}
\alias{summarize_numadmix}
\title{List the number of admixture events for each population}
\usage{
summarize_numadmix(graph)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
A data frame with columns \code{pop} and \code{nadmix}
}
\description{
List the number of admixture events for each population
}
\examples{
\dontrun{
summarize_numadmix(example_igraph)
}
}
