% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_numadmix_list}
\alias{summarize_numadmix_list}
\title{List number of admixture events for each population in a list of graphs}
\usage{
summarize_numadmix_list(graphlist)
}
\arguments{
\item{graphlist}{A list of admixture graphs}
}
\value{
A data frame with columns \code{pop}, \code{mean}, \code{mean_nonzero}, \code{min}, \code{max}
}
\description{
List number of admixture events for each population in a list of graphs
}
\examples{
\dontrun{
summarize_numadmix_list(graphlist)
}
}
