% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_proxies}
\alias{summarize_proxies}
\title{Assign proxy populations to admixed populations}
\usage{
summarize_proxies(graph)
}
\arguments{
\item{graph}{An admixture graph in igraph format, or as edge list with column \code{weight}}
}
\value{
A data frame with columns \code{pop}, \code{nadmix}, \code{proxy}, \code{nproxies} (and \code{weight})
}
\description{
Assign proxy populations to admixed populations
}
\examples{
\dontrun{
summarize_proxies(example_igraph)
}
}
