% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_proxies_list}
\alias{summarize_proxies_list}
\title{List proxy populations in graphlist}
\usage{
summarize_proxies_list(graphlist)
}
\arguments{
\item{graphlist}{A list of admixture graphs}
}
\value{
A data frame with columns \code{pop}, \code{proxy}, \code{nadmix}, \code{nproxies}, \code{n}, \code{frac}
}
\description{
List proxy populations in graphlist
}
\examples{
\dontrun{
summarize_proxies_list(graphlist)
}
}
