% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_zerof4}
\alias{summarize_zerof4}
\title{List all f4-statistics which have to be zero}
\usage{
summarize_zerof4(graph, eps = 1e-07)
}
\arguments{
\item{graph}{An admixture graph}

\item{eps}{Used to determine what counts as zero}
}
\value{
A data frame with all (non-redundant) zero f4-statistics
}
\description{
List all f4-statistics which have to be zero
}
\examples{
\dontrun{
summarize_zerof4(example_igraph)
}
}
