% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_zerof4_list}
\alias{summarize_zerof4_list}
\title{List clades in a list of graphs}
\usage{
summarize_zerof4_list(graphlist)
}
\arguments{
\item{graphlist}{A list of admixture graphs}
}
\value{
A data frame with columns \code{pop1}, \code{pop2}, \code{pop3}, \code{pop4}, \code{n}, \code{frac}
}
\description{
List clades in a list of graphs
}
\examples{
\dontrun{
summarize_zerof4_list(graphlist)
}
}
