% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{swap_leaves}
\alias{swap_leaves}
\title{Modify a graph by swapping two leaf nodes}
\usage{
swap_leaves(graph, fix_outgroup = TRUE)
}
\arguments{
\item{graph}{An admixture graph}

\item{fix_outgroup}{Keep outgroup in place}
}
\value{
A new admixture graph
}
\description{
Modify a graph by swapping two leaf nodes
}
