% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{test_cladality}
\alias{test_cladality}
\title{Test if two sets of populations form two clades}
\usage{
test_cladality(f2_data, left, right, fudge = 1e-04, boot = FALSE,
  cpp = TRUE)
}
\description{
Test if two sets of populations form two clades
}
