% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{tree_neighbors}
\alias{tree_neighbors}
\title{Find all trees within SPR distance of 1}
\usage{
tree_neighbors(tree)
}
\arguments{
\item{tree}{A tree in \code{igraph} format}
}
\value{
A data frame with all trees
}
\description{
Returns all trees which can be reached through one iteration of subtree-prune-and-regraft
}
