% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{unidentifiable_edges}
\alias{unidentifiable_edges}
\title{Find all unidentifiable edges}
\usage{
unidentifiable_edges(graph, gtof = 1)
}
\arguments{
\item{graph}{An admixture graph}

\item{gtof}{which version of \code{graph_to_function} to use}
}
\value{
A data frame with all unidentifiable graph parameters
}
\description{
This function generates and evaluates admixture graphs in \code{numgen} iterations
to find well fitting admixturegraphs.
}
