% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_f2}
\alias{write_f2}
\title{Write blocked f2 estimates to disk}
\usage{
write_f2(f2_block, outdir, overwrite = FALSE)
}
\arguments{
\item{f2_block}{3d array with f2 block jackknife estimates. The first two dimensions of
\code{f2_block} have to have population names. Each file will be stored under \code{{outdir}/{pop1}/{pop2}.rds}.}

\item{outdir}{directory into which to write the files.}

\item{overwrite}{should existing \code{.rds} files be overwritten?}
}
\description{
This function takes a 3d array of blocked f2 estimates, splits it by population pair,
and writes each pair to a separate \code{.rds} file.
}
\examples{
\dontrun{
write_f2(f2_block, outdir = 'path/to/f2stats/')
}
}
\seealso{
\code{\link{read_f2}}
}
