% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixtools.R
\docType{package}
\name{admixtools}
\alias{admixtools}
\title{Tools for inferring demographic history from genetic data}
\description{
admixtools is a collection of several programs for calculating f-statistics
and using them to make inferences about demographic history from genetic data
Aside from the core functions \code{\link{qp3pop}}, \code{\link{qpdstat}}, \code{\link{qpadm}},
and \code{\link{qpgraph}}, there are also wrapper functions and parsing functions around the original
ADMIXTOOLS software, as well as functions to read genotype files and precompute the data necessary
to quickly compute f statistics.
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics
}
\seealso{
\url{https://uqrmaie1.github.io/admixtools/index.html}
}
\author{
Robert Maier \email{<robertmaier@gmx.net>}

Nick Patterson
}
