% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{afs_to_f2_blocks}
\alias{afs_to_f2_blocks}
\title{Compute all pairwise f2 statistics}
\usage{
afs_to_f2_blocks(
  afdat,
  maxmem = 8000,
  blgsize = 0.05,
  poly_only = TRUE,
  pops1 = NULL,
  pops2 = NULL,
  outpop = NULL,
  outdir = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{afdat}{A list with three items with the same SNP in each row (generated by \code{\link{packedancestrymap_to_afs}} or \code{\link{plink_to_afs}})
\itemize{
\item \code{afs} A matrix of allele frequencies for all populations (columns) and SNPs (rows)
\item \code{counts} A matrix of allele counts for all populations (columns) and SNPs (rows)
\item \code{snpdat} A data frame with SNP metadata
}}

\item{maxmem}{split up allele frequency data into blocks, if memory requirements exceed \code{maxmem} MB.}

\item{blgsize}{SNP block size in Morgan. Default is 0.05 (50 cM). If \code{blgsize} is 100 or greater, if will be interpreted as base pair distance rather than centimorgan distance.}

\item{poly_only}{Exclude sites with identical allele frequencies in all populations}

\item{outpop}{If specified, f2-statistics will be weighted by heterozygosity in this population}

\item{outdir}{Directory into which to write f2 data (if \code{NULL}, data is returned instead)}

\item{overwrite}{Should existing files be overwritten? Only relevant if \code{outdir} is not \code{NULL}}

\item{verbose}{Print progress updates}

\item{pop1}{\code{pops1} and \code{pops2} can be specified if only a subset of pairs should be computed.}

\item{pop2}{\code{pops1} and \code{pops2} can be specified if only a subset of pairs should be computed.}
}
\description{
This function takes a allele frequency data and computes blocked f2 statistics for all population pairs,
which are written to \code{outdir}. \eqn{f2} for each SNP is computed as
\eqn{(p1 - p2)^2 - p1 (1 - p1)/(n1 - 1) - p2 (1 - p2)/(n2 - 1)}, where \eqn{p1} and \eqn{p2} are
allele frequencies in populations \eqn{1} and \eqn{2}, and \eqn{n1} and \eqn{n2} is the number of
non-missing haplotypes in populations \eqn{1} and \eqn{2}. See \code{details}
}
\details{
For each population pair, each of the \eqn{i = 1, \ldots, n} resutling values
(\eqn{n} is around 700 in practice) is the mean \eqn{f2} estimate across all SNPs except the ones in block \eqn{i}.

\eqn{- p1 (1 - p1)/(2 n1 - 1) - p2 (1 - p2)/(2 n2 - 1)} is a correction term which makes the estimates
unbiased at low sample sizes.
}
\examples{
\dontrun{
afdat = plink_to_afs('/my/geno/prefix')
afs_to_f2_blocks(afdat, outdir = '/my/f2/data/')
}
}
