% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{afs_to_f2_blocks}
\alias{afs_to_f2_blocks}
\title{Compute all pairwise f2 statistics}
\usage{
afs_to_f2_blocks(afdat, maxmem = 8000, dist = 0.05, poly_only = TRUE,
  outdir = NULL, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{afdat}{A list with three items with the same SNP in each row (generated by \code{\link{packedancestrymap_to_aftable}} or \code{\link{plink_to_aftable}})
\itemize{
\item \code{afs} A matrix of allele frequencies for all populations (columns) and SNPs (rows)
\item \code{counts} A matrix of allele counts for all populations (columns) and SNPs (rows)
\item \code{snpdat} A data frame with SNP metadata
}}

\item{maxmem}{split up allele frequency data into blocks, if memory requirements exceed \code{maxmem} MB.}

\item{dist}{Genetic distance in Morgan. Default is 0.05 (50 cM).}

\item{poly_only}{Exclude sites with identical allele frequencies in all populations}

\item{outdir}{Directory into which to write f2 data (if \code{NULL}, data is returned instead)}

\item{overwrite}{Should existing files be overwritten? Only relevant if \code{outdir} is not \code{NULL}}

\item{verbose}{Print progress updates}

\item{block_lengths}{vector with lengths of each block. \code{sum(block_lengths)}
has to match \code{nrow(afmat)}. See \code{\link{get_block_lengths}}}
}
\value{
\code{NULL} if \code{outdir} is specified. Otherwise, it should return a list with f2 data arrays
}
\description{
This function takes a allele frequency data and computes blocked f2 statistics for all population pairs,
which are written to \code{outdir}. \eqn{f2} for each SNP is computed as
\eqn{(p1 - p2)^2 - p1 (1 - p1)/(n1 - 1) - p2 (1 - p2)/(n2 - 1)}, where \eqn{p1} and \eqn{p2} are
allele frequencies in populations \eqn{1} and \eqn{2}, and \eqn{n1} and \eqn{n2} is the average number of
non-missing haplotypes in populations \eqn{1} and \eqn{2}. See \code{details}
}
\details{
For each population pair, each of the \eqn{i = 1, \ldots, n} resutling values
(\eqn{n} is around 700 in practice) is the mean \eqn{f2} estimate across all SNPs except the ones in block \eqn{i}.

\eqn{- p1 (1 - p1)/(2 n1 - 1) - p2 (1 - p2)/(2 n2 - 1)} is a correction term which makes the estimates
unbiased at low sample sizes.
}
\examples{
\dontrun{
afdat = plink_to_aftable('/my/geno/prefix')
afs_to_f2_blocks(afdat, outdir = '/my/f2/data/')
}
}
