% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{agraph_to_igraph}
\alias{agraph_to_igraph}
\title{Convert agraph to igraph}
\usage{
agraph_to_igraph(agraph)
}
\arguments{
\item{agraph}{An admixture graph in \code{\link{agraph}} format}
}
\value{
An admixture graph in \code{\link{igraph}} format
}
\description{
\code{agraph} is the format used by the \code{admixturegraph} packge. \code{igraph} is used by the \code{admixtools} package
}
\examples{
\dontrun{
agraph_to_igraph(agraph)
}
}
