% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ancestrymap_to_aftable}
\alias{ancestrymap_to_aftable}
\title{Read allele frequencies from ancestrymap files}
\usage{
ancestrymap_to_aftable(pref, inds = NULL, pops = NULL,
  adjust_pseudohaploid = TRUE, randomize_alleles = TRUE, seed = NULL,
  verbose = TRUE)
}
\arguments{
\item{pref}{Prefix of ancestrymap files (files have to end in \code{.geno}, \code{.ind}, \code{.snp})}

\item{inds}{Individuals from which to compute allele frequencies}

\item{pops}{Populations from which to compute allele frequencies. If \code{NULL} (default), populations will be extracted from the third column in the \code{.ind} file. If population labels are provided, they should have the same length as \code{inds}, and will be matched to them by position}

\item{adjust_pseudohaploid}{Genotypes of pseudohaploid samples are coded as \code{0} or \code{2}, although only one allele is observed. \code{adjust_pseudohaploid} ensures that the observed allele count increases only by \code{1} for each pseudohaploid sample. This leads to slightly more accurate estimates of f-statistics. Setting this parameter to \code{FALSE} is equivalent to the ADMIXTOOLS \code{inbreed: NO} option.}

\item{randomize_alleles}{Randomly report allele frequencies for either reference or for alternative allele for each SNP. This can furhter reduce bias when f-statistics are computed from allele frequency products. It does not affect f2 estimates (and other f-statistics computed from them).}

\item{seed}{Random seed for \code{randomize_alleles}}

\item{verbose}{Print progress updates}
}
\value{
A list with three data frames: allele frequency data, allele counts, and SNP metadata
}
\description{
Read allele frequencies from ancestrymap files
}
\examples{
\dontrun{
afdat = ancestrymap_to_aftable(prefix, pops = pops)
afs = afdat$afs
counts = afdat$counts
}
}
