% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{compare_fits}
\alias{compare_fits}
\title{Compare the fit of two qpgraph models}
\usage{
compare_fits(scores1, scores2)
}
\arguments{
\item{scores1}{Scores for the first graph}

\item{scores2}{Scores for the second graph}
}
\description{
Takes the bootstrap score distribution of two fits on the same populations and tests whether the scores of one graph are significantly better than the scores of the other. Used to be called \code{compare_fits3}
}
\examples{
\dontrun{
boo = boo_list(f2_blocks, nboot = 100)
fits1 = qpgraph_resample_snps2(boo$boo, graph1, boo$test)
fits2 = qpgraph_resample_snps2(boo$boo, graph2, boo$test)
compare_fits(fits1$score_test, fits2$score_test)
}
# Use all SNP blocks for f3 covariance matrix
\dontrun{
boo = boo_list(f2_blocks, nboot = 100)
ppinv1 = qpgraph_precompute_f3(f2_blocks, get_leafnames(graph1))$ppinv
ppinv2 = qpgraph_precompute_f3(f2_blocks, get_leafnames(graph2))$ppinv
fits1 = qpgraph_resample_snps2(boo$boo, graph1, boo$test, ppinv = ppinv1)
fits2 = qpgraph_resample_snps2(boo$boo, graph2, boo$test, ppinv = ppinv2)
compare_fits(fits1$score_test, fits2$score_test)
}
# Same as above
\dontrun{
fits = qpgraph_resample_multi(f2_blocks, list(graph1, graph2), nboot = 100)
compare_fits(fits[[1]]$score_test, fits[[2]]$score_test)
}
}
\seealso{
\code{\link{qpgraph_resample_snps2}} \code{\link{boo_list}}
}
