% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{compare_fits3}
\alias{compare_fits3}
\title{Compare the fit of two qpgraph models}
\usage{
compare_fits3(scores1, scores2)
}
\arguments{
\item{fits1}{The first graph}

\item{fits2}{The second graph}
}
\description{
Takes the bootstrap score distribution of two fits on the same populations and tests whether the scores of one graph are significantly better than the scores of the other.
}
\examples{
\dontrun{
boo = boo_list(f2_blocks, nboot = 100)
fits1 = qpgraph_resample_snps2(boo$boo, graph1, boo$test)
fits2 = qpgraph_resample_snps2(boo$boo, graph2, boo$test)
compare_fits3(fits1$score_test, fits2$score_test)
}
}
