% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{delete_admix}
\alias{delete_admix}
\title{Delete an admixture edge}
\usage{
delete_admix(graph, from, to)
}
\arguments{
\item{graph}{An admixture graph}

\item{from}{Edge source node}

\item{to}{Edge target node}
}
\value{
Admixture graph with one deleted edge
}
\description{
Delete an admixture edge
}
\seealso{
\code{\link{insert_edge}}
}
