% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{desimplify_graph}
\alias{desimplify_graph}
\title{Adds two nodes before each admixture node.}
\usage{
desimplify_graph(grph)
}
\description{
This is used to revert simplify_graph.
}
\examples{
simple = simplify_graph(igraph)
desimple = desimplify_graph(simple)
plot_graph2(simple)
plot_graph2(desimple)
}
