% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{desimplify_graph}
\alias{desimplify_graph}
\title{Add two nodes before each admixture node}
\usage{
desimplify_graph(graph)
}
\arguments{
\item{graph}{An admixture graph}
}
\description{
This is used to revert simplify_graph.
}
\examples{
simple = simplify_graph(example_igraph)
desimple = desimplify_graph(simple)
plot_graph(simple)
plot_graph(desimple)
}
