% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{extract_f2}
\alias{extract_f2}
\title{Compute and store blocked f2 statistics}
\usage{
extract_f2(pref, outdir, inds = NULL, pops = NULL, dist = 0.05,
  maxmem = 8000, maxmiss = 1, minmaf = 0, maxmaf = 0.5,
  transitions = TRUE, transversions = TRUE, keepsnps = NULL,
  overwrite = FALSE, format = NULL, poly_only = FALSE,
  adjust_pseudohaploid = FALSE, verbose = TRUE)
}
\arguments{
\item{pref}{Prefix of packedancestrymap or PLINK files. packedancestrymap has to end in \code{.geno}, \code{.snp}, \code{.ind},
PLINK has to end in \code{.bed}, \code{.bim}, \code{.fam}}

\item{outdir}{Directory data where data will be stored}

\item{inds}{Individuals for which data should be extracted}

\item{pops}{Populations for which data should be extracted. If both \code{pops} and \code{inds} are provided, they should have the same length and will be matched by position. If only \code{pops} is provided, all individuals from the \code{.ind} or \code{.fam} file in those populations will be extracted. If only \code{inds} is provided, each indivdual will be assigned to its own population of the same name. If neither \code{pops} nor \code{inds} is provided, all individuals and populations in the \code{.ind} or \code{.fam} file will be extracted.}

\item{dist}{Genetic distance in Morgan. Default is 0.05 (50 cM).}

\item{maxmem}{Maximum amount of memory to be used. If the required amount of memory exceeds \code{maxmem}, allele frequency data will be split into blocks, and the computation will be performed separately on each block pair.}

\item{maxmiss}{Discard SNPs which are missing in a fraction of populations higher than \code{maxmiss}}

\item{minmaf}{Discard SNPs with minor allele frequency less than \code{minmaf}}

\item{maxmaf}{Discard SNPs with minor allele frequency greater than than \code{maxmaf}}

\item{transitions}{Set this to \code{FALSE} to exclude transition SNPs}

\item{transversions}{Set this to \code{FALSE} to exclude transversion SNPs}

\item{keepsnps}{SNP IDs of SNPs to keep. Overrides other SNP filtering options}

\item{overwrite}{Overwrite existing files in \code{outdir}}

\item{format}{Supply this if the prefix can refer to genotype data in different formats
and you want to choose which one to read. Should be either \code{plink}, \code{ancestrymap}, or \code{packedancestrymap}}

\item{poly_only}{Exclude sites with identical allele frequencies in all populations}

\item{adjust_pseudohaploid}{Genotypes of pseudohaploid samples are coded as \code{0} or \code{2}, although only one allele is observed. \code{adjust_pseudohaploid} ensures that the observed allele count increases only by \code{1} for each pseudohaploid sample. This leads to slightly more accurate estimates of f-statistics. Setting this parameter to \code{FALSE} is equivalent to the ADMIXTOOLS \code{inbreed: NO} option. Pseudohaploid samples are identified as those that don't have any genotypes coded as \code{1} among the first 1000 SNPs.}

\item{verbose}{Print progress updates}
}
\value{
SNP metadata (invisibly)
}
\description{
Prepare data for various admixtools functions. Reads data from packedancestrymap or PLINK files,
and computes allele frequencies and f2 block jackknife statistics for selected populations.
Data is either written to disk in files for each population and population pair, or returned as list.
This function calls \code{\link{packedancestrymap_to_aftable}} or \code{\link{plink_to_aftable}}
and \code{\link{afs_to_f2_blocks}}.
}
\examples{
\dontrun{
pref = 'my/genofiles/prefix'
extract_f2(pref, pops = c('popA', 'popB', 'popC'))
}
}
