% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{extract_f2_subset}
\alias{extract_f2_subset}
\title{Copy f2-statistics}
\usage{
extract_f2_subset(from, to, pops, verbose = TRUE)
}
\arguments{
\item{from}{Directory with f2-statistics}

\item{to}{Target directory}

\item{pops}{Populations to copy}

\item{verbose}{Print progress updates}
}
\description{
Copy a subset of f2-statistics to a new directory
}
\examples{
\dontrun{
pref = 'my/genofiles/prefix'
outdir = 'dir/for/afdata/'
extract_f2_subset(pref, outdir)
}
}
