% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract samples from PLINK files}
\usage{
extract_samples(
  inpref,
  outpref,
  inds = NULL,
  pops = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{inpref}{Prefix of the \emph{PLINK} input files}

\item{outpref}{Prefix of the \emph{PLINK} output files}

\item{inds}{Individuals which should be extracted}

\item{pops}{Populations which should be extracted. Can not be provided together with \code{inds}}

\item{overwrite}{Set this to \code{TRUE} if \code{inpref == outpref} and you really want to overwrite the input files.}

\item{verbose}{Print progress updates}
}
\description{
This function reads \emph{PLINK} files, extracts a subset of samples, and writes new \emph{PLINK} files using \code{\link[genio]{write_plink}}. It's probably slower than running the equivalent command in \emph{PLINK} directly, but it can be useful to do this from within R.
When \code{inds} or \code{pops} is provided, only a subset of samples will be extracted.
}
