% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{f2}
\alias{f2}
\title{Estimate f2 statistics}
\usage{
f2(
  data,
  pop1 = NULL,
  pop2 = NULL,
  boot = FALSE,
  sure = FALSE,
  unique_only = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{data}{Input data in one of three forms: \enumerate{
\item A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{extract_f2}} (fastest option)
\item A directory which contains pre-computed f2-statistics
\item The prefix of genotype files (slowest option)
}}

\item{pop1}{One of the following four:
\enumerate{
\item \code{NULL}: all possible population combinations will be returned
\item A vector of population labels. All combinations with the other \code{pop} arguments will be returned
\item A matrix with population combinations to be tested, with one population per column and one
combination per row. Other \code{pop} arguments will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or
population combinations to be tested. Other \code{pop} arguments will be ignored.
}}

\item{pop2}{A vector of population labels}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix
of the f statistics will be computed using block-jackknife. Otherwise bootstrap resampling is performed \code{n} times,
where \code{n} is either equal to \code{boot} if it is an integer, or equal to the number of blocks if \code{boot} is \code{TRUE}.
The the covariance matrix of the f statistics will be computed using bootstrapping.}

\item{sure}{The number of population combinations can get very large. This is a safety option that stops you
from accidently computing all combinations if that number is large.}

\item{unique_only}{If \code{TRUE} (the default), redundant combinations will be excluded}

\item{verbose}{Print progress updates}
}
\value{
\code{f2} returns a data frame with f2 statistics
}
\description{
Computes f2 statistics from f2 blocks of the form \eqn{f2(A, B)}
}
\examples{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
f2(example_f2_blocks, pop1, pop2)
\dontrun{
f2(f2_dir, pop1, pop2)
}
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history} Genetics

Peter, B. (2016) \emph{Admixture, Population Structure, and F-Statistics} Genetics
}
