% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{f2_from_geno}
\alias{f2_from_geno}
\title{Read f2 jackknife blocks from genotype data}
\usage{
f2_from_geno(pref, inds = NULL, pops = NULL, na.action = "none",
  maxmem = 8000, format = NULL, verbose = TRUE)
}
\arguments{
\item{pref}{prefix of packedancestrymap or PLINK files. packedancestrymap has to end in \code{.geno}, \code{.snp}, \code{.ind},
PLINK has to end in \code{.bed}, \code{.bim}, \code{.fam}}

\item{inds}{the individuals for which data should be extracted.}

\item{pops}{the populations for which data should be extracted. \code{pops} and \code{inds} cannot be specified
at the same time. If none are specified, all populations will be extracted.}

\item{na.action}{what to do with SNPs with missing data.
\itemize{
\item \code{none} (default) output table will have missing data
\item \code{impute} allele frequencies will be imputed from other populations
\item \code{remove} SNPs with missing data will be removed
}}

\item{maxmem}{split up allele frequency data into blocks, if memory requirements exceed \code{maxmem} MB.}

\item{format}{supply this if the prefix can refer to genotype data in different formats
and you want to choose which one to read}

\item{verbose}{print progress updates}
}
\description{
Prepare data for various admixtools functions. Reads data from packedancestrymap or PLINK files,
and computes allele frequencies and f2 block jackknife statistics for selected populations.
Data is either written to disk in files for each population and population pair, or returned as list.
This function calls \code{\link{packedancestrymap_to_aftable}} / \code{\link{plink_to_aftable}}
and \code{\link{afs_to_f2_blocks}}. The default treatment of missing data is \code{na.action = 'none'}
which is similar to \code{useallsnps: YES}. To get results that are closer to \code{useallsnps: NO},
use \code{na.action = 'remove'}.
}
\examples{
\dontrun{
pref = 'my/genofiles/prefix'
f2_blocks = f2_from_geno(pref, pops = c('Protoss', 'Zerg', 'Terran'))
}
}
