% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{f2_from_geno}
\alias{f2_from_geno}
\title{Read f2 jackknife blocks from genotype data}
\usage{
f2_from_geno(pref, inds = NULL, pops = NULL, maxmem = 8000,
  dist = 0.05, format = NULL, verbose = TRUE)
}
\arguments{
\item{pref}{Prefix of packedancestrymap or PLINK files. packedancestrymap has to end in \code{.geno}, \code{.snp}, \code{.ind},
PLINK has to end in \code{.bed}, \code{.bim}, \code{.fam}}

\item{inds}{Individuals for which data should be extracted}

\item{pops}{Populations for which data should be extracted. If both \code{pops} and \code{inds} are provided, they should have the same length and will be matched by position. If only \code{pops} is provided, all individuals from the \code{.ind} or \code{.fam} file in those populations will be extracted. If only \code{inds} is provided, each indivdual will be assigned to its own population of the same name. If neither \code{pops} nor \code{inds} is provided, all individuals and populations in the \code{.ind} or \code{.fam} file will be extracted.}

\item{maxmem}{Maximum amount of memory to be used. If the required amount of memory exceeds \code{maxmem}, allele frequency data will be split into blocks, and the computation will be performed separately on each block pair.}

\item{dist}{Genetic distance in Morgan. Default is 0.05 (50 cM).}

\item{format}{Supply this if the prefix can refer to genotype data in different formats
and you want to choose which one to read. Should be either \code{plink}, \code{ancestrymap}, or \code{packedancestrymap}}

\item{verbose}{Print progress updates}
}
\description{
Prepare data for various admixtools functions. Reads data from packedancestrymap or PLINK files,
and computes allele frequencies and f2 block jackknife statistics for selected populations.
Data is either written to disk in files for each population and population pair, or returned as list.
This function calls \code{\link{packedancestrymap_to_aftable}} / \code{\link{plink_to_aftable}}
and \code{\link{afs_to_f2_blocks}}.
}
\examples{
\dontrun{
pref = 'my/genofiles/prefix'
f2_blocks = f2_from_geno(pref, pops = c('pop1', 'pop2', 'pop3'))
}
}
