% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{get_block_lengths}
\alias{get_block_lengths}
\title{Find LD-independent blocks}
\usage{
get_block_lengths(dat, blgsize = 0.05, distcol = "cm", cpp = F)
}
\arguments{
\item{dat}{Data frame with columns 'CHR' and either 'POS' or 'cm'}

\item{blgsize}{SNP block size in Morgan. Default is 0.05 (50 cM).}

\item{distcol}{Column to use as distance column}

\item{cpp}{Should the faster C++ version be used?}
}
\value{
A numeric vector where the ith element lists the number of SNPs in the ith block.
}
\description{
A new block begins at the SNP after the first SNP which is not within \code{blgsize} of the start of the last block.
\code{dat} needs to be ordered first by 'CHR', then by 'POS' or 'cm'
}
\examples{
\dontrun{
prefix = 'path/to/packedancestrymap_prefix'
pops = c('pop1', 'pop2', 'pop3')
afdat = packedancestrymap_to_afs(prefix, pops = pops)
block_lengths = get_block_lengths(afdat)
}
}
