% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{graph_minusplus}
\alias{graph_minusplus}
\title{Find all graphs which result from adding and removing one admixture edge}
\usage{
graph_minusplus(graph)
}
\arguments{
\item{graph}{Admixture graph in \code{igraph} format}
}
\value{
A data frame with columns \code{source_from}, \code{source_to}, \code{dest_from}, \code{dest_to}, and \code{graph}
}
\description{
Find all graphs which result from adding and removing one admixture edge
}
\examples{
\dontrun{
newgraphs = graph_minusplus(example_igraph)
# now evaluate the new graphs
newgraphs \%>\%
  rowwise \%>\%
  mutate(res = list(qpgraph(example_f2_blocks, graph))) \%>\%
  unnest_wider(res)
}
}
