% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{graph_splittrees}
\alias{graph_splittrees}
\title{Find all trees which are part of the admixture graph}
\usage{
graph_splittrees(graph)
}
\arguments{
\item{graph}{Admixture graph in \code{igraph} format}
}
\value{
A data frame with columns \code{name} and \code{graph}
}
\description{
Find all trees which are part of the admixture graph
}
\examples{
\dontrun{
trees = graph_splittrees(example_igraph)
# now evaluate the trees
trees \%>\%
  rowwise \%>\%
  mutate(res = list(qpgraph(example_f2_blocks, graph))) \%>\%
  unnest_wider(res)
}
}
