% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{graph_to_lgo}
\alias{graph_to_lgo}
\title{Convert a graph to .lgo file to be used in Legofit software}
\usage{
graph_to_lgo(
  graph,
  outpref = "out",
  neff = 1000,
  samples = 1,
  dates_internal = 1000,
  dates_terminal = 0,
  outpop = NULL
)
}
\arguments{
\item{graph}{Edgelist representation (i.e., a dataframe) of an admixture graph. It must contain the following columns: \code{from}, \code{to}, \code{type} and \code{weight}.}

\item{outpref}{A prefix of output file}

\item{neff}{Effective population size in diploid individuals (1000 by default). If a scalar value, it will be constant across all populations. Alternatively, it can be a named vector with a different value for each population (e.g., \code{c('R'=100, 'A'=50, 'B'=50)}).}

\item{samples}{The number of samples to be drawn for each population (1 by default). If a scalar value, it will be constant across all populations. Alternatively, it can be a named vector with a different value for each population.}

\item{dates_internal}{Either a scalar value (1000 generations by default) with the dates generated by \code{\link{pseudo_dates}} function, or a named vector with dates for each internal graph nodes (in generations).}

\item{dates_terminal}{Either a scalar value that will be constant across all terminal nodes, or a named vector with dates for each terminal graph nodes (in generations). The default is 0, meaning that all the samples will be drawn at the end of the simulation (i.e., from today).}

\item{outpop}{Name of the outgroup population to be removed, together with the root.}
}
\value{
The file name and path of the lgo file
}
\description{
Convert a graph to .lgo file to be used in Legofit software
}
\examples{
\dontrun{
out = random_sim(4, 1)
graph_to_lgo(out$edges, neff = out$neff, samples = 10)
}
}
