% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{graph_to_qpadm}
\alias{graph_to_qpadm}
\title{Get all qpadm models for a graph}
\usage{
graph_to_qpadm(
  graph,
  target = NULL,
  allpops = TRUE,
  more_right = TRUE,
  data = NULL,
  ...
)
}
\arguments{
\item{graph}{An admixture graph}

\item{target}{Name of the target population. If \code{NULL}, it will cycle through all admixed populations.}

\item{allpops}{Only consider models which include all populations in the graph}

\item{more_right}{Only consider models where the number of right populations is greater than the number of left populations}

\item{data}{If \code{data} is set, all models will be evaluated using \code{\link{qpadm_multi}}}

\item{...}{Arguments passed to \code{\link{qpadm_multi}}}
}
\value{
A data frame with one qpadm model per row
}
\description{
This function returns all valid qpadm models for a graph and a given target population, and evaluates them if \code{data} is provided.
}
\examples{
\dontrun{
graph_to_qpadm(example_igraph, 'Mbuti.DG', data = example_f2_blocks)
}
}
