% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{igraph_to_agraph}
\alias{igraph_to_agraph}
\title{Convert igraph to agraph}
\usage{
igraph_to_agraph(igraph)
}
\arguments{
\item{agraph}{An admixture graph in \code{\link{igraph}} format}
}
\value{
An admixture graph in \code{\link{agraph}} format
}
\description{
\code{agraph} is the format used by the \code{admixturegraph} packge. \code{igraph} is used by the \code{admixtools} package
}
\examples{
\dontrun{
igraph_to_agraph(example_igraph)
}
}
