% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{insert_admix}
\alias{insert_admix}
\title{Insert a single edge into graph}
\usage{
insert_admix(
  graph,
  source_from = NULL,
  source_to = NULL,
  dest_from = NULL,
  dest_to = NULL,
  substitute = FALSE
)
}
\arguments{
\item{graph}{An admixture graph}

\item{source_from}{Parent node of the source edge}

\item{source_to}{Child node of the source edge}

\item{dest_from}{Parent node of the destination edge}

\item{dest_to}{Child node of the destination edge}

\item{substitute}{Should another edge be inserted, if the one specified doesn't work?}
}
\value{
Adxmiture graph with inserted edge
}
\description{
Insert a single edge into graph
}
\seealso{
\code{\link{delete_admix}}, \code{\link{insert_admix_n}}
}
