% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{insert_admix_n}
\alias{insert_admix_n}
\title{Insert admixture edges into graph}
\usage{
insert_admix_n(graph, n = 1, fix_outgroup = TRUE)
}
\arguments{
\item{graph}{An admixture graph}

\item{from}{List of nodes. New edges will originate above these nodes.}

\item{to}{List of nodes. New edges will end above these nodes.}

\item{substitute}{Should another edge be inserted, if the one specified doesn't work?}
}
\value{
Admixture graph with inserted edges
}
\description{
Insert admixture edges into graph
}
\seealso{
\code{\link{insert_admix}} \code{\link{delete_admix}}
}
