% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{lazadm}
\alias{lazadm}
\title{Estimate admixture weights}
\usage{
lazadm(f2_data, left, right, target, boot = FALSE, constrained = TRUE)
}
\arguments{
\item{f2_data}{A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}}.}

\item{left}{Left populations (sources)}

\item{right}{Right populations (outgroups)}

\item{target}{Target population}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix
of f4 statistics, as well as the weight standard errors, will be computed using block-jackknife.
Otherwise bootstrap resampling is performed \code{n} times, where \code{n} is either equal to \code{boot} if it is an integer,
or equal to the number of blocks if \code{boot} is \code{TRUE}. The the covariance matrix of f4 statistics,
as well as the weight standard errors, will be computed using bootstrapping.}

\item{constrained}{If \code{TRUE} (default), admixture weights will all be non-negative.
If \code{FALSE}, they can be negative, as in \code{\link{qpadm}}}
}
\value{
A data frame with weights and standard errors for each left population
}
\description{
Models target as a mixture of left populations, and outgroup right populations. Uses Lazaridis method
based non-negative least squares of f4 matrix.
}
\examples{
target = 'Denisova.DG'
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
lazadm(example_f2_blocks, left, right, target)
lazadm(example_f2_blocks, left, right, target, constrained = FALSE)
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics

Haak, W. et al. (2015) \emph{Massive migration from the steppe was a source for Indo-European
languages in Europe.} Nature (SI 9)
}
\seealso{
\code{\link{qpadm}}
}
