% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{lazadm}
\alias{lazadm}
\title{Estimate admixture weights}
\usage{
lazadm(f2_data, target = NULL, left = NULL, right = NULL,
  boot = FALSE, getcov = FALSE, constrained = TRUE)
}
\arguments{
\item{f2_data}{a 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}}.}

\item{target}{target population}

\item{left}{source populations (or \code{leftlist} file)}

\item{right}{outgroup populations (or \code{rightlist} file)}

\item{getcov}{should standard errors be returned? Currently not implemented.}

\item{constrained}{if \code{TRUE} (default), admixture weights will all be non-negative.
if \code{FALSE}, they can be negative, as in \code{\link{qpadm}}}
}
\value{
a data frame with weights and standard errors for each left population
}
\description{
Models target as a mixture of left populations, and outgroup right populations. Uses Lazaridis method
based non-negative least squares of f4 matrix.
}
\examples{
target = 'Denisova.DG'
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
lazadm(example_f2_blocks, target, left, right)
lazadm(example_f2_blocks, target, left, right, constrained = FALSE)
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics

Haak, W. et al. (2015) \emph{Massive migration from the steppe was a source for Indo-European
languages in Europe.} Nature (SI 9)
}
\seealso{
\code{\link{qpadm}}
}
