% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{msprime_genome}
\alias{msprime_genome}
\title{Simulate an admixture graph in msprime v1.x.}
\usage{
msprime_genome(
  graph,
  outpref = "msprime_sim",
  neff = 1000,
  ind_per_pop = 1,
  mutation_rate = 1.25e-08,
  time = 1000,
  fix_leaf = FALSE,
  nchr = 1,
  recomb_rate_chr = 2e-08,
  seq_length = 1000,
  admix_default = 0.5,
  run = FALSE,
  ghost_lineages = TRUE,
  shorten_admixed_leaves = FALSE
)
}
\arguments{
\item{graph}{Graph as \code{igraph} object or edge list with columns 'from' and 'to'. If it's an edge list with column 'weight' (possibly from a fitted graph), the admixture weights will be used. Otherwise, all admixture edges will have weight 0.5.}

\item{outpref}{Prefix of output files}

\item{neff}{Effective population size. If a scalar value, it will be constant across all populations. Alternatively, it can be a named vector with a different value for each population (e.g., \code{c('R'=100, 'A'=50, 'B'=50)}).}

\item{ind_per_pop}{The number of individuals to simulate for each population. If a scalar value, it will be constant across all populations.
Alternatively, it can be a named vector with a different value for each population (e.g., \code{c('A'=10, 'B'=20)} to sample 10 and 20 diploid individuals from populations A and B, respectively).}

\item{mutation_rate}{Mutation rate per site per generation. The default is \code{1.25e-8}.}

\item{time}{Either a scalar value (1000 by default) which the dates generated by \code{\link{pseudo_dates}}, or a named vector with dates for each graph node (in generations).}

\item{fix_leaf}{Boolean specifying if leaf nodes will be fixed to 0. If \code{TRUE}, all samples will be drawn at the end of the simulation (i.e., from today).}

\item{nchr}{Number of chromosomes to simulate.}

\item{recomb_rate_chr}{A float value specifying recombination rate along the chromosomes. The default is \code{2e-8}.}

\item{seq_length}{The sequence length of the chromosomes. If it is a scalar value, the sequence length will be constant for all chromosomes.
Alternatively, it can be a vector with a length equal to number of chromosomes (i.e., \code{c(100,50)} to simulate 2 chromosomes with the lengths of 100 and 50).}

\item{admix_default}{A float value specifying default admixture proportion for all admixture nodes. The default is \code{0.5} (i.e., 50\\%-50\\%).
If another value between 0 and 1 is specified, admixture weights for each admixture event will be (value, 1-value).}

\item{run}{If \code{FALSE}, the function will terminate after writing the msprime script. If \code{TRUE}, it will try and execute the function with the default python installation. If you want to use some other python installation, you can set \verb{run = /my/python}.}

\item{ghost_lineages}{A boolean value specifying whether there should be a ghost lineages.
If \code{TRUE}, admixture events occur at time points shown on admixture graph by \code{\link{plot_graph}}.
If \code{FALSE} (default), admixture occurs at the previous generation.}

\item{shorten_admixed_leaves}{If \code{TRUE} simulate the behavior of treemix where drift after admixture is not allowed}
}
\value{
The file name and path of the simulation script
}
\description{
Unlike \code{\link{msprime_sim}}, this function can simulate continuous sequence and multiple chromosomes.
}
\examples{
results = qpgraph(example_f2_blocks, example_graph)
# Simulate 3 chromosomes whose lengths are 50, 100 and 100
msprime_genome(results$edges, nchr=3, seq_length=c(50, 100, 100))
}
