% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{msprime_sim}
\alias{msprime_sim}
\title{Generate msprime simulation code from admixture graph}
\usage{
msprime_sim(
  graph,
  outpref = "msprime_sim",
  nsnps = 1000,
  neff = 1000,
  ind_per_pop = 1,
  mutation_rate = 0.001,
  time = 10000,
  run = FALSE,
  python = "python"
)
}
\arguments{
\item{graph}{Graph as \code{igraph} object or edge list with columns 'from' and 'to' (optionally 'weight')}
}
\value{
msprime python code backbone
}
\description{
Generate msprime simulation code from admixture graph
}
\examples{
\dontrun{
results = qpgraph(example_f2_blocks, example_graph)
msprime_sim(results$edges)
}
}
