% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{msprime_sim}
\alias{msprime_sim}
\title{Simulate an admixture graph in msprime}
\usage{
msprime_sim(
  graph,
  outpref = "msprime_sim",
  nsnps = 1000,
  neff = 1000,
  ind_per_pop = 1,
  mutation_rate = 0.001,
  time = 1000,
  admix_default = 0.5,
  run = FALSE,
  numcores = NULL,
  shorten_admixed_leaves = FALSE
)
}
\arguments{
\item{graph}{Graph as \code{igraph} object or edge list with columns 'from' and 'to'. If it's an edge list with column 'weight' (possibly from a fitted graph), the admixture weights will be used. Otherwise, all admixture edges will have weight 0.5.}

\item{outpref}{Prefix of output files}

\item{nsnps}{The number of SNPs to simulate. All SNPs will be simulated independently of each other.}

\item{neff}{Effective population size. If a scalar value, it will be constant across all populations. Alternatively, it can be a named vector with a different value for each population.}

\item{ind_per_pop}{The number of individuals to simulate for each population. If a scalar value, it will be constant across all populations. Alternatively, it can be a named vector with a different value for each population.}

\item{mutation_rate}{Mutation rate per site per generation. The default is set to a high value (0.001) to obtain more polymorphic SNPs in order to speed up the simulation.}

\item{time}{Either a scalar value (1000 by default) which the dates generated by \code{\link{pseudo_dates}}, or a named vector with dates for each graph node.}

\item{admix_default}{The default admixture proportions for all admixture nodes}

\item{run}{If \code{FALSE}, the function will terminate after writing the msprime script. If \code{TRUE}, it will try and execute the function with the default python installation. If you want to use some other python installation, you can set \verb{run = /my/python}.}

\item{numcores}{The number of cores to use when simulating data.}
}
\value{
The file name and path of the simulation script
}
\description{
This function generates an msprime simulation script, and optionally executes it in python
}
\examples{
\dontrun{
results = qpgraph(example_f2_blocks, example_graph)
msprime_sim(results$edges)
}
}
