% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{node_counts}
\alias{node_counts}
\title{Count how often each node in graph occurs in other graphs}
\usage{
node_counts(graph, graphlist)
}
\description{
Count how often each node in graph occurs in other graphs
}
\examples{
\dontrun{
sigs = example_winners \%>\% mutate(sig = map(igraph, node_signature)) \%$\% sig \%>\% unlist \%>\% table \%>\% c
node_signature(example_winners$igraph[[1]])
}
}
