% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{packedancestrymap_to_aftable}
\alias{packedancestrymap_to_aftable}
\title{Read allele frequencies from packedancestrymap files}
\usage{
packedancestrymap_to_aftable(pref, inds = NULL, pops = NULL,
  blocksize = 1000, auto_only = TRUE, verbose = TRUE)
}
\arguments{
\item{pref}{prefix of packedancestrymap files (files have to end in \code{.geno}, \code{.ind}, \code{.snp}).}

\item{inds}{vector of samples from which to compute allele frequencies.}

\item{pops}{vector of populations from which to compute allele frequencies. If \code{NULL} (default), populations will be extracted from the third column in the \code{.ind} file}

\item{blocksize}{number of SNPs read in each block.}

\item{auto_only}{If \code{TRUE} (the default), all SNPs with chromosome outside the range of 1 to 22 will be discarded.}

\item{verbose}{print progress updates}
}
\value{
a list with two items: allele frequency data and allele counts.
}
\description{
This is currently slower than \code{plink_to_aftable} because it is implemented only in \code{R}, not in \code{C++}.
}
\examples{
\dontrun{
afdat = packedancestrymap_to_aftable(prefix, pops = pops)
afs = afdat$afs
counts = afdat$counts
}
}
