% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{packedancestrymap_to_aftable}
\alias{packedancestrymap_to_aftable}
\title{Read allele frequencies from packedancestrymap files}
\usage{
packedancestrymap_to_aftable(pref, inds = NULL, pops = NULL,
  blocksize = 1000, ignore_ploidy = FALSE, randomize_alleles = TRUE,
  seed = NULL, verbose = TRUE)
}
\arguments{
\item{pref}{prefix of packedancestrymap files (files have to end in \code{.geno}, \code{.ind}, \code{.snp}).}

\item{inds}{vector of samples from which to compute allele frequencies.}

\item{pops}{vector of populations from which to compute allele frequencies. If \code{NULL} (default), populations will be extracted from the third column in the \code{.ind} file}

\item{blocksize}{number of SNPs read in each block.}

\item{ignore_ploidy}{treat pseudohaploid samples as if they were diploid samples (each observed genotype in a sample increases the allele count by two). This is equivalent to the ADMIXTOOLS \code{inbreed: NO} option. The default (\code{FALSE}) produces slightly more accurate f2-statistics by increasing the allele count for each pseudohaploid sample only by one.}

\item{randomize_alleles}{randomly report allele frequencies for reference or for alternative allele for each SNP. For data with large amounts of missingness, this can help to reduce bias of f4-statistics computed from allele frequency products.}

\item{seed}{random seed for \code{randomize_alleles}}

\item{verbose}{print progress updates}
}
\value{
a list with two items: allele frequency data and allele counts.
}
\description{
Read allele frequencies from packedancestrymap files
}
\examples{
\dontrun{
afdat = packedancestrymap_to_aftable(prefix, pops = pops)
afs = afdat$afs
counts = afdat$counts
}
}
