% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{packedancestrymap_to_aftable}
\alias{packedancestrymap_to_aftable}
\title{Read allele frequencies from packedancestrymap files}
\usage{
packedancestrymap_to_aftable(pref, inds = NULL, pops = NULL,
  adjust_pseudohaploid = TRUE, randomize_alleles = FALSE,
  seed = NULL, verbose = TRUE)
}
\arguments{
\item{pref}{prefix of packedancestrymap files (files have to end in \code{.geno}, \code{.ind}, \code{.snp}).}

\item{inds}{vector of samples from which to compute allele frequencies.}

\item{pops}{vector of populations from which to compute allele frequencies. If \code{NULL} (default), populations will be extracted from the third column in the \code{.ind} file}

\item{adjust_pseudohaploid}{genotypes of pseudohaploid samples are coded as \code{0} or \code{2}, although only one allele is observed. \code{adjust_pseudohaploid} ensures that the observed allele count increases only by \code{1} for each pseudohaploid sample. This leads to slightly more accurate estimates of f-statistics. Setting this parameter to \code{FALSE} is equivalent to the ADMIXTOOLS \code{inbreed: NO} option.}

\item{randomize_alleles}{randomly report allele frequencies for reference or for alternative allele for each SNP. Changes allele frequency products, but not f2 estimates.}

\item{seed}{random seed for \code{randomize_alleles}}

\item{verbose}{print progress updates}
}
\value{
a list with two items: allele frequency data and allele counts.
}
\description{
Read allele frequencies from packedancestrymap files
}
\examples{
\dontrun{
afdat = packedancestrymap_to_aftable(prefix, pops = pops)
afs = afdat$afs
counts = afdat$counts
}
}
