% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{packedancestrymap_to_aftable}
\alias{packedancestrymap_to_aftable}
\title{Read allele frequencies from packedancestrymap genotype file}
\usage{
packedancestrymap_to_aftable(pref, pops = NULL, inds = NULL,
  blocksize = 1000, na.action = "none", return_matrix = FALSE)
}
\arguments{
\item{pref}{prefix of packedancestrymap files (has to end in .geno, .ind, .snp).}

\item{pops}{vector of populations from which to compute allele frequencies.}

\item{inds}{vector of samples from which to compute allele frequencies. \code{pops} or \code{inds} have to be specified.}

\item{blocksize}{number of SNPs read in each block.}

\item{na.action}{what to do with missing SNPs. \code{none} output table will have missing data. \code{impute} allele frequencies will be imputed from other populations. \code{remove} SNPs with missing data will be removed.}
}
\value{
a tibble with allele frequency data. first six columns are snpfile, remaining columns are allele frequencies for reach population.
}
\description{
Read allele frequencies from packedancestrymap genotype file
}
\examples{
\dontrun{
afdat = packedancestrymap_to_aftable(prefix, pops)
}
}
