% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{packedancestrymap_to_plink}
\alias{packedancestrymap_to_plink}
\alias{ancestrymap_to_plink}
\title{Convert (packed)ancestrymap to PLINK}
\usage{
packedancestrymap_to_plink(
  inpref,
  outpref,
  inds = NULL,
  pops = NULL,
  verbose = TRUE
)
}
\arguments{
\item{inpref}{Prefix of the (packed)ancestrymap input files}

\item{outpref}{Prefix of the \emph{PLINK} output files}

\item{inds}{Individuals which should be extracted}

\item{pops}{Populations which should be extracted. Can not be provided together with \code{inds}}

\item{verbose}{Print progress updates}
}
\description{
This function converts \emph{(packed)ancestrymap/EIGENSTRAT} format files to \emph{PLINK} files, using \code{\link[genio]{write_plink}}.
When \code{inds} or \code{pops} is provided, only a subset of samples will be extracted.
}
\section{Alias}{

\code{ancestrymap_to_plink}
}

