% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{parse_qpadm_output}
\alias{parse_qpadm_output}
\title{Read qpAdm output file}
\usage{
parse_qpadm_output(outfile)
}
\arguments{
\item{outfile}{Output file generated by qpAdm.}
}
\value{
Data frame with output data.
}
\description{
Read qpAdm output file
}
