% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{parse_qpgraph_graphfile}
\alias{parse_qpgraph_graphfile}
\title{Read qpGraph graph file}
\usage{
parse_qpgraph_graphfile(graphfile)
}
\arguments{
\item{graphfile}{file with admixture graph in qpGraph format.}
}
\value{
graph represented as two column edge matrix.
}
\description{
Read qpGraph graph file
}
