% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parse_qpgraph_graphfile}
\alias{parse_qpgraph_graphfile}
\title{Reads graph in qpGraph format}
\usage{
parse_qpgraph_graphfile(graphfile)
}
\arguments{
\item{graphfile}{file with admixture graph in qpGraph format.}
}
\value{
graph represented as two column edge matrix.
}
\description{
Reads graph in qpGraph format
}
