% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{parse_qpgraph_graphfile}
\alias{parse_qpgraph_graphfile}
\title{Read qpGraph graph file}
\usage{
parse_qpgraph_graphfile(graphfile, split_multi = TRUE)
}
\arguments{
\item{graphfile}{File with admixture graph in qpGraph format.}
}
\value{
Graph represented as two column edge matrix. Can have four columns if edges are locked
}
\description{
Read qpGraph graph file
}
