% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parse_qpgraph_output}
\alias{parse_qpgraph_output}
\title{Reads qpGraph output file}
\usage{
parse_qpgraph_output(outfile)
}
\arguments{
\item{outfile}{output file generated by qpGraph.}
}
\value{
list of output data.
}
\description{
Reads qpGraph output file
}
