% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{plink_to_aftable}
\alias{plink_to_aftable}
\title{Read allele frequencies from PLINK files}
\usage{
plink_to_aftable(pref, pops = NULL, inds = NULL, na.action = "none",
  return_matrix = FALSE, verbose = FALSE)
}
\arguments{
\item{pref}{prefix of PLINK files (has to end in \code{.bed}, \code{.bim}, \code{.fam}).}

\item{pops}{vector of populations from which to compute allele frequencies.}

\item{inds}{vector of samples from which to compute allele frequencies. \code{pops} and \code{inds} cannot be specified at the same time. If none are specified, populations will be extracted from the first column in the \code{.fam} file}

\item{na.action}{what to do with SNPs with missing data.
\itemize{
\item \code{none} (default) output table will have missing data
\item \code{impute} allele frequencies will be imputed from other populations
\item \code{remove} SNPs with missing data will be removed
}}

\item{return_matrix}{return data as list of matrices}
}
\value{
a list with two items: allele frequency data and individual counts. Unless \code{return_matrix} is specified, first six columns are bimfile, remaining columns are data for each population.
}
\description{
Read allele frequencies from PLINK files
}
\examples{
\dontrun{
afdat = plink_to_aftable(prefix, pops)
afs = afdat$afs
counts = afdat$counts
}
}
