% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Plot a comparison of two lists of qpgraph output}
\usage{
plot_comparison(out1, out2)
}
\arguments{
\item{out1}{first qpgraph output.}

\item{out2}{second qpgraph output.}
}
\value{
a ggplot object.
}
\description{
Plot a comparison of two lists of qpgraph output
}
\examples{
fit1 = qpgraph(graph1, f2_blocks, block_lengths, cpp = TRUE)
fit2 = qpgraph(graph1, f2_blocks, block_lengths, cpp = FALSE)
plot_comparison(fit1, fit2)
}
