% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot an admixture graph}
\usage{
plot_graph(
  graph,
  fix = NULL,
  title = "",
  color = TRUE,
  textsize = 2.5,
  highlight_unidentifiable = FALSE
)
}
\arguments{
\item{graph}{An admixture graph. If it's an edge list with a \code{label} column,
those values will displayed on the edges}

\item{fix}{If \code{TRUE}, there will be an attempt to rearrange the nodes to minimize
the number of intersecting edges. This can take very long for large graphs.
By default this is only done for graphs with fewer than 10 leaves.}

\item{title}{Plot title}

\item{color}{Plot it in color or greyscale}

\item{textsize}{Size of edge and node labels}

\item{highlight_unidentifiable}{Highlight unidentifiable edges in red. Can be slow for large graphs. See \code{\link{unidentifiable_edges}}.}
}
\value{
A ggplot object
}
\description{
Plot an admixture graph
}
\examples{
plot_graph(example_graph)
}
