% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot an admixture graph}
\usage{
plot_graph(graph, fix = TRUE, title = "", color = TRUE)
}
\arguments{
\item{graph}{an admixture graph. If it's an edge list with a \code{label} column, those values will displayed on the edges}

\item{fix}{if \code{TRUE} (the default), there will be an attempt to rearrange the nodes to minimize the number of intersecting edges. This can take very long for large graphs.}

\item{title}{plot title}

\item{color}{plot it in color or greyscale}
}
\value{
a ggplot object
}
\description{
Plot an admixture graph
}
\examples{
plot_graph(example_graph)
}
