% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_graph}
\alias{plot_graph}
\title{Plots an admixture graph.}
\usage{
plot_graph(edges, fix = TRUE, title = "", color = TRUE)
}
\arguments{
\item{edges}{a two column matrix specifying the admixture graph. first column is source node, second column is target node. the first edge has to be root -> outgroup. admixture nodes are inferred as those nodes which are the targets of two different sources.}

\item{title}{plot title}
}
\value{
a ggplot object.
}
\description{
Plots an admixture graph.
}
\examples{
plot_graph(graph1)
}
