% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_map}
\alias{plot_map}
\title{Plot samples on a map}
\usage{
plot_map(leafcoords, map_layout = 1)
}
\arguments{
\item{leafcoords}{data frame with columns \code{group}, \code{lon}, \code{lat}}

\item{map_layout}{1 or 2}
}
\value{
a plotly object.
}
\description{
Plot samples on a map
}
\examples{
\dontrun{
plot_map(example_anno, 1)
}
}
