% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_map}
\alias{plot_map}
\title{Plots samples on a map.}
\usage{
plot_map(leafcoords, map_layout = 1)
}
\arguments{
\item{leafcoords}{data frame with columns \code{group}, \code{lon}, \code{lat}}

\item{map_layout}{1 or 2}
}
\value{
a ggplot object.
}
\description{
Plots samples on a map.
}
\examples{
\dontrun{
p1 = plot_map(igraph1, anno_v41, 1)
}
}
