% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotly_graph}
\alias{plotly_graph}
\title{Plot an admixture graph using plotly}
\usage{
plotly_graph(
  graph,
  collapse_threshold = 0,
  fix = FALSE,
  shift_down = TRUE,
  print_highlow = FALSE,
  highlight_unidentifiable = FALSE,
  pos = NULL
)
}
\arguments{
\item{graph}{An admixture graph}

\item{collapse_threshold}{Collapse nodes if they are separated by less than this distance (for fitted graphs)}

\item{fix}{If \code{TRUE}, there will be an attempt to rearrange the nodes to minimize
the number of intersecting edges. This can take very long for large graphs.
By default this is only done for graphs with fewer than 10 leaves.}

\item{shift_down}{Shift descendent nodes down}

\item{highlight_unidentifiable}{Highlight unidentifiable edges in red. Can be slow for large graphs. See \code{\link{unidentifiable_edges}}.}

\item{pos}{Optional data frame with node coordinates (columns \code{node}, \code{x}, \code{y})}
}
\value{
A plotly object
}
\description{
Plot an admixture graph using plotly
}
\examples{
plotly_graph(example_graph)
}
