% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qp3pop}
\alias{qp3pop}
\alias{f3}
\title{Estimate f3 statistics}
\usage{
qp3pop(
  data,
  pop1 = NULL,
  pop2 = NULL,
  pop3 = NULL,
  boot = FALSE,
  sure = FALSE,
  unique_only = TRUE,
  blgsize = NULL,
  block_lengths = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Input data in one of three forms: \enumerate{
\item A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{f2_from_geno}} (fastest option)
\item A directory which contains pre-computed f2-statistics
\item The prefix of genotype files (slowest option)
}}

\item{pop1}{One of the following four:
\enumerate{
\item \code{NULL}: all possible population combinations will be returned
\item A vector of population labels. All combinations with the other \code{pop} arguments will be returned
\item A matrix with population combinations to be tested, with one population per column and one
combination per row. Other \code{pop} arguments will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or
population combinations to be tested. Other \code{pop} arguments will be ignored.
}}

\item{pop2}{A vector of population labels}

\item{pop3}{A vector of population labels}

\item{boot}{If \code{FALSE} (the default), block-jackknife resampling will be used to compute standard errors.
Otherwise, block-bootstrap resampling will be used to compute standard errors. If \code{boot} is an integer, that number
will specify the number of bootstrap resamplings. If \code{boot = TRUE}, the number of bootstrap resamplings will be
equal to the number of SNP blocks.}

\item{sure}{The number of population combinations can get very large. This is a safety option that stops you
from accidently computing all combinations if that number is large.}

\item{unique_only}{If \code{TRUE} (the default), redundant combinations will be excluded}

\item{verbose}{Print progress updates}

\item{...}{Additional arguments passed to \code{\link{f3blockdat_from_geno}} if \code{data} is a genotype prefix, or to \code{\link{get_f2}} otherwise}

\item{apply_corr}{With \code{apply_corr = FALSE}, no bias correction is performed. With \code{apply_corr = TRUE} (the default), a bias correction term based on the heterozygosity in the first population is subtracted from the f3 estimate. The option is ineffective if the first argument is an array of pre-computed f2-statistics. In that case, the bias correction can be toggled by the \code{apply_corr} parameter in \code{\link{extract_f2}} or \code{\link{f2_from_geno}}. The heterozygosity calculation used in \code{apply_corr = TRUE} requires at least two haploid samples or one diploid sample in the first population. Otherwise, \code{apply_corr = TRUE} will result in missing values.}

\item{outgroupmode}{With \code{outgroupmode = FALSE}, estimates of f3 will be normalized by estimates of the heterozygosity of the target population. This is the default option if the first argument is the prefix of genotype data. If the first argument is an array of pre-computed f2-statistics, then no normalization can be performed, which corresponds to \code{outgroupmode = TRUE}. As with \code{apply_corr = TRUE}, the heterozygosity calculation used in \code{outgroupmode = FALSE} requires at least two pseudodiploid samples or one diploid sample in the first population. Otherwise, \code{outgroupmode = FALSE} will result in missing values.}

\item{poly_only}{Only keep SNPs with mean allele frequency not equal to 0 or 1. Defaults to \code{FALSE}. Only effective if the first argument is the prefix of genotype data. If the first argument is an array of pre-computed f2-statistics, use the \code{poly_only} argument in \code{\link{extract_f2}} or \code{\link{f2_from_geno}}}
}
\value{
\code{qp3pop} returns a data frame with f3 statistics, with columns for populations 1 to 3,
f3-estimate (\code{est}), standard error of the estimate (\code{se}), z-score (\code{z}, \code{est}/\code{se}), p-value (\code{2*(1-pnorm(z))}),
and the number of SNPs used (\code{n}; only if first argument is genotype prefix)
}
\description{
Computes f3 statistics of the form \eqn{f3(A; B, C)}. This is generally equivalent to
\eqn{(f2(A, B) + f2(A, C) - f2(B, C)) / 2} and to \eqn{f4(A, B; A, C)}.
However, the exact estimates depend on the type of input data, selection of SNPs,
and on the values of some of the arguments, which are described below.
}
\details{
The default values of the parameters \code{apply_corr}, \code{outgroupmode}, \code{poly_only} depend on the first argument (\code{data}), and they can affect the estimated f3-statistics. When the \code{data} is the prefix of genotype data, the default parameters are generally the same as in the original qp3pop program. When \code{data} is an array of pre-computed f2-statistics, the parameters may be set while computing f2-statistics. If the first population is a single pseudodiploid sample, it is not possible to get unbiased estimates of f3. To get biased estimates for a single pseudodiploid sample from genotype data, set \code{outgroupmode = TRUE} and \code{apply_corr = FALSE}. Under this combination of parameters, \verb{f3(A; B, C)} should also be identical to \verb{f4(A, B; A, C)}, since f4-statistics generally do not require any bias correction.
See \code{examples} for more information.
}
\section{Alias}{

\code{f3}
}

\examples{
\dontrun{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
pop3 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
qp3pop(example_f2_blocks, pop1, pop2, pop3)

pops = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
qp3pop(example_f2_blocks, pops)
qp3pop(example_f2_blocks, pops, unique_only = FALSE)
qp3pop(f2_dir, pop1, pop2, pop3)

# Below are three scenarios, and in each one `qp3pop()` and `qp3pop_wrapper()`
# should give the same or very similar estimates. Note that to compute `f3(A; B, C)`,
# `qp3pop_wrapper()`, and the original qp3pop program, expect populations to be in the order `B`, `C`, `A`.

prefix = '/path/to/geno/prefix'
qp3popbin = '/path/to/AdmixTools/bin/qp3Pop'
pops = dimnames(example_f2_blocks)[[1]]

# 1. target diploid, outgroupmode NO (this is the default when passing a geno file prefix)
qp3pop_wrapper(prefix, pops[2], pops[3], pops[1], bin = qp3popbin, outgroupmode = FALSE)
qp3pop(prefix, pops[1], pops[2], pops[3])
  # est, se, z match well; n is higher
qp3pop(prefix, pops[1], pops[2], pops[3], poly_only = TRUE)
  # est, se, z match less well; n is identical

# 2. target diploid, outgroupmode YES (this is the only option with precomputed f2-stats)
qp3pop_wrapper(prefix, pops[2], pops[3], pops[1], bin = qp3popbin, outgroupmode = TRUE)
qp3pop(prefix, pops[1], pops[2], pops[3], outgroupmode = TRUE)
  # est, se, z match (except for factor 1000)
f2b = f2_from_geno(prefix, pops = pops[1:3], poly_only = FALSE)
qp3pop(f2b, pops[1], pops[2], pops[3])

# 3. target pseudodiploid (no heterozygotes means heterozygosity rate correction is not possible)
qp3pop_wrapper(prefix, pops[1], pops[3], pops[2], bin = qp3popbin, outgroupmode = TRUE)
qp3pop(prefix, pops[2], pops[1], pops[3], outgroupmode = TRUE, apply_corr = FALSE)
  # est, se, z match (except for factor 1000)
}
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history} Genetics

Peter, B. (2016) \emph{Admixture, Population Structure, and F-Statistics} Genetics
}
