% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qp3pop}
\alias{qp3pop}
\alias{f3}
\title{Estimate f3 statistics}
\usage{
qp3pop(f2_data, pop1 = NULL, pop2 = NULL, pop3 = NULL,
  f2_denom = 1, sure = FALSE, unique_only = TRUE, verbose = FALSE)
}
\arguments{
\item{f2_data}{a 3d array of block-jackknife leave-one-block-out estimates of f2 statistics, output of \code{\link{afs_to_f2_blocks}}. alternatively, a directory with f2 statistics. see \code{\link{extract_data}}.}

\item{pop1}{one of the following four:
\enumerate{
\item \code{NULL}: all possible triples of the populations in \code{f2_blocks} will be returned
\item a vector of population labels
\item a data frame in which the first four columns specify the population quadruples to be tested. \code{pop2} and \code{pop3} will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or population quadruples to be tested. \code{pop2} and \code{pop3} will be ignored.
}}

\item{pop2}{a vector of population labels}

\item{pop3}{a vector of population labels}

\item{f2_denom}{scales f2-statistics. 1 correspondes to \code{f4mode: YES}. 1/4.75 is similar to \code{f4mode: NO}}
}
\value{
If \code{printonly}, the \code{qp3Pop} command, otherwise a data frame with parsed \code{qp3Pop} output
}
\description{
Computes f3 statistics from f2 jackknife blocks of the form \eqn{F_3(A; B, C)}. Equivalent to  \eqn{(F_2(A, B) + F_2(A, C) - F_2(B, C)) / 2} and to \eqn{F_4(A, B; A, C)}
}
\section{Alias}{

\code{f3}
}

\examples{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
pop3 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
qp3pop(example_f2_blocks, pop1, pop2, pop3)
\dontrun{
qp3pop(f2_dir, pop1, pop2, pop3)
}
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics
}
