% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qp3pop}
\alias{qp3pop}
\alias{f3}
\title{Estimate f3 statistics}
\usage{
qp3pop(
  f2_data,
  pop1 = NULL,
  pop2 = NULL,
  pop3 = NULL,
  boot = FALSE,
  sure = FALSE,
  unique_only = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{f2_data}{A 3d array with blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{f2_from_geno}}
Alternatively, a directory with precomputed data. See \code{\link{extract_f2}} and \code{\link{extract_counts}}.}

\item{pop1}{One of the following four:
\enumerate{
\item \code{NULL}: all possible pairs of the populations in \code{f2_blocks} will be returned
\item A vector of population labels
\item A matrix with population combinations to be tested, with one population per column and one
combination per row. Other \code{pop} arguments will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or
population combinations to be tested. Other \code{pop} arguments will be ignored.
}}

\item{pop2}{A vector of population labels}

\item{pop3}{A vector of population labels}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix
of the f statistics will be computed using block-jackknife. Otherwise bootstrap resampling is performed \code{n} times,
where \code{n} is either equal to \code{boot} if it is an integer, or equal to the number of blocks if \code{boot} is \code{TRUE}.
The the covariance matrix of the f statistics will be computed using bootstrapping.}

\item{sure}{The number of population combinations can get very large. This is a safety option that stops you
from accidently computing all combinations if that number is large.}

\item{unique_only}{If \code{TRUE} (the default), redundant combinations will be excluded}

\item{verbose}{Print progress updates}
}
\value{
A data frame with f3 statistics
}
\description{
Computes f3 statistics from f2 blocks of the form \eqn{f3(A; B, C)}. Equivalent to
\eqn{(f2(A, B) + f2(A, C) - f2(B, C)) / 2} and to \eqn{f4(A, B; A, C)}
}
\section{Alias}{

\code{f3}
}

\examples{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
pop3 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
qp3pop(example_f2_blocks, pop1, pop2, pop3)
\dontrun{
qp3pop(f2_dir, pop1, pop2, pop3)
}
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history} Genetics

Peter, B. (2016) \emph{Admixture, Population Structure, and F-Statistics} Genetics
}
