% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qp3pop_wrapper}
\alias{qp3pop_wrapper}
\title{Wrapper function around the original qp3Pop program}
\usage{
qp3pop_wrapper(
  pref,
  source1,
  source2 = NULL,
  target = NULL,
  bin = "~np29/o2bin/qp3Pop",
  outdir = ".",
  parfile = NULL,
  inbreed = "NO",
  outgroupmode = "NO",
  printonly = FALSE,
  env = "",
  verbose = TRUE
)
}
\arguments{
\item{pref}{Path to and prefix of the packedancestrymap genotype files}

\item{source1}{One of the following four:
\enumerate{
\item \code{NULL}: Populations will be read from \code{poplistname} or \code{popfilename} specified in \code{parfile}
\item A vector of population labels
\item A data frame in which the first four columns specify the population triples to be tested.
\code{source2}, \code{target} will be ignored.
\item The location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or
population combinations to be tested. \code{source2} and \code{target} will be ignored.
}}

\item{source2}{A vector of population labels}

\item{target}{A vector of population labels}

\item{bin}{Path to the qp3Pop binary file}

\item{outdir}{Output directory. files \code{out}, \code{parfile}, \code{poplistname},
\code{popfilename} may be overwritten}

\item{parfile}{qp3Pop parameter file. If this is specified, \code{source1}, \code{source2},
\code{target} will be ignored.}

\item{inbreed}{inbreed}

\item{outgroupmode}{outgroupmode}

\item{printonly}{Should the command be printed or executed?}

\item{env}{Export environmental variables. See examples.}

\item{verbose}{Print progress updates}
}
\value{
If \code{printonly}, the \code{qp3Pop} command, otherwise a data frame with parsed \code{qp3Pop} output
}
\description{
Computes f3 statistics of the form \eqn{f3(A; B, C)}. Equivalent to \eqn{(f2(A, B) + f2(A, C) - f2(B, C)) / 2}
and to \eqn{f4(A, B; A, C)}. Requires a working installation of qp3Pop, which will be called
using \code{\link{system}}
}
\examples{
\dontrun{
source1 = c('Altai_Neanderthal.DG', 'Vindija.DG')
source2 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
target = 'Denisova.DG'
qp3pop_wrapper('genotype_prefix', source1, source2, target,
  bin = 'path/to/qp3Pop')
}
}
