% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{qpadm}
\alias{qpadm}
\alias{qpwave}
\title{Estimate admixture weights}
\usage{
qpadm(
  f2_data,
  left,
  right,
  target = NULL,
  fudge = 1e-04,
  boot = FALSE,
  constrained = FALSE,
  cpp = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{f2_data}{A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}}. Alternatively, a directory with f2 statistics.}

\item{left}{Left populations (sources)}

\item{right}{Right populations (outgroups)}

\item{target}{Target population}

\item{fudge}{value added to diagonal matrix elements before inverting}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix
of f4 statistics, as well as the weight standard errors, will be computed using block-jackknife.
Otherwise bootstrap resampling is performed \code{n} times, where \code{n} is either equal to \code{boot} if it is an integer,
or equal to the number of blocks if \code{boot} is \code{TRUE}. The the covariance matrix of f4 statistics,
as well as the weight standard errors, will be computed using bootstrapping.}

\item{constrained}{Constrain admixture weights to be non-negative}

\item{cpp}{Use C++ functions. Setting this to \code{FALSE} will be slower but can help with debugging.}

\item{verbose}{Print progress updates}
}
\value{
A list with output describing the model fit:
\enumerate{
\item \code{weights} A data frame with estimated admixture proportions where each row is a left population.
Returned only if \code{target} is specified
\item \code{f4} A data frame with estimated (and fitted) f4-statistics
\item \code{rankdrop} A data frame describing model fits with different ranks, including p-values for the overall fit
and for nested models (comparing two models with rank difference of one).
\item \code{popdrop} A data frame describing model fits with different populations. Returned only if \code{target} is specified
}
}
\description{
Models target as a mixture of left populations, given a set of outgroup right populations.
Can be used to estimate admixture proportions, and to estimate the number of independent
admixture events.
}
\section{Alias}{

\code{qpwave}
}

\examples{
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
target = 'Denisova.DG'
qpadm(example_f2_blocks, left, right, target)
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics

Haak, W. et al. (2015) \emph{Massive migration from the steppe was a source for Indo-European
languages in Europe.} Nature (SI 10)
}
\seealso{
\code{\link{lazadm}}
}
