% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{qpadm}
\alias{qpadm}
\title{Estimate admixture weights}
\usage{
qpadm(
  data,
  left,
  right,
  target,
  f4blocks = NULL,
  fudge = 1e-04,
  boot = FALSE,
  getcov = TRUE,
  constrained = FALSE,
  cpp = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{The input data in the form of:
\itemize{
\item A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{extract_f2}}
\item A directory with f2 statistics
\item The prefix of a genotype file
}}

\item{left}{Left populations (sources)}

\item{right}{Right populations (outgroups)}

\item{target}{Target population}

\item{fudge}{value added to diagonal matrix elements before inverting}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix
of f4 statistics, as well as the weight standard errors, will be computed using block-jackknife.
Otherwise bootstrap resampling is performed \code{n} times, where \code{n} is either equal to \code{boot} if it is an integer,
or equal to the number of blocks if \code{boot} is \code{TRUE}. The the covariance matrix of f4 statistics,
as well as the weight standard errors, will be computed using bootstrapping.}

\item{getcov}{Compute weights covariance. Setting \code{getcov = FALSE} will speed up the computation.}

\item{constrained}{Constrain admixture weights to be non-negative}

\item{cpp}{Use C++ functions. Setting this to \code{FALSE} will be slower but can help with debugging.}

\item{verbose}{Print progress updates}

\item{...}{If \code{data} is the prefix of genotype files, additional arguments will be passed to \code{\link{f4blockdat_from_geno}}}
}
\value{
A list with output describing the model fit:
\itemize{
\item \code{weights} A data frame with estimated admixture proportions where each row is a left population.
\item \code{f4} A data frame with estimated and fitted f4-statistics
\item \code{rankdrop} A data frame describing model fits with different ranks, including p-values for the overall fit
and for nested models (comparing two models with rank difference of one). A model with \code{l} left populations and \code{r} right populations has an f4-matrix of dimensions \code{(l-1)*(r-1)}. If no two left population form a clade with respect to all right populations, this model will have rank \code{(l-1)*(r-1)}.
\itemize{
\item \code{f4rank}: Number of ranks below full rank
\item \code{dof}: Degrees of freedom of the chi-squared null distribution: \code{(l-1-f4rank)*(r-1-f4rank)}
\item \code{chisq}: Chi-sqaured statistic, obtained as \verb{E'QE}, where \code{E} is the difference between estimated and fitted f4-statistics, and \code{Q} is the f4-statistic covariance matrix.
\item \code{p}: p-value obtained from \code{chisq} as \code{pchisq(chisq, df = dof, lower.tail = FALSE)}
\item \code{dofdiff}: Difference in degrees of freedom between this model and the model with one less rank
\item \code{chisqdiff}: Difference in chi-squared statistics
\item \code{p_nested}: p-value testing whether the difference between two models of rank difference 1 is significant
}
\item \code{popdrop} A data frame describing model fits with different populations. Note that all models use the same set of SNPs as the first model.
\itemize{
\item \code{pat}: A binary code indicating which populations are present in this model. \code{1} represents dropped populations.
\item \code{wt}:
\item \code{dof}: Degrees of freedom
\item \code{chisq}: Chi-sqaured statistic, obtained as \verb{E'QE}, where \code{E} is the difference between estimated and fitted f4-statistics, and \code{Q} is the f4-statistic covariance matrix.
\item \code{p}: p-value obtained from \code{chisq} as \code{pchisq(chisq, df = dof, lower.tail = FALSE)}
\item \code{f4rank}:
\item \code{feasible}: A model is feasible if all weights fall between 0 and 1
\item \verb{<population name>}: The weights for each population in this model
}
}
}
\description{
Models target as a mixture of left populations, given a set of outgroup right populations.
Can be used to estimate admixture proportions, and to estimate the number of independent
admixture events.
}
\examples{
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
target = 'Denisova.DG'
qpadm(example_f2_blocks, left, right, target)
\dontrun{
qpadm("/my/geno/prefix", left, right, target)
}
}
\references{
Haak, W. et al. (2015) \emph{Massive migration from the steppe was a source for Indo-European
languages in Europe.} Nature (SI 10)
}
\seealso{
\code{\link{qpwave}}, \code{\link{lazadm}}
}
