% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{qpadm}
\alias{qpadm}
\title{Estimate admixture weights}
\usage{
qpadm(target, left, right, f2_blocks = NULL, block_lengths = NULL,
  f2_dir = NULL, fudge = 1e-04, getcov = TRUE, cpp = TRUE)
}
\arguments{
\item{target}{the target population}

\item{left}{the source population}

\item{right}{outgroup populations}

\item{f2_blocks}{3d array of block-jackknife leave-one-block-out estimates of f2 statistics. output of \code{\link{afs_to_f2_blocks}}. they are weighted by inverse of outgroup heterozygosity, if outgroup was specified.}

\item{block_lengths}{the jackknife block lengths used in computing the f2 statistics. see \code{\link{get_block_lengths}}.}

\item{f2_dir}{a directory with f2 statistics for each population pair in the graph. must contain 'block_lengths.RData'.}

\item{fudge}{value added to diagonal matrix elements before inverting}

\item{getcov}{should standard errors be returned? Setting this to FALSE makes this function much faster.}

\item{cpp}{should optimization be done using C++ or R function? cpp = TRUE is much faster.}
}
\value{
a data frame with weights and standard errors for each left population
}
\description{
Models target as a mixture of left populations, and outgroup right populations.
}
\examples{
target = 'Denisova.DG'
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
qpadm(target, left, right, f2_blocks, block_lengths)
}
