% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{qpadm_models}
\alias{qpadm_models}
\title{Return all valid qpAdm models for an admixturegraph}
\usage{
qpadm_models(grph, add_outgroup = FALSE, nested = TRUE, abbr = -1)
}
\arguments{
\item{grph}{an admixture graph as igraph object}

\item{add_outgroup}{should the graph outgroup be added to the qpAdm right populations? Technically this shouldn't be an informative outgroup for qpAdm.}

\item{nested}{should a nested data frame be returned (\code{TRUE}), or should populations be concatenated into strings (\code{FALSE})?}

\item{abbr}{maximum number of characters to print for each population. The default (-1) doesn't abbreviate the names.}
}
\description{
For large admixturegraph, there may be a large number of valid qpAdm models!
}
\examples{
\dontrun{
qpadm_models(igraph2, add_outgroup = TRUE)
qpadm_models(igraph2, add_outgroup = TRUE) \%>\% slice(1) \%$\% list(target, left, right)
}
}
