% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{qpadm_models}
\alias{qpadm_models}
\title{Partition a list of populations into left and right populations}
\usage{
qpadm_models(pops, allpops = TRUE, more_right = TRUE)
}
\arguments{
\item{pops}{A vector of populations}

\item{allpops}{Return only models which use all provided populations}

\item{more_right}{Return only models with more right than left populations}
}
\value{
A data frame with one qpadm model per row
}
\description{
Partition a list of populations into left and right populations
}
\examples{
\dontrun{
graph_to_qpadm(get_leafnames(example_igraph))
}
}
\seealso{
\code{\link{qpadm_models}}, \code{\link{graph_f2_function}}
}
