% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{qpadm_models}
\alias{qpadm_models}
\title{Return all valid qpAdm models for an admixturegraph.}
\usage{
qpadm_models(grph, add_outgroup = FALSE, nested = TRUE, abbr = -1)
}
\arguments{
\item{grph}{an admixture graph as igraph object}

\item{add_outgroup}{should the graph outgroup be added to the qpAdm right populations? Technically this shouldn't be an informative outgroup for qpAdm.}
}
\description{
For large admixturegraph, there may be a large number of valid qpAdm models!
}
\examples{
qpadm_models(igraph2, add_outgroup = TRUE)
qpadm_models(igraph2, add_outgroup = TRUE) \%>\% slice(1) \%$\% list(target, left, right)
}
