% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{qpadm_multi}
\alias{qpadm_multi}
\title{Run multiple qpadm models}
\usage{
qpadm_multi(
  data,
  models,
  allsnps = FALSE,
  full_results = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{The input data in the form of:
\itemize{
\item A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{extract_f2}}
\item A directory with f2 statistics
\item The prefix of a genotype file
}}

\item{models}{A nested list (or data frame) with qpadm models. It should consist of two or three other named lists (or columns) containing \code{left}, \code{right}, (and \code{target}) populations.}

\item{allsnps}{Use all SNPs with allele frequency estimates in every population of any given population quadruple. If \code{FALSE} (the default) only SNPs which are present in all populations in \code{popcombs} (or any given model in it) will be used. When there are populations with lots of (non-randomly) missing data, \code{allsnps = TRUE} can lead to false positive results. This option only has an effect if \code{data} is the prefix of a genotype file. If \code{data} are f2-statistics, the behavior will be determined by the options that were used in computing the f2-statistics.}

\item{verbose}{Print progress updates}

\item{...}{Further arguments passed to \code{\link{qpadm}}}
}
\value{
A list where each element is the output of one qpadm model.
}
\description{
This function runs multiple qpadm models, re-using f4-statistics where possible. Supports parallel evaluation of models, which can be turned on with \code{future::plan('multisession')} or similar, and turned off with \code{future::plan('sequential')}.
}
\examples{
\dontrun{
# the following specifies two models: one with 2/3/1 and one with 1/2/1 left/right/target populations
models = tibble(
           left = list(c('pop1', 'pop2'), c('pop3')),
           right = list(c('pop5', 'pop6', 'pop7'), c('pop7', 'pop8')),
           target = c('pop10', 'pop10'))
results = qpadm_multi('/my/geno/prefix', models)
}
}
