% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{qpadm_p}
\alias{qpadm_p}
\title{Faster version of \code{\link{qpadm}} with reduced output}
\usage{
qpadm_p(
  f2_data,
  left,
  right,
  target = NULL,
  fudge = 1e-04,
  boot = FALSE,
  constrained = FALSE,
  rnk = length(setdiff(left, target)) - 1,
  cpp = TRUE
)
}
\arguments{
\item{f2_data}{A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}}. Alternatively, a directory with f2 statistics.}

\item{left}{Left populations (sources)}

\item{right}{Right populations (outgroups)}

\item{target}{Target population}

\item{fudge}{Value added to diagonal matrix elements before inverting}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix
of f4 statistics, as well as the weight standard errors, will be computed using block-jackknife.
Otherwise bootstrap resampling is performed \code{n} times, where \code{n} is either equal to \code{boot} if it is an integer,
or equal to the number of blocks if \code{boot} is \code{TRUE}. The the covariance matrix of f4 statistics,
as well as the weight standard errors, will be computed using bootstrapping.}

\item{constrained}{Constrain admixture weights to be non-negative}

\item{rnk}{Rank of f4-matrix. Defaults to one less than full rank.}

\item{cpp}{Use C++ functions. Setting this to \code{FALSE} will be slower but can help with debugging.}
}
\value{
Data frame with \code{f4rank}, \code{dof}, \code{chisq}, \code{p}, \code{feasible}
}
\description{
Models target as a mixture of left populations, given a set of outgroup right populations.
Can be used to estimate admixture proportions, and to estimate the number of independent
admixture events.
}
\examples{
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
target = 'Denisova.DG'
qpadm_p(example_f2_blocks, left, right, target)
}
\seealso{
\code{\link{qpadm}}
}
