% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpadm_wrapper}
\alias{qpadm_wrapper}
\title{Wrapper function around the original qpAdm program}
\usage{
qpadm_wrapper(
  pref,
  left,
  right,
  target = NULL,
  bin = "~np29/o2bin/qpAdm",
  outdir = "./",
  parfile = NULL,
  allsnps = "NO",
  blgsize = 0.05,
  fancyf4 = "NO",
  f4mode = "YES",
  inbreed = "NO",
  printonly = FALSE,
  env = "",
  verbose = TRUE
)
}
\arguments{
\item{pref}{Path to and prefix of the packedancestrymap genotype files}

\item{left}{Left populations (or leftlist file)}

\item{right}{Right populations (or rightlist file)}

\item{target}{Target population}

\item{bin}{Path to the qpAdm binary file}

\item{outdir}{Output directory. files \code{out}, \code{parfile}, \code{leftlist},
\code{rightlist} will be overwritten}

\item{parfile}{qpAdm parameter file}

\item{allsnps}{allsnps}

\item{blgsize}{blgsize}

\item{fancyf4}{fancyf4}

\item{f4mode}{f4mode}

\item{inbreed}{inbreed}

\item{printonly}{Should the command be printed or executed?}

\item{env}{Export environmental variables. See examples.}

\item{verbose}{Print progress updates}
}
\value{
If not printonly, a data frame with parsed qpAdm output
}
\description{
This requires a working installation of qpAdm, which will be called using \code{\link{system}}
}
\examples{
\dontrun{
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
target = 'Denisova.DG'
qpadm_wrapper('genotype_prefix', left, right, target,
  bin = 'path/to/qpAdm')
}
}
