% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{qpadm_wrapper}
\alias{qpadm_wrapper}
\title{Wrapper function around the original qpAdm program}
\usage{
qpadm_wrapper(target = NULL, left = NULL, right = NULL, bin,
  pref = NULL, outdir = "./", parfile = NULL, printonly = FALSE,
  env = "")
}
\arguments{
\item{target}{target population}

\item{left}{source populations (or leftlist file)}

\item{right}{outgroup populations (or rightlist file)}

\item{bin}{path to the qpAdm binary file}

\item{pref}{path to and prefix of the packedancestrymap genotype files}

\item{outdir}{the output directory. files \code{out}, \code{parfile}, \code{leftlist}, \code{rightlist} will be overwritten}

\item{printonly}{should the command be printed or executed?}
}
\value{
if not printonly, a data frame with parsed qpAdm output
}
\description{
This requires a working installation of qpAdm, which will be called using \code{\link{system}}
}
\examples{
\dontrun{
target = 'Denisova.DG'
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
qpadm_wrapper(target, left, right,
  bin = 'path/to/qpAdm', pref = 'path/to/packedancestrymap_prefix',
  env = 'export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/path/to/blas/')
}
}
