% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qpdstat}
\alias{qpdstat}
\alias{f4}
\title{Estimate f4 statistics}
\usage{
qpdstat(
  data,
  pop1 = NULL,
  pop2 = NULL,
  pop3 = NULL,
  pop4 = NULL,
  boot = FALSE,
  sure = FALSE,
  unique_only = TRUE,
  comb = TRUE,
  blgsize = NULL,
  block_lengths = NULL,
  f4mode = TRUE,
  afprod = TRUE,
  cpp = TRUE,
  verbose = is.character(data),
  ...
)
}
\arguments{
\item{data}{Input data in one of three forms: \enumerate{
\item A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{f2_from_geno}} (fastest option)
\item A directory which contains pre-computed f2-statistics
\item The prefix of genotype files (slowest option)
}}

\item{pop1}{One of the following four:
\enumerate{
\item \code{NULL}: all possible population combinations will be returned
\item A vector of population labels. All combinations with the other \code{pop} arguments will be returned
\item A matrix with population combinations to be tested, with one population per column and one
combination per row. Other \code{pop} arguments will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or
population combinations to be tested. Other \code{pop} arguments will be ignored.
}}

\item{pop2}{A vector of population labels}

\item{pop3}{A vector of population labels}

\item{pop4}{A vector of population labels}

\item{boot}{If \code{FALSE} (the default), block-jackknife resampling will be used to compute standard errors.
Otherwise, block-bootstrap resampling will be used to compute standard errors. If \code{boot} is an integer, that number
will specify the number of bootstrap resamplings. If \code{boot = TRUE}, the number of bootstrap resamplings will be
equal to the number of SNP blocks.}

\item{sure}{The number of population combinations can get very large. This is a safety option that stops you
from accidently computing all combinations if that number is large.}

\item{unique_only}{If \code{TRUE} (the default), redundant combinations will be excluded}

\item{comb}{Generate all combinations of \code{pop1}, \code{pop2}, \code{pop3}, \code{pop4}. If \code{FALSE}, \code{pop1}, \code{pop2}, \code{pop3}, \code{pop4} should all be vectors of the same length.}

\item{blgsize}{SNP block size in Morgan. Default is 0.05 (5 cM). Only used when \code{data} is the prefix of genotype files}

\item{block_lengths}{Vector with lengths of each jackknife block. \code{sum(block_lengths)} has to
match the number of SNPs. only used when \code{data} is the prefix of genotype files}

\item{f4mode}{Set this to \code{FALSE} to compute D-statistics instead of f4. This only has an effect if the first argument is a genotype prefix. D-statistics are computed as \code{(a-b)*(c-d) / ((a + b - 2*a*b) * (c + d - 2*c*d))}, which is the same as \code{(P(ABBA) - P(BABA)) / (P(ABBA) + P(BABA))}}

\item{afprod}{Compute f4 from allele frequency products instead of f2 (default \code{TRUE}). Only used if \code{data} is a directory with precomputed data. For populations with lots of missing data, this option reduces bias that can result from setting \code{maxmiss} to values greater than 0. In all other cases it should not make a difference.}

\item{cpp}{Use C++ functions. Setting this to \code{FALSE} will be slower but can help with debugging.}

\item{verbose}{Print progress updates}

\item{...}{Additional arguments passed to \code{\link{f4blockdat_from_geno}} if \code{data} is a genotype file prefix or \code{\link{f2_from_precomp}} if \code{data} is a directory with f2-statistics}
}
\value{
\code{qpdstat} returns a data frame with f4 statistics
}
\description{
Computes f4-statistics of the form \eqn{f4(A, B; C, D)}. For allele frequencies \code{a}, \code{b}, \code{c}, \code{d}, \verb{f4(A, B; C, D)} is computed as the average of \code{(a-b)*(c-d)} over all SNPs in each SNP block. This is equivalent to \eqn{(f2(A, D) + f2(B, C) - f2(A, C) - f2(B, D)) / 2} (assuming no missing data). The input of this function can either be a 3d array of f2-statistics generated by \code{\link{f2_from_precomp}} or \code{\link{f2_from_geno}}, a directory with f2-statistics, or the prefix of genotype files. Computing f4 from genotype files directly is slower, but provides more flexibility in dealing with missing data (see details).
}
\details{
f4- and D-statistics are informative about how four populations are related to one another. Estimates of f4 are unbiased as long as assumptions about SNP ascertainment and mutation rates are met. Missing data can violate the ascertainment assumptions: an f4-statistic may be significantly different when it is calculated from all non-missing SNPs, compared to what it would be if it were calculated from all SNPs in the genome. However, because this difference is often small, f4 is often calculated using samples or populations with missing data, on a particular subset of all SNPs. There are different strategies for choosing the SNPs in this case, and these strategies differ in how many SNPs they use, how likely they lead to bias, and whether pre-computed f2-statistics can be used.
\itemize{
\item Use the same SNPs for every f4-statistic\cr
This is the most conservative option, but also the option which will use the smallest number of SNPs, which may result in a lack of power. It is the default option when pre-computing f2-statistics (\code{maxmiss = 0} in \code{\link{extract_f2}} or \code{\link{f2_from_geno}}).
\item Use different SNPs for each f4-statistic\cr
This option strikes a balance between avoiding bias and using a larger number of SNPs. For each f4-statistic it selects all SNPs which are present in all four populations. This option only works when the first argument is a genotype prefix (it doesn't work with pre-computed f2-statistics). In that case, this option is used by default. To turn it off and instead use the same SNPs for every f4-statistic, set \code{allsnps = FALSE}. This option is the default option in the original qpDstat program (it is in fact the only option for selecting SNPs in the original qpDstat; however in the original qpAdm and qpGraph programs, this mode of selecting SNPs can be enabled with the \code{allsnps} option, whereas the default mode in qpAdm and qpGraph is equivalent to \code{maxmiss = 0})
\item Use different SNPs for each f2-statistic\cr
This is the least conservative option, but also the option which uses most of the available information. It makes it possible to use pre-computed f2-statistics for a large number of populations without losing a large number of SNPs. To use this option, set the \code{maxmiss} parameter to a value greater than 0 (and not larger than 1) in \code{\link{extract_f2}} or \code{\link{f2_from_geno}}. When using this option, be aware that bias is possible, in particular for f4-statistics where some populations have large amounts of missing data. To reduce the bias that can result from using this option, you may want to combine it with using the option \code{afprod = TRUE} in \code{\link{f2_from_precomp}}.
In summary, whenever you work with populations with missing data, there is no guarantee that f4- or D-statistics involving these populations are not skewed in some way. If you choose to analyze these populations anyway, and you decide which SNPs to use, there is a trade-off between maximizing power and minimizing the risk of bias. One strategy might be to first use the least conservative option (setting \code{maxmiss = 1} in \code{\link{extract_f2}}) to get an overview, and then spot-check individual results using more conservative options.
}
}
\section{Alias}{

\code{f4}
}

\examples{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
pop3 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
pop4 = 'Switzerland_Bichon.SG'
qpdstat(example_f2_blocks, pop1, pop2, pop3, pop4)
\dontrun{
qpdstat(f2_dir, pop1, pop2, pop3, pop4)
}
\dontrun{
# compute D-statistics instead
qpdstat("/geno/prefix", pop1, pop2, pop3, pop4)
}
\dontrun{
# make a data frame with the population combinations for which f4 should be computed
combinations = tibble(pop1 = pop1, pop2 = pop2[1], pop3 = pop3, pop4 = pop4)
qpdstat(example_f2_blocks, combinations)
}
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history} Genetics

Peter, B. (2016) \emph{Admixture, Population Structure, and F-Statistics} Genetics
}
