% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qpdstat}
\alias{qpdstat}
\alias{f4}
\title{Estimate f4 statistics}
\usage{
qpdstat(pop1 = NULL, pop2 = NULL, pop3 = NULL, pop4 = NULL,
  f2_blocks = NULL, block_lengths = NULL, f2_dir = NULL,
  f2_denom = 1, sure = FALSE, unique_only = TRUE, verbose = FALSE)
}
\arguments{
\item{pop1}{one of the following four:
\enumerate{
\item \code{NULL}: all possible quadruples of the populations in \code{f2_blocks} will be returned
\item a vector of population labels
\item a data frame in which the first four columns specify the population quadruples to be tested. \code{pop2}, \code{pop3}, \code{pop4} will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or population quadruples to be tested. \code{pop2}, \code{pop3}, \code{pop4} will be ignored.
}}

\item{pop2}{a vector of population labels}

\item{pop3}{a vector of population labels}

\item{pop4}{a vector of population labels}

\item{f2_blocks}{3d array of block-jackknife leave-one-block-out estimates of f2 statistics. output of \code{\link{afs_to_f2_blocks}}. they are weighted by inverse of outgroup heterozygosity, if outgroup was specified.}

\item{block_lengths}{the jackknife block lengths used in computing the f2 statistics. see \code{\link{get_block_lengths}}.}

\item{f2_dir}{a directory with f2 statistics for each population pair in the graph. must contain 'block_lengths.RData'.}

\item{f2_denom}{scales f2-statistics. 1 correspondes to \code{f4mode: YES}. 1/4.75 is similar to \code{f4mode: NO}}
}
\value{
If \code{printonly}, the \code{qpDstat} command, otherwise a data frame with parsed \code{qpDstat} output
}
\description{
Computes f4 statistics from f2 jackknife blocks of the form \eqn{F_4(P_1, P_2; P_3, P_4)}. Equivalent to \eqn{(F_2(P_1, P_4) + F_2(P_2, P_3) - F_2(P_1, P_3) - F_2(P_2, P_4)) / 2}
}
\section{Alias}{

\code{f4}
}

\examples{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
pop3 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
pop4 = 'Switzerland_Bichon.SG'
qpdstat(pop1, pop2, pop3, pop4,
  example_f2_blocks, example_block_lengths)
\dontrun{
qpdstat(pop1, pop2, pop3, pop4, f2_dir = f2_dir)
}
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics
}
