% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qpdstat}
\alias{qpdstat}
\alias{f4}
\title{Estimate f4 statistics}
\usage{
qpdstat(f2_data, pop1 = NULL, pop2 = NULL, pop3 = NULL,
  pop4 = NULL, f2_denom = 1, boot = FALSE, sure = FALSE,
  unique_only = TRUE, comb = TRUE, verbose = FALSE)
}
\arguments{
\item{f2_data}{a 3d array of block-jackknife leave-one-block-out estimates of f2 statistics,
output of \code{\link{afs_to_f2_blocks}}. Alternatively, a directory with f2 statistics.
see \code{\link{extract_counts}}.}

\item{pop1}{one of the following four:
\enumerate{
\item \code{NULL}: all possible pairs of the populations in \code{f2_blocks} will be returned
\item a vector of population labels
\item a data frame with population combinations to be tested, with one population per column and one
combination per row. Other \code{pop} arguments will be ignored.
\item the location of a file (\code{poplistname} or \code{popfilename}) which specifies the populations or
population combinations to be tested. Other \code{pop} arguments will be ignored.
}}

\item{pop2}{a vector of population labels}

\item{pop3}{a vector of population labels}

\item{pop4}{a vector of population labels}

\item{f2_denom}{scales f2-statistics. 1 correspondes to \code{f4mode: YES}. 1/4.75 is similar to \code{f4mode: NO}.}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix
of the f statistics will be computed using block-jackknife. Otherwise bootstrap resampling is performed \code{n} times,
where \code{n} is either equal to \code{boot} if it is an integer, or equal to the number of blocks if \code{boot} is \code{TRUE}.
The the covariance matrix of the f statistics will be computed using bootstrapping.}

\item{sure}{The number of population combinations can get very large. This is a safety option that stops you
from accidently computing all combinations if that number is large.}

\item{unique_only}{If \code{TRUE} (the default), redundant combinations will be returned as well.}

\item{verbose}{print progress updates}
}
\value{
a data frame with f4 statistics
}
\description{
Computes f4 statistics from f2 blocks of the form \eqn{f4(A, B; C, D)}. Equivalent to
\eqn{(f2(A, D) + f2(B, C) - f2(A, C) - f2(B, D)) / 2}
}
\section{Alias}{

\code{f4}
}

\examples{
pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
pop3 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
pop4 = 'Switzerland_Bichon.SG'
qpdstat(example_f2_blocks, pop1, pop2, pop3, pop4)
\dontrun{
qpdstat(f2_dir, pop1, pop2, pop3, pop4)
}
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics
}
