% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qpf4ratio}
\alias{qpf4ratio}
\title{Estimate admixture proportions via f4 ratios}
\usage{
qpf4ratio(data, pops, boot = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{Input data in one of three forms:
\enumerate{
\item A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{f2_from_geno}} (fastest option)
\item A directory which contains pre-computed f2-statistics
\item The prefix of genotype files (slowest option)
}}

\item{pops}{A vector of 5 populations or a five column population matrix.
The following ratios will be computed: \verb{f4(1, 2; 3, 4)/f4(1, 2; 5, 4)}}

\item{boot}{If \code{FALSE} (the default), block-jackknife resampling will be used to compute standard errors.
Otherwise, block-bootstrap resampling will be used to compute standard errors. If \code{boot} is an integer, that number
will specify the number of bootstrap resamplings. If \code{boot = TRUE}, the number of bootstrap resamplings will be
equal to the number of SNP blocks.}

\item{verbose}{Print progress updates}
}
\value{
\code{qpf4ratio} returns a data frame with f4 ratios
}
\description{
Estimate admixture proportions via f4 ratios
}
