% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdstat.R
\name{qpf4ratio}
\alias{qpf4ratio}
\title{Estimate admixture proportions via f4 ratios}
\usage{
qpf4ratio(data, pops, boot = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{Input data in one of three forms:
\enumerate{
\item A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{extract_f2}} (fastest option)
\item A directory which contains pre-computed f2-statistics
\item The prefix of genotype files (slowest option)
}}

\item{pops}{A vector of 5 populations or a five column population matrix.
The following ratios will be computed: \verb{f4(1, 2; 3, 4)/f4(1, 2; 5, 4)}}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix
of the f statistics will be computed using block-jackknife. Otherwise bootstrap resampling is performed \code{n} times,
where \code{n} is either equal to \code{boot} if it is an integer, or equal to the number of blocks if \code{boot} is \code{TRUE}.
The the covariance matrix of the f statistics will be computed using bootstrapping.}

\item{verbose}{Print progress updates}
}
\value{
\code{qpf4ratio} returns a data frame with f4 ratios
}
\description{
Estimate admixture proportions via f4 ratios
}
