% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{qpgraph}
\alias{qpgraph}
\title{Compute the fit of an admixture graph}
\usage{
qpgraph(
  data,
  graph,
  lambdascale = 1,
  boot = FALSE,
  diag = 1e-04,
  diag_f3 = 1e-05,
  lsqmode = FALSE,
  numstart = 10,
  seed = NULL,
  cpp = TRUE,
  return_f4 = FALSE,
  f3precomp = NULL,
  f3basepop = NULL,
  constrained = TRUE,
  ppinv = NULL,
  f2_blocks_test = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{Input data in one of three forms:
\enumerate{
\item A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{extract_f2}} (fastest option)
\item A directory which contains pre-computed f2-statistics
\item The prefix of genotype files (slowest option)
}}

\item{graph}{An admixture graph represented as a matrix of edges, an \code{\link{igraph}} object, or the path to a \emph{qpGraph} graph file. Edges can be constrained by providing a matrix or data frame of edges with columns titled \code{lower} and \code{upper} with lower and upper bounds, respectively. By default, admixture edges are constrained to be between zero and one (with paired edges summing to one), and drift edges have a lower bound at zero.}

\item{lambdascale}{Scales f2-statistics. This has no effect on the fit, but is used in the original \emph{qpGraph} program to display branch weights on a scale that corresponds to FST distances.}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix of
f3 statistics will be computed using block-jackknife. Otherwise bootstrap resampling is performed \code{n} times,
where \code{n} is either equal to \code{boot} if it is an integer, or equal to the number of blocks if \code{boot} is \code{TRUE}.
The covariance matrix of f3 statistics will be computed using bootstrap resampling.}

\item{diag}{Regularization term added to the diagonal elements of the covariance matrix of fitted branch lengths (after scaling by the matrix trace). Default is 0.0001.}

\item{diag_f3}{Regularization term added to the diagonal elements of the covariance matrix of estimated f3 statistics (after scaling by the matrix trace). In the original \emph{qpGraph} program, this is fixed at 0.00001.}

\item{lsqmode}{Least-squares mode. If \code{TRUE}, the likelihood score will be computed using a diagonal matrix with \code{1/(sum(diag(f3_var)) * diag_f3)}, in place of the inverse f3-statistic covariance matrix.

\code{lsqmode = 2} will use the identity matrix instead, which is equivalent to computing the score as the sum of squared residuals (\code{sum((f3_est-f3_fit)^2)}).

Both of these options do not take the covariance of f3-statistics into account. This can lead to bias, but is more stable in cases where the inverse f3-statistics covariance matrix can not be estimated precisely (for example because the number of populations is large). An alternative to \code{lsqmode = TRUE} that doesn't completely ignore the covariance of f3-statistics is to increase \code{diag_f3}.}

\item{numstart}{Number of random initializations of starting weights. Defaults to 10. Increasing this number will make the optimization slower, but reduce the risk of not finding the optimal weights. Check the \code{opt} output to see how much the optimization depends on the starting weights.}

\item{seed}{Random seed for generating starting weights.}

\item{cpp}{Use C++ functions. Setting this to \code{FALSE} will be slower but can help with debugging.}

\item{return_f4}{Return all f4-statistics, as well as the z-score of the worst f4-statistic residual. Defaults to \code{FALSE} because this can be slow.}

\item{f3precomp}{Optional precomputed f3-statistics. This should be the output of \code{\link{qpgraph_precompute_f3}} and can be provided instead of \code{data}. This can speed things up if many graphs are evaluated using the same set of f3-statistics.}

\item{f3basepop}{Optional f3-statistics base population. Inference will be based on f3-statistics of the form \verb{f3(f3basepop; i, j)} for all population pairs \verb{(i, j)}. Defaults to the outgroup population if the graph has one. This option is ignored if \code{f3precomp} is provided. Changing \code{f3basepop} should make very little difference.}

\item{ppinv}{Optional inverse f3-statistics covariance matrix. Can be used for \code{\link{compare_fits3}}.}

\item{f2_blocks_test}{An optional 3d array of f2-statistics used for computing an out-of-sample score. This should contain only SNP blocks which are not part of \code{f2_blocks}. This allows to estimate the fit of a graph without overfitting and will not be used during the optimization step}

\item{verbose}{Print progress updates}
}
\value{
\code{qpgraph} returns a list with data describing the model fit:
\itemize{
\item \code{edges}: A data frame where each row is an edge in the graph. For regular edges,
the column \code{weight} is the estimated edge length, and for admixture edges, it is the estimated admixture weight.
\item \code{score}: The likelihood score of the fitted graph. Lower values correspond to better fits.
The score is calculated as the inner product of the residuals (difference between estimated and
fitted f3 statistics), weighted by the inverse of the f3 covariance matrix. See \code{\link{qpgraph_score}}
\item \code{f2}: Estimated and fitted f2 statistics. p-values and z-scores test the significance of the difference.
\item \code{f3}: Estimated and fitted f3 statistics. p-values and z-scores test the significance of the difference.
\item \code{f4}: Estimated and fitted f4 statistics (if \code{return_f4 = TRUE}). p-values and z-scores test the significance of the difference.
\item \code{opt}: A data frame with details of the weight-fitting step, including the randomly sampled starting weights. The column \code{value} contains the score for each set of starting weights. Columns starting with \code{x} denote initial weights, and columns starting with \code{y} denote fitted weights.
\item \code{worst_residual}: The highest absolute z-score of f4-statistics residuals (fitted - estimated f4); (returned if \code{return_f4 = TRUE})
}
}
\description{
Computes the fit of a given admixturegraph from f2-statistics. Drift edge weights and admixture edges weights are optimized until the (negative) likelihood score is minimized. The likelihood score is based on the squared difference between estimated and fitted f3-statistics.
}
\examples{
out = qpgraph(example_f2_blocks, example_graph)
plot_graph(out$edges)
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics
}
\seealso{
\code{\link{qpgraph_wrapper}} for a wrapper functions which calls the original \emph{qpGraph} program.
}
