% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{qpgraph}
\alias{qpgraph}
\title{Compute the fit of an admixture graph}
\usage{
qpgraph(f2_data, graph, f2_denom = 1, fnscale = 1e-06, fudge = 0.001,
  lsqmode = FALSE, numstart = NULL, seed = NULL, verbose = FALSE,
  cpp = TRUE)
}
\arguments{
\item{f2_data}{a 3d array of block-jackknife leave-one-block-out estimates of f2 statistics, output of \code{\link{afs_to_f2_blocks}}. alternatively, a directory with f2 statistics. see \code{\link{extract_data}}.}

\item{graph}{an admixture graph represented as a matrix of edges, an \code{\link{igraph}} object, or the path to a qpGraph graph file.}

\item{f2_denom}{scales f2-statistics. A value of around 0.278 converts F2 to Fst.}

\item{fnscale}{try increasing or decreasing this, if optimization does not converge.}

\item{fudge}{try increasing this, if you get the error message \code{constraints are inconsistent, no solution!}.}

\item{lsqmode}{least-squares mode. sets the offdiagonal elements of the block-jackknife covariance matrix to zero.}

\item{numstart}{number of random initializations. defaults to 10 times the number of admixture nodes.}

\item{seed}{seed for generating starting weights.}

\item{verbose}{print optimization iterations}

\item{cpp}{should optimization be done using C++ or R function? \code{cpp = TRUE} is much faster.}
}
\value{
a list of qpGraph output data
}
\description{
Computes the fit of an admixturegraph for a given graph topology and empirical f2-block-jackknife statistics.
}
\examples{
out = qpgraph(example_graph, example_f2_blocks)
plot_graph(out$edges)
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics
}
\seealso{
\code{\link{qpgraph_wrapper}} for a wrapper functions which call the original qpGraph program; \code{\link{qpgraph_slim}} for a faster function function which requires f3 estimates and an inverted covariance matrix as input instead.
}
