% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{qpgraph_precompute_f3}
\alias{qpgraph_precompute_f3}
\title{Compute f3-statistics from f2-statistics.}
\usage{
qpgraph_precompute_f3(
  f2_data,
  pops,
  f3basepop = NULL,
  f2_denom = 1,
  boot = FALSE,
  seed = NULL,
  fudge_cov = 1e-05,
  lsqmode = FALSE
)
}
\arguments{
\item{f2_data}{A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}}.
alternatively, a directory with precomputed data. see \code{\link{extract_f2}} and \code{\link{extract_counts}}.}

\item{pops}{Populations for which to compute f3-statistics}

\item{f3basepop}{f3-statistics base population. If \code{NULL} (the default),
the first population in \code{pops} will be used as the basis.}

\item{f2_denom}{Scales f2-statistics. A value of around 0.278 converts F2 to Fst.}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix of f3 statistics
will be computed using block-jackknife. Otherwise bootstrap resampling is performed \code{n} times, where \code{n} is either
equal to \code{boot} if it is an integer, or equal to the number of blocks if \code{boot} is \code{TRUE}. The the covariance matrix
of f3 statistics will be computed using bootstrapping.}

\item{seed}{Random seed used if \code{boot} is \code{TRUE}.}

\item{fudge_cov}{Regularization term added to the diagonal elements of the covariance matrix of estimated f3 statistics (after scaling by the matrix trace).}

\item{lsqmode}{Least-squares mode. If \code{TRUE}, the likelihood score will be computed using a diagonal matrix with \code{1/(sum(diag(f3_var)) * fudge_cov)}, in place of the inverse f3-statistic covariance matrix. \code{lsqmode = 2} will use the identity matrix instead, which is equivalent to computing the score as the sum of squared residuals (\code{sum((f3_est-f3_fit)^2)}). Both of these options do not take the covariance of f3-statistics into account. This can lead to bias, but is more stable in cases where the inverse f3-statistics covariance matrix can not be estimated precisely (for example because the number of populations is large). An alternative to using \code{lsqmode = TRUE} which doesn't completely ignore the covariance of f3-statistics is to increase \code{fudge_cov}.}
}
\value{
A list with four items
\enumerate{
\item \code{f3_est} a matrix with f3-statistics for all population pairs with the output
\item \code{ppinv} a matrix with the inverse of the f3-statistic covariance matrix
\item \code{f2out} a data frame with f2 estimates
\item \code{f3out} a data frame with f3 estimates
}
}
\description{
Takes a 3d array of f2 block jackknife estimates and computes f3-statistics between the
first population \eqn{p1} and all population pairs \eqn{i, j}: \eqn{f3(p1; p_i, p_j)}
}
\examples{
pops = get_leafnames(example_igraph)
qpgraph_precompute_f3(example_f2_blocks, pops)$f3_est
\dontrun{
qpgraph_precompute_f3(f2_dir, pops, f2_denom = 0.278)
}
}
