% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{qpgraph_precompute_f3}
\alias{qpgraph_precompute_f3}
\title{Compute f3-statistics from f2-statistics.}
\usage{
qpgraph_precompute_f3(
  data,
  pops,
  f3basepop = NULL,
  lambdascale = 1,
  boot = FALSE,
  seed = NULL,
  diag_f3 = 1e-05,
  lsqmode = FALSE
)
}
\arguments{
\item{data}{Input data in one of three forms:
\enumerate{
\item A 3d array of blocked f2 statistics, output of \code{\link{f2_from_precomp}} or \code{\link{extract_f2}} (fastest option)
\item A directory which contains pre-computed f2-statistics
\item The prefix of genotype files (slowest option)
}}

\item{pops}{Populations for which to compute f3-statistics}

\item{f3basepop}{f3-statistics base population. If \code{NULL} (the default),
the first population in \code{pops} will be used as the basis.}

\item{lambdascale}{Scales f2-statistics. This has no effect on the fit, but is used in the original \emph{qpGraph} program to display branch weights on a scale that corresponds to FST distances.}

\item{boot}{If \code{FALSE} (the default), block-jackknife resampling will be used to compute standard errors.
Otherwise, block-bootstrap resampling will be used to compute standard errors. If \code{boot} is an integer, that number
will specify the number of bootstrap resamplings. If \code{boot = TRUE}, the number of bootstrap resamplings will be
equal to the number of SNP blocks.}

\item{seed}{Random seed used if \code{boot} is \code{TRUE}.}

\item{diag_f3}{Regularization term added to the diagonal elements of the covariance matrix of estimated f3 statistics (after scaling by the matrix trace). In the original \emph{qpGraph} program, this is fixed at 0.00001.}

\item{lsqmode}{Least-squares mode. If \code{TRUE}, the likelihood score will be computed using a diagonal matrix with \code{1/(sum(diag(f3_var)) * diag_f3)}, in place of the inverse f3-statistic covariance matrix. \code{lsqmode = 2} will use the identity matrix instead, which is equivalent to computing the score as the sum of squared residuals (\code{sum((f3_est-f3_fit)^2)}). Both of these options do not take the covariance of f3-statistics into account. This can lead to bias, but is more stable in cases where the inverse f3-statistics covariance matrix can not be estimated precisely (for example because the number of populations is large). An alternative to using \code{lsqmode = TRUE} which doesn't completely ignore the covariance of f3-statistics is to increase \code{diag_f3}.}
}
\value{
A list with four items
\enumerate{
\item \code{f3_est} a matrix with f3-statistics for all population pairs with the output
\item \code{ppinv} a matrix with the inverse of the f3-statistic covariance matrix
\item \code{f2out} a data frame with f2 estimates
\item \code{f3out} a data frame with f3 estimates
}
}
\description{
Takes a 3d array of f2 block jackknife estimates and computes f3-statistics between the
first population \eqn{p1} and all population pairs \eqn{i, j}: \eqn{f3(p1; p_i, p_j)}
}
\examples{
pops = get_leafnames(example_igraph)
qpgraph_precompute_f3(example_f2_blocks, pops)$f3_est
\dontrun{
qpgraph_precompute_f3(f2_dir, pops)
}
}
