% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{qpgraph_slim}
\alias{qpgraph_slim}
\title{Compute the fit of an admixturegraph.}
\usage{
qpgraph_slim(graph, f3_jest, ppinv, pops, fnscale = 1e-06,
  numstart = 10, seed = NULL, verbose = FALSE, cpp = TRUE)
}
\arguments{
\item{graph}{an admixture graph represented as an \code{\link{igraph}} object}

\item{f3_jest}{matrix of f3 estimates. output of \code{\link{qpgraph_precompute_f3}}}

\item{ppinv}{inverted covariance matrix of f3-statistics}

\item{pops}{populations in \code{f3_jest} and \code{ppinv}}

\item{fnscale}{try increasing or decreasing this, if optimization does not converge.}

\item{numstart}{number of random initializations. defaults to 10 times the number of admixture nodes.}

\item{seed}{seed for generating starting weights.}

\item{verbose}{print optimization iterations}

\item{cpp}{should optimization be run in C++ or in R? C++ is faster.}

\item{fudge}{try increasing this, if you get the error message \code{constraints are inconsistent, no solution!}.}
}
\value{
a list of qpGraph output data
}
\description{
Computes the fit of an admixture graph for a given graph topology given f3 estimates and an inverted covariance matrix.
}
\examples{
pops = get_leafnames(example_igraph)
precomp = qpgraph_precompute_f3(pops, example_f2_blocks, example_block_lengths)
f3_jest = precomp$f3_jest
ppinv = precomp$ppinv
out = qpgraph_slim(example_igraph, f3_jest, ppinv, pops)
plot_graph(out$edges)
}
\seealso{
\code{\link{qpgraph_wrapper}}for a wrapper functions which call the original qpGraph program; \code{\link{qpgraph}} for a slower function which requires f2 block jackknife statistics as input instead. \code{\link{qpgraph_precompute_f3}} computes the required input from a 3d array of \code{f2_statistics}
}
