% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{qpgraph_wrapper}
\alias{qpgraph_wrapper}
\title{A wrapper around qpGraph which requires an existing 'parfile' and 'graphfile'}
\usage{
qpgraph_wrapper(bin = "./qpGraph", parfile = "./parfile",
  graphfile = "./graphfile", outfile = "./out", printonly = FALSE)
}
\arguments{
\item{bin}{location of the qpGraph binary file.}

\item{parfile}{qpGraph parameter file.}

\item{graphfile}{qpGraph graph file.}

\item{outfile}{output file.}

\item{printonly}{should output be executed or the command just be printed?}
}
\value{
a list of qpGraph output data.
}
\description{
A wrapper around qpGraph which requires an existing 'parfile' and 'graphfile'
}
