% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{qpwave}
\alias{qpwave}
\title{Estimate admixture waves}
\usage{
qpwave(
  f2_data,
  left,
  right,
  fudge = 1e-04,
  boot = FALSE,
  constrained = FALSE,
  cpp = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{left}{Left populations (sources)}

\item{right}{Right populations (outgroups)}

\item{fudge}{value added to diagonal matrix elements before inverting}

\item{boot}{If \code{FALSE} (the default), each block will be left out at a time and the covariance matrix
of f4 statistics, as well as the weight standard errors, will be computed using block-jackknife.
Otherwise bootstrap resampling is performed \code{n} times, where \code{n} is either equal to \code{boot} if it is an integer,
or equal to the number of blocks if \code{boot} is \code{TRUE}. The the covariance matrix of f4 statistics,
as well as the weight standard errors, will be computed using bootstrapping.}

\item{constrained}{Constrain admixture weights to be non-negative}

\item{cpp}{Use C++ functions. Setting this to \code{FALSE} will be slower but can help with debugging.}

\item{verbose}{Print progress updates}
}
\value{
A list with output describing the model fit:
\enumerate{
\item \code{f4} A data frame with estimated f4-statistics
\item \code{rankdrop} A data frame describing model fits with different ranks, including p-values for the overall fit
and for nested models (comparing two models with rank difference of one).
}
}
\description{
\code{qpwave} compares two sets of populations (\code{left} and \code{right}) to each other. It estimates a lower bound on the number of admixtue waves that went from \code{left} into \code{right}, by comparing a matrix of f4-statistics to low-rank approximations. For a rank of 0 this is equivalent to testing whether \code{left} and \code{right} form clades relative to each other.
}
\examples{
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
qpwave(example_f2_blocks, left, right)
}
\references{
Patterson, N. et al. (2012) \emph{Ancient admixture in human history.} Genetics

Haak, W. et al. (2015) \emph{Massive migration from the steppe was a source for Indo-European
languages in Europe.} Nature (SI 10)
}
\seealso{
\code{\link{qpadm}}
}
