% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{random_newick}
\alias{random_newick}
\title{Generate a random, binary graph with n terminal nodes}
\usage{
random_newick(n, start = "", end = "", outpop = NULL)
}
\arguments{
\item{n}{the number of terminal nodes, or a vector of population labels.}

\item{start}{prefix.}

\item{end}{postfix.}

\item{outpop}{outgroup (if \code{n} is a vector of labels).}
}
\value{
tree in newick format.
}
\description{
Generate a random, binary graph with n terminal nodes
}
\examples{
random_newick(5)
random_newick(c('a', 'b', 'c', 'd'))
}
