% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_ancestrymap}
\alias{read_ancestrymap}
\title{Read genotype data from ancestrymap files}
\usage{
read_ancestrymap(pref, inds = NULL, verbose = TRUE)
}
\arguments{
\item{pref}{Prefix of packedancestrymap files (files have to end in \code{.geno}, \code{.ind}, \code{.snp})}

\item{inds}{Individuals for which data should be read. Defaults to all individuals}

\item{verbose}{Print progress updates}
}
\value{
A list with the genotype data matrix, the \code{.ind} file, and the \code{.snp} file
}
\description{
Read genotype data from ancestrymap files
}
\examples{
\dontrun{
samples = c('Ind1', 'Ind2', 'Ind3')
geno = read_packedancestrymap(prefix, samples)
}
}
