% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_f2}
\alias{read_f2}
\title{Read blocked f2 estimates from disk}
\usage{
read_f2(f2_dir, pops = NULL, afprod = FALSE, remove_na = TRUE, verbose = FALSE)
}
\arguments{
\item{f2_dir}{Directory from which to read files}

\item{pops}{Populations for which f2 statistics should be read. Defaults to all populations,
which may require a lot of memory.}

\item{afprod}{Return allele frequency products instead of f2 estimates}

\item{remove_na}{Remove blocks with missing values}

\item{verbose}{Print progress updates}
}
\value{
A 3d array of block jackknife estimates
}
\description{
This function reads blocked f2 estimates (or allele frequency products) which were writtend to disk by \code{\link{write_f2}}
and returns them as a 3d array.
}
\examples{
\dontrun{
read_f2(f2_dir, pops = c('pop1', 'pop2', 'pop3'))
}
}
\seealso{
\code{\link{write_f2}}
}
