% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_packedancestrymap}
\alias{read_packedancestrymap}
\title{Read genotype data from packedancestrymap files}
\usage{
read_packedancestrymap(pref, inds = NULL, blocksize = 1000,
  verbose = TRUE)
}
\arguments{
\item{pref}{prefix of packedancestrymap files (files have to end in \code{.geno}, \code{.ind}, \code{.snp}).}

\item{inds}{optional vector of samples to read in}

\item{blocksize}{number of SNPs read in each block.}

\item{verbose}{print progress updates}
}
\value{
a list with the genotype data matrix, the \code{.ind} file, and the \code{.snp} file
}
\description{
This is currently slower than \code{read_plink} because it is implemented only in \code{R}, not in \code{C++}.
}
\examples{
\dontrun{
samples = c('Ind1', 'Ind2', 'Ind3')
geno = read_packedancestrymap(prefix, samples)
}
}
