% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_packedancestrymap}
\alias{read_packedancestrymap}
\title{Read genotype data from packedancestrymap files}
\usage{
read_packedancestrymap(
  pref,
  inds = NULL,
  pops = NULL,
  first = 1,
  last = Inf,
  transpose = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{pref}{Prefix of the packedancestrymap files}

\item{inds}{Individuals for which data should be read. Defaults to all individuals}

\item{pops}{Populations for which data should be read. Cannot be provided together with 'inds'}

\item{first}{Index of first SNP to read}

\item{last}{Index of last SNP to read}

\item{transpose}{Transpose genotype matrix (default is \code{snps} x \code{individuals})}

\item{verbose}{Print progress updates}
}
\value{
A list with the genotype data matrix, the \code{.ind} file, and the \code{.snp} file
}
\description{
Read genotype data from packedancestrymap files
}
\examples{
\dontrun{
samples = c('Ind1', 'Ind2', 'Ind3')
geno = read_packedancestrymap(prefix, samples)
}
}
