% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_plink}
\alias{read_plink}
\title{Read genotype data from \code{PLINK} files}
\usage{
read_plink(pref, inds = NULL, auto_only = TRUE, verbose = FALSE)
}
\arguments{
\item{pref}{prefix of packedancestrymap files (files have to end in \code{.geno}, \code{.ind}, \code{.snp}).}

\item{inds}{optional vector of samples to read in}

\item{auto_only}{If \code{TRUE} (the default), all SNPs with chromosome outside the range of 1 to 22 will be discarded.}

\item{verbose}{print progress updates}

\item{blocksize}{number of SNPs read in each block.}
}
\value{
a list with the genotype data matrix, the \code{.ind} file, and the \code{.snp} file
}
\description{
This is based on Gad Abraham's \code{\href{https://github.com/gabraham/plink2R}{plink2R}} package,
but genotypes are \code{m} x \code{n}, not \code{n} x \code{m}.
See \code{\href{https://www.rdocumentation.org/packages/genio}{genio}} for a dedicated \code{R} package for
reading and writing \code{PLINK} files.
}
\examples{
\dontrun{
samples = c('Ind1', 'Ind2', 'Ind3')
geno = read_packedancestrymap(prefix, samples)
}
}
