% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{split_mat}
\alias{split_mat}
\title{Split a matrix into blocks}
\usage{
split_mat(mat, cols_per_block, outdir, overwrite = FALSE,
  verbose = TRUE)
}
\arguments{
\item{mat}{the matrix to be split}

\item{cols_per_block}{the number of columns per block}

\item{outdir}{output directory}
}
\description{
This function splits a large matrix into smaller blocks with \code{cols_per_block} columns per block,
and saves them as \code{.rds} files in \code{outdir}
}
\examples{
\dontrun{
afmatall = packedancestrymap_to_aftable('path/to/packedancestrymap_prefix', allpopulations,
                                         na.action = 'none', return_matrix = TRUE)
split_mat(afmatall, cols_per_block = 20, outdir = 'afmatall_split_v42.1/')
}
}
\seealso{
\code{\link{packedancestrymap_to_aftable}}, \code{\link{write_split_f2_block}}
}
