% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{spr_all}
\alias{spr_all}
\title{Modify a graph by regrafting a subcomponent}
\usage{
spr_all(graph, fix_outgroup = TRUE)
}
\arguments{
\item{graph}{An admixture graph}
}
\value{
A new admixture graph
}
\description{
Modify a graph by regrafting a subcomponent
}
