% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_triples}
\alias{summarize_triples}
\title{Summarize triples across graphs}
\usage{
summarize_triples(results, maxscore = NA)
}
\arguments{
\item{results}{the output of \code{optimize_admixturegraph}}

\item{maxscore}{restrict summary to graphs with score not larger than \code{maxscore}}
}
\description{
This summarizes topologies of population triples across graphs
}
\examples{
\dontrun{
summarize_triples(opt_results)
}
}
