% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toposearch.R
\name{summarize_zerof4}
\alias{summarize_zerof4}
\title{List clades in a graph}
\usage{
summarize_zerof4(graph, eps = 1e-07)
}
\arguments{
\item{graph}{An admixture graph}

\item{eps}{Used to determine what counts as zero}
}
\value{
A data frame with all (non-redundant) zero f4-statistics
}
\description{
List clades in a graph
}
\examples{
\dontrun{
summarize_zerof4(example_igraph)
}
}
