% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpadm.R
\name{test_cladality}
\alias{test_cladality}
\title{Test if two sets of populations form two clades}
\usage{
test_cladality(f2_data, left, right, fudge = 1e-04, boot = FALSE, cpp = TRUE)
}
\arguments{
\item{left}{Left populations (sources)}

\item{right}{Right populations (outgroups)}

\item{fudge}{Value added to diagonal matrix elements before inverting}

\item{boot}{If \code{FALSE} (the default), block-jackknife resampling will be used to compute standard errors.
Otherwise, block-bootstrap resampling will be used to compute standard errors. If \code{boot} is an integer, that number
will specify the number of bootstrap resamplings. If \code{boot = TRUE}, the number of bootstrap resamplings will be
equal to the number of SNP blocks.}

\item{cpp}{Use C++ functions. Setting this to \code{FALSE} will be slower but can help with debugging.}
}
\description{
A thin wrapper around \code{\link{qpadm_p}} with \code{rnk} set to zero
}
