% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_block_lengths}
\alias{write_block_lengths}
\title{Compute and write block lengths}
\usage{
write_block_lengths(pref, outdir, dist = 0.05, distcol = "cm",
  auto_only = TRUE)
}
\description{
Compute and write block lengths
}
\examples{
\dontrun{
pref = 'path/to/packedancestrymap_prefix'
genodir = 'split_v42.1'
outdir = 'indpairs_v42.1'
write_block_lengths(pref, outdir)
dat = read_packedancestrymap(pref, inds = inds)
split_mat(dat$geno, cols_per_block = 20, outdir = genodir)
write_split_inddat(genodir, outdir)
numblocks = 185 # this should be the number of split allele frequency files
for(j in 1:numblocks) {
  for(j in i:numblocks) {
    write_split_pairdat(genodir, outdir, block1 = i, block2 = j)
    }
  }
}
}
