% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{write_dot}
\alias{write_dot}
\title{Convert graph to dot format}
\usage{
write_dot(edges, outfile = stdout())
}
\arguments{
\item{edges}{Graph as edge list (columns labelled 'from', 'to', 'weight')}

\item{outfile}{Output file name}
}
\description{
Convert graph to dot format
}
\examples{
\dontrun{
results = qpgraph(example_f2_blocks, example_graph)
write_dot(results$edges)
}
}
