% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_f2}
\alias{write_f2}
\title{Write f2 block jackknife estimates to disk}
\usage{
write_f2(f2_block, path, overwrite = FALSE)
}
\arguments{
\item{f2_block}{3d array with f2 block jackknife estimates. The first two dimensions of \code{f2_block} have to have population names. Each file will be stored under \code{{path}/{pop1}/{pop2}.RData}.}

\item{path}{directory into which to write the files.}

\item{overwrite}{should existing \code{.RData} files be overwritten?}
}
\description{
This function takes a 3d array of f2 block jackknife estimates, splits it by population pair, and writes each pair to a separate \code{.RData} file.
}
\examples{
\dontrun{
write_f2(f2_block, path = 'path/to/f2stats/')
}
}
\seealso{
\code{\link{read_f2}}
}
