% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_split_f2_block}
\alias{write_split_f2_block}
\title{Compute f2 blocks and write them to disk}
\usage{
write_split_f2_block(
  afdir,
  outdir,
  chunk1,
  chunk2,
  blgsize = 0.05,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{afdir}{Directory with allele frequency and counts \code{.rds} files created by \code{\link{split_mat}}}

\item{outdir}{Directory where data will be stored}

\item{chunk1}{Index of the first chunk of populations}

\item{chunk2}{Index of the second chunk of populations}

\item{blgsize}{SNP block size in Morgan. Default is 0.05 (50 cM).}

\item{overwrite}{Overwrite existing files (default \code{FALSE})}

\item{verbose}{Print progress updates}
}
\description{
This is intended for computing f2-statistics for a large number of populations,
too many to do everything in working memory.
It assumes that the allele frequencies have already been computed and are stored in \code{.rds} files,
split into consecutive blocks for a set of populations. This function calls \code{\link{write_f2}},
which takes a (sub-)chunk of pairwise f2-statistics, and writes one pair at a time to disk.
}
\examples{
\dontrun{
afdir = 'tmp_af_dir/'
f2dir = 'f2_dir'
extract_afs('path/to/packedancestrymap_prefix', afdir)
numchunks = length(list.files(afdir, 'afs.+rds'))
# numchunks should be the number of split allele frequency files
for(i in 1:numchunks) {
  for(j in i:numchunks) {
    write_split_f2_block(afdir, f2dir, chunk1 = i, chunk2 = j)
  }
}
}
# Alternatively, the following code will do the same, while submitting each chunk as a separate job.
# (if \code{\link[future]{plan}} has been set up appropriately)
\dontrun{
furrr::future_map(1:numchunks, ~{i=.; map(i:numchunks, ~{
  write_split_f2_block(afdir, f2dir, chunk1 = i, chunk2 = .)
  })})
  }
}
\seealso{
\code{\link{extract_f2}} Does the same thing in one step for smaller data.
}
